% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workflow.R
\name{ActivityArgument}
\alias{ActivityArgument}
\title{ActivityArgument}
\description{
Class representing an activity argument object.
}
\examples{
\dontrun{
ActivityArgument$new()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{argumentName}}{The name of the argument.}

\item{\code{acceptableArgumentTypes}}{A list of strings containing the acceptable
argument types.}

\item{\code{defaultArgumentType}}{The default argument type. One of:
\code{TimeSignalUnit}, \code{DepthSignalUnit}, \code{StaticSignalUnit},
\code{StringSignalUnit}, \code{StringTimeSignalUnit}, \code{PVTSignalUnit},
\code{Hierarchy}, \code{TagEntries}, \code{WorkspaceValue}, \code{Model}}

\item{\code{defaultSignalName}}{The name of the default signal.}

\item{\code{defaultUnitName}}{The name of the default unit.}

\item{\code{defaultString}}{The default string.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ActivityArgument-new}{\code{ActivityArgument$new()}}
\item \href{#method-ActivityArgument-toList}{\code{ActivityArgument$toList()}}
\item \href{#method-ActivityArgument-clone}{\code{ActivityArgument$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActivityArgument-new"></a>}}
\if{latex}{\out{\hypertarget{method-ActivityArgument-new}{}}}
\subsection{Method \code{new()}}{
Create a new activity argument instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivityArgument$new(
  argumentName = NULL,
  acceptableArgumentTypes = NULL,
  defaultArgumentType = c("TimeSignalUnit", "DepthSignalUnit", "StaticSignalUnit",
    "StringSignalUnit", "StringTimeSignalUnit", "PVTSignalUnit", "Hierarchy",
    "TagEntries", "WorkspaceValue", "Model"),
  defaultSignalName = NULL,
  defaultUnitName = NULL,
  defaultString = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{argumentName}}{The name of the argument.}

\item{\code{acceptableArgumentTypes}}{A list of strings containing the
acceptable argument types.}

\item{\code{defaultArgumentType}}{The default argument type. One of:
\code{TimeSignalUnit}, \code{DepthSignalUnit}, \code{StaticSignalUnit},
\code{StringSignalUnit}, \code{StringTimeSignalUnit},
\code{PVTSignalUnit}, \code{Hierarchy}, \code{TagEntries},
\code{WorkspaceValue}, \code{Model}}

\item{\code{defaultSignalName}}{The name of the default signal.}

\item{\code{defaultUnitName}}{The name of the default unit.}

\item{\code{defaultString}}{The default string.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActivityArgument-toList"></a>}}
\if{latex}{\out{\hypertarget{method-ActivityArgument-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivityArgument$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActivityArgument-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ActivityArgument-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivityArgument$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
