% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workflow.R
\name{ActivityMappedArgument}
\alias{ActivityMappedArgument}
\title{ActivityMappedArgument}
\description{
Class representing an activity mapped argument object.
}
\examples{
\dontrun{
ActivityMappedArgument$new()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{argumentName}}{The name of the argument.}

\item{\code{argumentType}}{The argument type. One of:
\code{TimeSignalUnit}, \code{DepthSignalUnit}, \code{StaticSignalUnit},
\code{StringSignalUnit}, \code{StringTimeSignalUnit}, \code{PVTSignalUnit},
\code{Hierarchy}, \code{TagEntries}, \code{WorkspaceValue}, \code{Model}}

\item{\code{acceptableArgumentTypes}}{A list of strings containing the acceptable
argument types.}

\item{\code{mappedSignalName}}{The name of the mapped signal.}

\item{\code{mappedUnitName}}{The name of the mapped unit.}

\item{\code{mappedString}}{The mapped string.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ActivityMappedArgument-new}{\code{ActivityMappedArgument$new()}}
\item \href{#method-ActivityMappedArgument-toList}{\code{ActivityMappedArgument$toList()}}
\item \href{#method-ActivityMappedArgument-clone}{\code{ActivityMappedArgument$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActivityMappedArgument-new"></a>}}
\if{latex}{\out{\hypertarget{method-ActivityMappedArgument-new}{}}}
\subsection{Method \code{new()}}{
Create a new activity argument instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivityMappedArgument$new(
  argumentName = NULL,
  argumentType = c("TimeSignalUnit", "DepthSignalUnit", "StaticSignalUnit",
    "StringSignalUnit", "StringTimeSignalUnit", "PVTSignalUnit", "Hierarchy",
    "TagEntries", "WorkspaceValue", "Model"),
  acceptableArgumentTypes = NULL,
  mappedSignalName = NULL,
  mappedUnitName = NULL,
  mappedString = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{argumentName}}{The name of the argument.}

\item{\code{argumentType}}{The argument type. One of:
\code{TimeSignalUnit}, \code{DepthSignalUnit}, \code{StaticSignalUnit},
\code{StringSignalUnit}, \code{StringTimeSignalUnit},
\code{PVTSignalUnit}, \code{Hierarchy}, \code{TagEntries},
\code{WorkspaceValue}, \code{Model}}

\item{\code{acceptableArgumentTypes}}{A list of strings containing the
acceptable argument types.}

\item{\code{mappedSignalName}}{The name of the mapped signal.}

\item{\code{mappedUnitName}}{The name of the mapped unit.}

\item{\code{mappedString}}{The mapped string.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActivityMappedArgument-toList"></a>}}
\if{latex}{\out{\hypertarget{method-ActivityMappedArgument-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivityMappedArgument$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActivityMappedArgument-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ActivityMappedArgument-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivityMappedArgument$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
