% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApiRequests.R
\name{ApiRequests}
\alias{ApiRequests}
\title{ApiRequests}
\description{
Performs REST API calls
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ApiRequests-new}{\code{ApiRequests$new()}}
\item \href{#method-ApiRequests-clone}{\code{ApiRequests$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiRequests-new"></a>}}
\if{latex}{\out{\hypertarget{method-ApiRequests-new}{}}}
\subsection{Method \code{new()}}{
Create a new ApiRequests instance.
description Issue a POST request to the api

param body The body of the post request.
param url The base url of the API.
param route The route to use. E.g. Signal, Files, etc.
param token_type The type of the used token.
param token The token used for authenticating the request.
param expect_data Whether to expect data being returned from the request.
param query Named list of additional query parameters.
description Upload a file using a POST request to the api

param file The file to upload (path incl. file name).
param url The base url of the API.
param route The route to use. E.g. Signal, Files, etc.
param token_type The type of the used token.
param token The token used for authenticating the request.
description Issue a GET request to the api

param url The base url of the API.
param route The route to use. E.g. Signal, Files, etc.
param token_type The type of the used token.
param token The token used for authenticating the request.
param query Additional parameters for the query string.
param parse_json Whether to parse the result of the request. Defaults to
 \code{TRUE}. If set to \code{FALSE}, the raw content will be returned.

returns The result of the GET request. Either as received, or parsed to
 an object using \code{jsonlite::fromJSON}.
description Issue a DELETE request to the api

param name Name of the item to delete.
param url The base url of the API.
param route The route to use. E.g. Signal, Files, etc.
param token_type The type of the used token.
param token The token used for authenticating the request.
param query named list of query parameters.
description Issue a PUT request to the api

param body The body of the request.
param url The base url of the API.
param route The route to use. E.g. Signal, Files, etc.
param token_type The type of the used token.
param token The token used for authenticating the request.
param query named list of query parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiRequests$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiRequests-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ApiRequests-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiRequests$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
