% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chart.R
\name{Axis}
\alias{Axis}
\title{Axis}
\description{
Class representing a chart axis object. Axes are used when
creating a chart.
}
\examples{
\dontrun{
Axis$new(title = "Fruits eaten",
         isOpposed = false,
         isLog = false,
         axisId = "fruits",
         manualMin = NULL,
         manualMax = NULL,
         isInversed = false,
         ticksPerInterval = 10)
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{title}}{The title of the axis.}

\item{\code{isOpposed}}{Flag specifying whether the axis is located on the opposing
side of the chart (left or top).}

\item{\code{isLog}}{Flag specifying whether the axis is scaled logarithmically.}

\item{\code{axisId}}{The id of the axis.}

\item{\code{manualMin}}{Integer specifying the manually set minimum of the axis.}

\item{\code{manualMax}}{Integer specifying the manually set maximum of the axis.}

\item{\code{isInversed}}{Flag specifying whether the axis is inversed.}

\item{\code{ticksPerInterval}}{Integer specifying the number of ticks per interval.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Axis-new}{\code{Axis$new()}}
\item \href{#method-Axis-toList}{\code{Axis$toList()}}
\item \href{#method-Axis-clone}{\code{Axis$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Axis-new"></a>}}
\if{latex}{\out{\hypertarget{method-Axis-new}{}}}
\subsection{Method \code{new()}}{
Create a new Axis instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Axis$new(
  title = NULL,
  isOpposed = NULL,
  isLog = NULL,
  axisId = NULL,
  manualMin = NULL,
  manualMax = NULL,
  isInversed = NULL,
  ticksPerInterval = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{The title of the axis.}

\item{\code{isOpposed}}{Flag specifying whether the axis is located on the
opposing side of the chart (left or top).}

\item{\code{isLog}}{Flag specifying whether the axis is scaled logarithmically.}

\item{\code{axisId}}{The id of the axis.}

\item{\code{manualMin}}{Integer specifying the manually set minimum of the axis.}

\item{\code{manualMax}}{Integer specifying the manually set maximum of the axis.}

\item{\code{isInversed}}{Flag specifying whether the axis is inversed.}

\item{\code{ticksPerInterval}}{Integer specifying the number of ticks per
interval.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Axis-toList"></a>}}
\if{latex}{\out{\hypertarget{method-Axis-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Axis$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Axis-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Axis-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Axis$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
