% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chart.R
\name{Chart}
\alias{Chart}
\title{Chart}
\description{
Class representing a PetroVisor chart object.
}
\examples{
\dontrun{
Chart$new()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the chart.}

\item{\code{dataType}}{The datatype of the chart (string).}

\item{\code{legendLocation}}{A string specifying the location of the legend.}

\item{\code{isLegendVisible}}{A flag specifying whether the legend is visible.}

\item{\code{isLegendOrientationHorizontal}}{A flag specifying whether the legend is
oriented horizontal.}

\item{\code{isLegendPlacementInside}}{A flag specifying whether the legend is placed
inside or outside the chart.}

\item{\code{chartType}}{The type of the chart (string).}

\item{\code{isStacked}}{A flag specifying whether the chart is stacked.}

\item{\code{isStacked100}}{A flag specifying whether the chart is stacked.}

\item{\code{isGridLinesOnXAxis}}{A flag specifying whether gridlines are visible on
the x-axis.}

\item{\code{isGridLinesOnYAxis}}{A flag specifying whether gridlines are visible on
the y-axis.}

\item{\code{dataAxesMappings}}{A list giving the axes mappings.}

\item{\code{seriesSettings}}{A list of series settings. The individual settings
must be instances of the class \code{SeriesSetting}.}

\item{\code{axes}}{A list of axes. The individual axes must be instances of the
class \code{Axis}.}

\item{\code{showToolTipTitle}}{A flag specifying whether the tooltip title is shown.}

\item{\code{showToolTipX}}{A flag specifying whether to show the x-value in the
tooltip.}

\item{\code{showToolTipY}}{A flag specifying whether to show the y-value in the
tooltip.}

\item{\code{isLocked}}{This flag specifies whether the chart is locked.
Defaults to \code{FALSE}.}

\item{\code{user}}{The user the chart belongs to.}

\item{\code{isFavorite}}{This flag specifies whether the chart is marked as favorite
item, and thus shown in the favorites tab on the home module in PetroVisor.
Defaults to \code{FALSE}.}

\item{\code{labels}}{A list of strings holding the labels of the chart.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Chart-new}{\code{Chart$new()}}
\item \href{#method-Chart-toList}{\code{Chart$toList()}}
\item \href{#method-Chart-clone}{\code{Chart$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Chart-new"></a>}}
\if{latex}{\out{\hypertarget{method-Chart-new}{}}}
\subsection{Method \code{new()}}{
Create a new Scope instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chart$new(
  name = NULL,
  dataType = NULL,
  legendLocation = NULL,
  isLegendVisible = NULL,
  isLegendOrientationHorizontal = NULL,
  isLegendPlacementInside = NULL,
  chartType = NULL,
  isStacked = NULL,
  isStacked100 = NULL,
  isGridLinesOnXAxis = NULL,
  isGridLinesOnYAxis = NULL,
  dataAxesMappings = NULL,
  seriesSettings = NULL,
  axes = NULL,
  showToolTipTitle = NULL,
  showToolTipX = NULL,
  showToolTipY = NULL,
  isLocked = FALSE,
  user = NULL,
  isFavorite = FALSE,
  labels = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the chart.}

\item{\code{dataType}}{The datatype of the chart (string).}

\item{\code{legendLocation}}{A string specifying the location of the legend.}

\item{\code{isLegendVisible}}{A flag specifying whether the legend is visible.}

\item{\code{isLegendOrientationHorizontal}}{A flag specifying whether the legend is
oriented horizontal.}

\item{\code{isLegendPlacementInside}}{A flag specifying whether the legend is placed
inside or outside the chart.}

\item{\code{chartType}}{The type of the chart (string).}

\item{\code{isStacked}}{A flag specifying whether the chart is stacked.}

\item{\code{isStacked100}}{A flag specifying whether the chart is stacked.}

\item{\code{isGridLinesOnXAxis}}{A flag specifying whether gridlines are visible on
the x-axis.}

\item{\code{isGridLinesOnYAxis}}{A flag specifying whether gridlines are visible on
the y-axis.}

\item{\code{dataAxesMappings}}{A list giving the axes mappings.}

\item{\code{seriesSettings}}{A list of series settings. The individual settings
must be instances of the class \code{SeriesSetting}.}

\item{\code{axes}}{A list of axes. The individual axes must be instances of the
class \code{Axis}.}

\item{\code{showToolTipTitle}}{A flag specifying whether the tooltip title is shown.}

\item{\code{showToolTipX}}{A flag specifying whether to show the x-value in the
tooltip.}

\item{\code{showToolTipY}}{A flag specifying whether to show the y-value in the
tooltip.}

\item{\code{isLocked}}{This flag specifies whether the chart is locked.
Defaults to \code{FALSE}.}

\item{\code{user}}{The user the chart belongs to.}

\item{\code{isFavorite}}{This flag specifies whether the chart is marked as favorite
item, and thus shown in the favorites tab on the home module in PetroVisor.
Defaults to \code{FALSE}.}

\item{\code{labels}}{A list of strings holding the labels of the chart.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Chart-toList"></a>}}
\if{latex}{\out{\hypertarget{method-Chart-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chart$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Chart-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Chart-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chart$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
