% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculation.R
\name{CleansingFilter}
\alias{CleansingFilter}
\title{CleansingFilter}
\description{
Class representing a cleansing filter object.
}
\examples{
\dontrun{
CleansingFilter$new()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the cleansing filter.}

\item{\code{formula}}{The cleansing filter's formula.}

\item{\code{unitName}}{The cleansing filter's unit name.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CleansingFilter-new}{\code{CleansingFilter$new()}}
\item \href{#method-CleansingFilter-toList}{\code{CleansingFilter$toList()}}
\item \href{#method-CleansingFilter-clone}{\code{CleansingFilter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CleansingFilter-new"></a>}}
\if{latex}{\out{\hypertarget{method-CleansingFilter-new}{}}}
\subsection{Method \code{new()}}{
Create a new CleansingFilter instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CleansingFilter$new(name = NULL, formula = NULL, unitName = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the cleansing filter.}

\item{\code{formula}}{The cleansing filter's formula.}

\item{\code{unitName}}{The cleansing filter's unit name.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CleansingFilter-toList"></a>}}
\if{latex}{\out{\hypertarget{method-CleansingFilter-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CleansingFilter$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CleansingFilter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CleansingFilter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CleansingFilter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
