% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfigurationSetting.R
\name{ConfigurationSetting}
\alias{ConfigurationSetting}
\title{ConfigurationSetting}
\description{
Class representing a PetroVisor ConfigurationSetting object.
}
\examples{
\dontrun{
ConfigurationSetting$new(name = "MyStringValue",
                         value_type = "String",
                         string_value = "Hello World!")
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the configuration setting.}

\item{\code{numeric_value}}{The numeric value of the configuration setting.}

\item{\code{string_value}}{The string value of the configuration setting.}

\item{\code{list_value}}{The list value of the configuration setting (type list).}

\item{\code{enumeration_value}}{The enumeration value of the configuration setting
(named list).}

\item{\code{dictionary_value}}{The dictionary value of the configuration setting
(named list).}

\item{\code{value_type}}{The value type of the configuration setting.
Must be one of: \code{"Numeric"}, \code{"String"},
\code{"NumericWithUnit"}, \code{"List"}, \code{"Enumeration"},
\code{"Dictionary"}, \code{"StringWithType"}, \code{"ListWithType"},
\code{"StringFromValues"}, \code{"ListFromValues"}}

\item{\code{unit_name}}{The name of the unit associated with the numeric value if
the valueType is \code{"Numeric with unit"}.}

\item{\code{possible_values}}{A list containing the possible values of the
configuration setting.}

\item{\code{is_system}}{This flaf specifies, whether the configuration setting is a
PetroVisor system setting (immutable).}

\item{\code{description}}{The description of the item.}

\item{\code{labels}}{A list of strings holding the labels of the item.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ConfigurationSetting-new}{\code{ConfigurationSetting$new()}}
\item \href{#method-ConfigurationSetting-toList}{\code{ConfigurationSetting$toList()}}
\item \href{#method-ConfigurationSetting-clone}{\code{ConfigurationSetting$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfigurationSetting-new"></a>}}
\if{latex}{\out{\hypertarget{method-ConfigurationSetting-new}{}}}
\subsection{Method \code{new()}}{
Create a new ConfigurationSetting instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigurationSetting$new(
  name = NULL,
  numeric_value = NULL,
  string_value = NULL,
  list_value = list(),
  enumeration_value = list(),
  dictionary_value = list(),
  value_type = c("Numeric", "String", "NumericWithUnit", "List", "Enumeration",
    "Dictionary", "StringWithType", "ListWithType", "StringFromValues", "ListFromValues"),
  unit_name = NULL,
  possible_values = list(),
  is_system = NULL,
  description = NULL,
  labels = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the configuration setting.}

\item{\code{numeric_value}}{The numeric value of the configuration setting.}

\item{\code{string_value}}{The string value of the configuration setting.}

\item{\code{list_value}}{The list value of the configuration setting.}

\item{\code{enumeration_value}}{The enumeration value of the configuration
setting (named list).}

\item{\code{dictionary_value}}{The dictionary value of the configuration setting
(named list).}

\item{\code{value_type}}{The value type of the configuration setting.
Must be one of: \code{"Numeric"}, \code{"String"},
\code{"NumericWithUnit"}, \code{"List"}, \code{"Enumeration"},
\code{"Dictionary"}, \code{"StringWithType"}, \code{"ListWithType"},
\code{"StringFromValues"}, \code{"ListFromValues"}}

\item{\code{unit_name}}{The name of the unit associated with the numeric value
if the valueType is \code{"Numeric with unit"}.}

\item{\code{possible_values}}{A list containing the possible values of the
configuration setting.}

\item{\code{is_system}}{This flaf specifies, whether the configuration setting
is a PetroVisor system setting (immutable).}

\item{\code{description}}{The description of the item.}

\item{\code{labels}}{A list of strings holding the labels of the item.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfigurationSetting-toList"></a>}}
\if{latex}{\out{\hypertarget{method-ConfigurationSetting-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigurationSetting$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfigurationSetting-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ConfigurationSetting-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigurationSetting$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
