% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossPlot.R
\name{CrossPlot}
\alias{CrossPlot}
\title{CrossPlot}
\description{
Class representing a PetroVisor cross plot object.
}
\examples{
\dontrun{
CrossPlot$new()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the cross plot.}

\item{\code{dataSets}}{A list of data sets in the cross plot. The items in the
list must be instances of the class \code{CrossPlotDataSet}.}

\item{\code{dataStart}}{The start date of the data.}

\item{\code{dataEnd}}{The end date of the data.}

\item{\code{dataStep}}{The time increment of the data. One of:
\code{"EveryMinute"}, \code{"EveryFiveMinutes"},
\code{"EveryFifteenMinutes"}, \code{"Hourly"}, \code{"Daily"},
\code{"Monthly"}, \code{"Quarterly"}, \code{"Yearly"}.}

\item{\code{hierarchyName}}{The name of a hierarchy.}

\item{\code{mode}}{The mode of the cross plot.}

\item{\code{xAxisSignalName}}{The name of the signal on the x-axis.}

\item{\code{xAxisUnitName}}{The name of the unit on the x-axis.}

\item{\code{xAxisEntityName}}{The name of the entity on the x-axis.}

\item{\code{yAxisSignalName}}{The name of the signal on the y-axis.}

\item{\code{yAxisUnitName}}{The name of the unit on the y-axis.}

\item{\code{yAxisEntityName}}{The name of the entity on the y-axis.}

\item{\code{skippedEntityNames}}{A list of the names of skipped entities.}

\item{\code{timeMin}}{The minimum time.}

\item{\code{timeMax}}{The maximum time.}

\item{\code{isLocked}}{This flag specifies whether the cross plot is locked.
Defaults to \code{FALSE}.}

\item{\code{user}}{The user the cross plot belongs to.}

\item{\code{isFavorite}}{This flag specifies whether the cross plot is marked as
favorite item, and thus shown in the favorites tab on the home module in
PetroVisor. Defaults to \code{FALSE}.}

\item{\code{labels}}{A list of strings holding the labels of the cross plot.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CrossPlot-new}{\code{CrossPlot$new()}}
\item \href{#method-CrossPlot-toList}{\code{CrossPlot$toList()}}
\item \href{#method-CrossPlot-clone}{\code{CrossPlot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CrossPlot-new"></a>}}
\if{latex}{\out{\hypertarget{method-CrossPlot-new}{}}}
\subsection{Method \code{new()}}{
Create a new CrossPlot instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossPlot$new(
  name = NULL,
  dataSets = NULL,
  dataStart = NULL,
  dataEnd = NULL,
  dataStep = c("EveryMinute", "EveryFiveMinutes", "EveryFifteenMinutes", "Hourly",
    "Daily", "Monthly", "Quarterly", "Yearly"),
  hierarchyName = NULL,
  mode = NULL,
  xAxisSignalName = NULL,
  xAxisUnitName = NULL,
  xAxisEntityName = NULL,
  yAxisSignalName = NULL,
  yAxisUnitName = NULL,
  yAxisEntityName = NULL,
  skippedEntityNames = NULL,
  timeMin = NULL,
  timeMax = NULL,
  isLocked = FALSE,
  user = NULL,
  isFavorite = FALSE,
  labels = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the cross plot.}

\item{\code{dataSets}}{A list of data sets in the cross plot. The items in the
list must be instances of the class \code{CrossPlotDataSet}.}

\item{\code{dataStart}}{The start date of the data.}

\item{\code{dataEnd}}{The end date of the data.}

\item{\code{dataStep}}{The time increment of the data. One of:
\code{"EveryMinute"}, \code{"EveryFiveMinutes"},
\code{"EveryFifteenMinutes"}, \code{"Hourly"}, \code{"Daily"},
\code{"Monthly"}, \code{"Quarterly"}, \code{"Yearly"}.}

\item{\code{hierarchyName}}{The name of a hierarchy.}

\item{\code{mode}}{The mode of the cross plot.}

\item{\code{xAxisSignalName}}{The name of the signal on the x-axis.}

\item{\code{xAxisUnitName}}{The name of the unit on the x-axis.}

\item{\code{xAxisEntityName}}{The name of the entity on the x-axis.}

\item{\code{yAxisSignalName}}{The name of the signal on the y-axis.}

\item{\code{yAxisUnitName}}{The name of the unit on the y-axis.}

\item{\code{yAxisEntityName}}{The name of the entity on the y-axis.}

\item{\code{skippedEntityNames}}{A list of the names of skipped entities.}

\item{\code{timeMin}}{The minimum time.}

\item{\code{timeMax}}{The maximum time.}

\item{\code{isLocked}}{This flag specifies whether the cross plot is locked.
Defaults to \code{FALSE}.}

\item{\code{user}}{The user the cross plot belongs to.}

\item{\code{isFavorite}}{This flag specifies whether the cross plot is marked as
favorite item, and thus shown in the favorites tab on the home module in
PetroVisor. Defaults to \code{FALSE}.}

\item{\code{labels}}{A list of strings holding the labels of the cross plot.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CrossPlot-toList"></a>}}
\if{latex}{\out{\hypertarget{method-CrossPlot-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossPlot$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CrossPlot-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CrossPlot-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossPlot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
