% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossPlot.R
\name{CrossPlotDataSet}
\alias{CrossPlotDataSet}
\title{CrossPlotDataSet}
\description{
Class representing a cross plot data set object.
}
\examples{
\dontrun{
CrossPlotDataSet$new()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{signalName}}{The name of the signal.}

\item{\code{unitName}}{The name of the unit.}

\item{\code{entityName}}{The name of the entity.}

\item{\code{symbol}}{A string giving the symbol to use for the data set.}

\item{\code{color}}{The color to use for the data set (string).}

\item{\code{min}}{The min value of the data set.}

\item{\code{max}}{The max value of the data set.}

\item{\code{trendLine}}{A string specifying whether to show a trend line.
Either \code{"On"} or \code{"Off"}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CrossPlotDataSet-new}{\code{CrossPlotDataSet$new()}}
\item \href{#method-CrossPlotDataSet-toList}{\code{CrossPlotDataSet$toList()}}
\item \href{#method-CrossPlotDataSet-clone}{\code{CrossPlotDataSet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CrossPlotDataSet-new"></a>}}
\if{latex}{\out{\hypertarget{method-CrossPlotDataSet-new}{}}}
\subsection{Method \code{new()}}{
Create a new CrossPlotDataSet instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossPlotDataSet$new(
  signalName = NULL,
  unitName = NULL,
  entityName = NULL,
  symbol = NULL,
  color = NULL,
  min = NULL,
  max = NULL,
  trendLine = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{signalName}}{The name of the signal.}

\item{\code{unitName}}{The name of the unit.}

\item{\code{entityName}}{The name of the entity.}

\item{\code{symbol}}{A string giving the symbol to use for the data set.}

\item{\code{color}}{The color to use for the data set (string).}

\item{\code{min}}{The min value of the data set.}

\item{\code{max}}{The max value of the data set.}

\item{\code{trendLine}}{A string specifying whether to show a trend line.
Either \code{"On"} or \code{"Off"}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CrossPlotDataSet-toList"></a>}}
\if{latex}{\out{\hypertarget{method-CrossPlotDataSet-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossPlotDataSet$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CrossPlotDataSet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CrossPlotDataSet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossPlotDataSet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
