% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessTemplate.R
\name{CustomField}
\alias{CustomField}
\title{CustomField}
\description{
Class representing a process template custom field.
}
\examples{
\dontrun{
CustomField$new()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the custom field.}

\item{\code{evaluatedColumn}}{The evaluated column. Must be an instance of the
class \code{Column}.}

\item{\code{evaluatedDate}}{The date of the evaluation (string).}

\item{\code{evaluatedValue}}{The result of the evaluation (string).}

\item{\code{fieldType}}{The type of the custom field (integer).}

\item{\code{scriptName}}{The name of the script used by the custom field.}

\item{\code{scriptColumnName}}{The name of the column used by the custom field.}

\item{\code{currentManualValue}}{The current manually entered value.}

\item{\code{initialValue}}{The initial value of the custom field.}

\item{\code{expressionTime}}{The time of the expression.}

\item{\code{activeStep}}{The name of the step the custom field is active in.}

\item{\code{settingName}}{The name of the custom field's setting.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CustomField-new}{\code{CustomField$new()}}
\item \href{#method-CustomField-toList}{\code{CustomField$toList()}}
\item \href{#method-CustomField-clone}{\code{CustomField$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomField-new"></a>}}
\if{latex}{\out{\hypertarget{method-CustomField-new}{}}}
\subsection{Method \code{new()}}{
Create a new CustomField instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomField$new(
  name = NULL,
  evaluatedColumn = NULL,
  evaluatedDate = NULL,
  evaluatedValue = NULL,
  fieldType = NULL,
  scriptName = NULL,
  scriptColumnName = NULL,
  currentManualValue = NULL,
  initialValue = NULL,
  expressionTime = NULL,
  activeStep = NULL,
  settingName = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the custom field.}

\item{\code{evaluatedColumn}}{The evaluated column. Must be an instance of the
class \code{Column}.}

\item{\code{evaluatedDate}}{The date of the evaluation (string).}

\item{\code{evaluatedValue}}{The result of the evaluation (string).}

\item{\code{fieldType}}{The type of the custom field (integer).}

\item{\code{scriptName}}{The name of the script used by the custom field.}

\item{\code{scriptColumnName}}{The name of the column used by the custom field.}

\item{\code{currentManualValue}}{The current manually entered value.}

\item{\code{initialValue}}{The initial value of the custom field.}

\item{\code{expressionTime}}{The time of the expression.}

\item{\code{activeStep}}{The name of the step the custom field is active in.}

\item{\code{settingName}}{The name of the custom field's setting.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomField-toList"></a>}}
\if{latex}{\out{\hypertarget{method-CustomField-toList}{}}}
\subsection{Method \code{toList()}}{
Convert the object to a list. This function is mainly used
by the Services to convert the objects to lists and then
call the web API.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomField$toList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomField-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CustomField-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomField$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
