% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workflow.R
\name{CustomWorkflowActivity}
\alias{CustomWorkflowActivity}
\title{CustomWorkflowActivity}
\description{
Class representing a custom workflow activity object.
}
\examples{
\dontrun{
CustomWorkflowActivity$new()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the custom workflow activity.}

\item{\code{className}}{The custom workflow activity's class name.}

\item{\code{assemblyContent}}{The content of the custom workflow activity.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CustomWorkflowActivity-new}{\code{CustomWorkflowActivity$new()}}
\item \href{#method-CustomWorkflowActivity-toList}{\code{CustomWorkflowActivity$toList()}}
\item \href{#method-CustomWorkflowActivity-clone}{\code{CustomWorkflowActivity$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomWorkflowActivity-new"></a>}}
\if{latex}{\out{\hypertarget{method-CustomWorkflowActivity-new}{}}}
\subsection{Method \code{new()}}{
Create a new custom workflow activity instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomWorkflowActivity$new(
  name = NULL,
  className = NULL,
  assemblyContent = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the custom workflow activity.}

\item{\code{className}}{The custom workflow activity's class name.}

\item{\code{assemblyContent}}{The content of the custom workflow activity.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomWorkflowActivity-toList"></a>}}
\if{latex}{\out{\hypertarget{method-CustomWorkflowActivity-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomWorkflowActivity$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomWorkflowActivity-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CustomWorkflowActivity-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomWorkflowActivity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
