% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataIntegration.R
\name{DataConnection}
\alias{DataConnection}
\title{DataConnection}
\description{
Class representing a PetroVisor data connection object.
}
\examples{
\dontrun{
DataConnection$new()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the data connection.}

\item{\code{connectionType}}{The type of the data connection.}

\item{\code{settings}}{The settings of the data connection.}

\item{\code{isLocked}}{This flag specifies whether the data connection is locked.
Defaults to \code{FALSE}.}

\item{\code{user}}{The user the data connection belongs to.}

\item{\code{isFavorite}}{This flag specifies whether the data connection is marked
as favorite item, and thus shown in the favorites tab on the home module in
PetroVisor. Defaults to \code{FALSE}.}

\item{\code{labels}}{A list of strings holding the labels of the data connection.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataConnection-new}{\code{DataConnection$new()}}
\item \href{#method-DataConnection-toList}{\code{DataConnection$toList()}}
\item \href{#method-DataConnection-clone}{\code{DataConnection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataConnection-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataConnection-new}{}}}
\subsection{Method \code{new()}}{
Create a new DataConnection instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataConnection$new(
  name = NULL,
  connectionType = NULL,
  settings = NULL,
  isLocked = FALSE,
  user = NULL,
  isFavorite = FALSE,
  labels = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the data connection.}

\item{\code{connectionType}}{The type of the data connection.}

\item{\code{settings}}{The settings of the data connection.}

\item{\code{isLocked}}{This flag specifies whether the data connection is
locked. Defaults to \code{FALSE}.}

\item{\code{user}}{The user the data connection belongs to.}

\item{\code{isFavorite}}{This flag specifies whether the data connection is
marked as favorite item, and thus shown in the favorites tab on the
home module in PetroVisor. Defaults to \code{FALSE}.}

\item{\code{labels}}{A list of strings holding the labels of the data
connection.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataConnection-toList"></a>}}
\if{latex}{\out{\hypertarget{method-DataConnection-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataConnection$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataConnection-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataConnection-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataConnection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
