% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataIntegration.R
\name{DataMapping}
\alias{DataMapping}
\title{DataMapping}
\description{
Class representing a data mapping as used in data sources.
}
\examples{
\dontrun{
DataMapping$new()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{sourceRef}}{The reference in the source.}

\item{\code{entityName}}{The name of the mapped entity.}

\item{\code{signalName}}{The name of the mapped signal.}

\item{\code{unitName}}{The name of the mapped unit.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataMapping-new}{\code{DataMapping$new()}}
\item \href{#method-DataMapping-toList}{\code{DataMapping$toList()}}
\item \href{#method-DataMapping-clone}{\code{DataMapping$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataMapping-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataMapping-new}{}}}
\subsection{Method \code{new()}}{
Create a new DataMapping instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataMapping$new(
  sourceRef = NULL,
  entityName = NULL,
  signalName = NULL,
  unitName = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sourceRef}}{The reference in the source.}

\item{\code{entityName}}{The name of the mapped entity.}

\item{\code{signalName}}{The name of the mapped signal.}

\item{\code{unitName}}{The name of the mapped unit.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataMapping-toList"></a>}}
\if{latex}{\out{\hypertarget{method-DataMapping-toList}{}}}
\subsection{Method \code{toList()}}{
Convert the object to a list. This function is mainly used
by the Services to convert the objects to lists and then
call the web API.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataMapping$toList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataMapping-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataMapping-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataMapping$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
