% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataTypes.R
\name{DataPoint}
\alias{DataPoint}
\title{DataPoint}
\description{
Class representing a time-dependent data point.
}
\examples{
\dontrun{
DataPoint$new(date = "2020-01-01T00:00:00.000Z", value = 20)
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{date}}{The timestamp of the data point.}

\item{\code{value}}{The value at the specified timestamp.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataPoint-new}{\code{DataPoint$new()}}
\item \href{#method-DataPoint-toList}{\code{DataPoint$toList()}}
\item \href{#method-DataPoint-clone}{\code{DataPoint$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataPoint-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataPoint-new}{}}}
\subsection{Method \code{new()}}{
Create a new DataPoint instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataPoint$new(date = NULL, value = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{date}}{The timestamp of the data point.}

\item{\code{value}}{The value at the specified timestamp.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataPoint-toList"></a>}}
\if{latex}{\out{\hypertarget{method-DataPoint-toList}{}}}
\subsection{Method \code{toList()}}{
Convert the object to a list. This function is mainly used
by the DataServices to convert the objects to lists and then
call the web API.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataPoint$toList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataPoint-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataPoint-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataPoint$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
