% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DcaFit.R
\name{DcaFit}
\alias{DcaFit}
\title{DcaFit}
\description{
Class representing a PetroVisor DcaFit object.
}
\examples{
\dontrun{
DcaFit$new( )
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the DCA fit.}

\item{\code{entity}}{The entity for which the DCA fit is created. Must be an
instance of the class \code{Entity}.}

\item{\code{rateSignal}}{The signal giving the production rate. Must be an
instance of the class \code{Signal}.}

\item{\code{rateUnit}}{The unit the rate signal is given in. Must be an
instance of the class \code{Unit}.}

\item{\code{cumulativeSignal}}{The signal giving the cumulative data. Must be
an instance of the class \code{Signal}.}

\item{\code{cumulativeUnit}}{The unit the cumulative signal is given in.
Must be an instance of the class \code{Unit}.}

\item{\code{timeScope}}{The scope of the fit. Must be an instance of the class
\code{Scope}.}

\item{\code{dataPoints}}{A list holding the data points for the fit. The items
in the list must be instances of the class \code{DataPoint}.}

\item{\code{excludedDataPoints}}{A list of excluded data points. The items
in the list must be instances of the class \code{DataPoint}.}

\item{\code{cumulativeDataPoints}}{A list holding the cumulative data points
for the fit. The items in the list must be instances of the class
\code{DataPoint}.}

\item{\code{fitRangeStart}}{The start date of the fit range (string).}

\item{\code{fitRangeEnd}}{The end date of the fit range (string).}

\item{\code{fitType}}{The type of the fit.}

\item{\code{isAutomatic}}{This flag specifies whether the fit was created
automatically.}

\item{\code{isReviewed}}{This flag specifies whether the fit was reviewed.}

\item{\code{isReviewNeeded}}{This flag specifies whether a review of the fit
is needed.}

\item{\code{intercept}}{The value of the fit's y-intercept.}

\item{\code{slope}}{The value of the fit's slope.}

\item{\code{hyperbolicConstant}}{The value of the fit's hyperbolic constant.}

\item{\code{rootSquared}}{The RS value of the fit.}

\item{\code{meanAbsoluteError}}{The fit's mean absolute error.}

\item{\code{rootMeanSquareError}}{The fit's RMS error.}

\item{\code{estimatedUltimateRecovery}}{The EUR calculated from the fit.}

\item{\code{remainingReserves}}{The remaining reserves based on the fit.}

\item{\code{economicLimit}}{The economic limit for the calculation of EUR and
remaining reserves.}

\item{\code{leftBehind}}{The reserves that are left behind.}

\item{\code{nowDate}}{The current date.}

\item{\code{forecastEndDate}}{The end date of the forecast.}

\item{\code{isLocked}}{This flag specifies whether the DCA fit is locked.
Defaults to \code{FALSE}.}

\item{\code{user}}{The user the DCA fit belongs to.}

\item{\code{isFavorite}}{This flag specifies whether the DCA fit is marked as
favorite item, and thus shown in the favorites tab on the home module in
PetroVisor. Defaults to \code{FALSE}.}

\item{\code{labels}}{A list of strings holding the labels of the DCA fit.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DcaFit-new}{\code{DcaFit$new()}}
\item \href{#method-DcaFit-toList}{\code{DcaFit$toList()}}
\item \href{#method-DcaFit-clone}{\code{DcaFit$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DcaFit-new"></a>}}
\if{latex}{\out{\hypertarget{method-DcaFit-new}{}}}
\subsection{Method \code{new()}}{
Create a new DCAFit instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DcaFit$new(
  name = NULL,
  entity = NULL,
  rateSignal = NULL,
  rateUnit = NULL,
  cumulativeSignal = NULL,
  cumulativeUnit = NULL,
  timeScope = NULL,
  dataPoints = NULL,
  excludedDataPoints = NULL,
  cumulativeDataPoints = NULL,
  fitRangeStart = NULL,
  fitRangeEnd = NULL,
  fitType = NULL,
  isAutomatic = NULL,
  isReviewed = NULL,
  isReviewNeeded = NULL,
  intercept = NULL,
  slope = NULL,
  hyperbolicConstant = NULL,
  rootSquared = NULL,
  meanAbsoluteError = NULL,
  rootMeanSquareError = NULL,
  estimatedUltimateRecovery = NULL,
  remainingReserves = NULL,
  economicLimit = NULL,
  leftBehind = NULL,
  nowDate = NULL,
  forecastEndDate = NULL,
  isLocked = FALSE,
  user = NULL,
  isFavorite = FALSE,
  labels = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the DCA fit.}

\item{\code{entity}}{The entity for which the DCA fit is created. Must be an
instance of the class \code{Entity}.}

\item{\code{rateSignal}}{The signal giving the production rate. Must be an
instance of the class \code{Signal}.}

\item{\code{rateUnit}}{The unit the rate signal is given in. Must be an
instance of the class \code{Unit}.}

\item{\code{cumulativeSignal}}{The signal giving the cumulative data. Must be
an instance of the class \code{Signal}.}

\item{\code{cumulativeUnit}}{The unit the cumulative signal is given in.
Must be an instance of the class \code{Unit}.}

\item{\code{timeScope}}{The scope of the fit. Must be an instance of the class
\code{Scope}.}

\item{\code{dataPoints}}{A list holding the data points for the fit. The items
in the list must be instances of the class \code{DataPoint}.}

\item{\code{excludedDataPoints}}{A list of excluded data points. The items
in the list must be instances of the class \code{DataPoint}.}

\item{\code{cumulativeDataPoints}}{A list holding the cumulative data points
for the fit. The items in the list must be instances of the class
\code{DataPoint}.}

\item{\code{fitRangeStart}}{The start date of the fit range (string).}

\item{\code{fitRangeEnd}}{The end date of the fit range (string).}

\item{\code{fitType}}{The type of the fit.}

\item{\code{isAutomatic}}{This flag specifies whether the fit was created
automatically.}

\item{\code{isReviewed}}{This flag specifies whether the fit was reviewed.}

\item{\code{isReviewNeeded}}{This flag specifies whether a review of the fit
is needed.}

\item{\code{intercept}}{The value of the fit's y-intercept.}

\item{\code{slope}}{The value of the fit's slope.}

\item{\code{hyperbolicConstant}}{The value of the fit's hyperbolic constant.}

\item{\code{rootSquared}}{The RS value of the fit.}

\item{\code{meanAbsoluteError}}{The fit's mean absolute error.}

\item{\code{rootMeanSquareError}}{The fit's RMS error.}

\item{\code{estimatedUltimateRecovery}}{The EUR calculated from the fit.}

\item{\code{remainingReserves}}{The remaining reserves based on the fit.}

\item{\code{economicLimit}}{The economic limit for the calculation of EUR and
remaining reserves.}

\item{\code{leftBehind}}{The reserves that are left behind.}

\item{\code{nowDate}}{The current date.}

\item{\code{forecastEndDate}}{The end date of the forecast.}

\item{\code{isLocked}}{This flag specifies whether the DCA fit is locked.
Defaults to \code{FALSE}.}

\item{\code{user}}{The user the DCA fit belongs to.}

\item{\code{isFavorite}}{This flag specifies whether the DCA fit is marked as
favorite item, and thus shown in the favorites tab on the home module
in PetroVisor. Defaults to \code{FALSE}.}

\item{\code{labels}}{A list of strings holding the labels of the DCA fit.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DcaFit-toList"></a>}}
\if{latex}{\out{\hypertarget{method-DcaFit-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DcaFit$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DcaFit-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DcaFit-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DcaFit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
