% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntityType.R
\name{EntityType}
\alias{EntityType}
\title{EntityType}
\description{
Class representing a PetroVisor entityType object.
}
\examples{
\dontrun{
EntityType$new(name = "Section")
}
}
\seealso{
\itemize{
\item \link{Entity} for entity definitions
\item \link{RepositoryService} for loading entity types
\item \code{vignette("repository-service")} for examples
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the entity type.}

\item{\code{image}}{The image of the entity type represented as string.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EntityType-new}{\code{EntityType$new()}}
\item \href{#method-EntityType-toList}{\code{EntityType$toList()}}
\item \href{#method-EntityType-clone}{\code{EntityType$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EntityType-new"></a>}}
\if{latex}{\out{\hypertarget{method-EntityType-new}{}}}
\subsection{Method \code{new()}}{
Create a new EntityType instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EntityType$new(name = NULL, image = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of entity type.}

\item{\code{image}}{The image of the entity type represented as string.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EntityType-toList"></a>}}
\if{latex}{\out{\hypertarget{method-EntityType-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EntityType$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EntityType-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EntityType-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EntityType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
