% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculation.R
\name{Event}
\alias{Event}
\title{Event}
\description{
Class representing a event object.
}
\examples{
\dontrun{
Event$new()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the event.}

\item{\code{formula}}{The event's formula.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Event-new}{\code{Event$new()}}
\item \href{#method-Event-toList}{\code{Event$toList()}}
\item \href{#method-Event-clone}{\code{Event$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Event-new"></a>}}
\if{latex}{\out{\hypertarget{method-Event-new}{}}}
\subsection{Method \code{new()}}{
Create a new Event instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Event$new(name = NULL, formula = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the cleansing filter.}

\item{\code{formula}}{The cleansing filter's formula.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Event-toList"></a>}}
\if{latex}{\out{\hypertarget{method-Event-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Event$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Event-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Event-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Event$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
