% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculation.R
\name{EventCalculation}
\alias{EventCalculation}
\title{EventCalculation}
\description{
Class representing a PetroVisor event calculation object.
}
\examples{
\dontrun{
EventCalculation$new()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the event calculation.}

\item{\code{events}}{A list of events of the event calculation. Each item has
to be an instance of the class \code{Event}.}

\item{\code{inputTableNames}}{A list containing the names of all input tables for
the event calculation.}

\item{\code{isLocked}}{This flag specifies whether the event calculation is
locked. Defaults to \code{FALSE}.}

\item{\code{user}}{The user the event calculation belongs to.}

\item{\code{isFavorite}}{This flag specifies whether the event calculation is
marked as favorite item, and thus shown in the favorites tab on the home
module in PetroVisor. Defaults to \code{FALSE}.}

\item{\code{labels}}{A list of strings holding the labels of the event
calculation.}

\item{\code{formula}}{The event calculation's formula as string.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EventCalculation-new}{\code{EventCalculation$new()}}
\item \href{#method-EventCalculation-toList}{\code{EventCalculation$toList()}}
\item \href{#method-EventCalculation-clone}{\code{EventCalculation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventCalculation-new"></a>}}
\if{latex}{\out{\hypertarget{method-EventCalculation-new}{}}}
\subsection{Method \code{new()}}{
Create a new EventCalculation instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventCalculation$new(
  name = NULL,
  events = NULL,
  inputTableNames = NULL,
  isLocked = FALSE,
  user = NULL,
  isFavorite = FALSE,
  labels = NULL,
  formula = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the event calculation.}

\item{\code{events}}{A list of events of the event calculation. Each item has
to be an instance of the class \code{Event}.}

\item{\code{inputTableNames}}{A list containing the names of all input tables for
the event calculation.}

\item{\code{isLocked}}{This flag specifies whether the event calculation is
locked. Defaults to \code{FALSE}.}

\item{\code{user}}{The user the event calculation belongs to.}

\item{\code{isFavorite}}{This flag specifies whether the event calculation is
marked as favorite item, and thus shown in the favorites tab on the home
module in PetroVisor. Defaults to \code{FALSE}.}

\item{\code{labels}}{A list of strings holding the labels of the event
calculation.}

\item{\code{formula}}{The event calculation's formula as string.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventCalculation-toList"></a>}}
\if{latex}{\out{\hypertarget{method-EventCalculation-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventCalculation$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventCalculation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EventCalculation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventCalculation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
