% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\name{Filter}
\alias{Filter}
\title{Filter}
\description{
Class representing a PetroVisor filter object.
}
\examples{
\dontrun{
Filter$new(name = "MyFilter",
           entitySetName = "All Wells",
           checkedSignalNames = list("oil rate", "gas rate"),
           checkedUnitNames = list("m3/d", "m3/d"),
           start = "2020-01-01T00:00:00.000Z",
           end = "2020-03-01T00:00:00.000Z",
           step = "Daily")
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the filter.}

\item{\code{entityNamePattern}}{The pattern to filter the available entities by.}

\item{\code{signalNamePattern}}{The pattern to filter the available signals by.}

\item{\code{entitySetName}}{The name of the selected entity set.}

\item{\code{checkedEntityNames}}{The names of the checked entities.}

\item{\code{checkedSignalNames}}{The names of the checked signals.}

\item{\code{checkedUnitNames}}{The names of the checked units.}

\item{\code{start}}{The filter's start date.}

\item{\code{end}}{The filter's end date.}

\item{\code{step}}{The filter's time increment. Allowed values are:
\code{"EveryMinute"}, \code{"EveryFiveMinutes"},
\code{"EveryFifteenMinutes"}, \code{"Hourly"}, \code{"Daily"},
\code{"Monthly"}, \code{"Quarterly"}, \code{"Yearly"}}

\item{\code{isLocked}}{This flag specifies whether the filter is locked.
Defaults to \code{FALSE}.}

\item{\code{user}}{The user the filter belongs to.}

\item{\code{isFavorite}}{This flag specifies whether the filter is marked as
favorite item, and thus shown in the favorites tab on the home module in
PetroVisor. Defaults to \code{FALSE}.}

\item{\code{labels}}{A list of strings holding the labels of the filter.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Filter-new}{\code{Filter$new()}}
\item \href{#method-Filter-toList}{\code{Filter$toList()}}
\item \href{#method-Filter-clone}{\code{Filter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Filter-new"></a>}}
\if{latex}{\out{\hypertarget{method-Filter-new}{}}}
\subsection{Method \code{new()}}{
Create a new Filter instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Filter$new(
  name = NULL,
  entityNamePattern = NULL,
  signalNamePattern = NULL,
  entitySetName = NULL,
  checkedEntityNames = NULL,
  checkedSignalNames = NULL,
  checkedUnitNames = NULL,
  start = NULL,
  end = NULL,
  step = c("EveryMinute", "EveryFiveMinutes", "EveryFifteenMinutes", "Hourly", "Daily",
    "Monthly", "Quarterly", "Yearly"),
  isLocked = FALSE,
  user = NULL,
  isFavorite = FALSE,
  labels = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the filter.}

\item{\code{entityNamePattern}}{The pattern to filter the available entities by.}

\item{\code{signalNamePattern}}{The pattern to filter the available signals by.}

\item{\code{entitySetName}}{The name of the selected entity set.}

\item{\code{checkedEntityNames}}{The names of the checked entities.}

\item{\code{checkedSignalNames}}{The names of the checked signals.}

\item{\code{checkedUnitNames}}{The names of the checked units.}

\item{\code{start}}{The filter's start date.}

\item{\code{end}}{The filter's end date.}

\item{\code{step}}{The filter's time increment. Allowed values are:
\code{"EveryMinute"}, \code{"EveryFiveMinutes"},
\code{"EveryFifteenMinutes"}, \code{"Hourly"}, \code{"Daily"},
\code{"Monthly"}, \code{"Quarterly"}, \code{"Yearly"}}

\item{\code{isLocked}}{This flag specifies whether the filter is locked.
Defaults to \code{FALSE}.}

\item{\code{user}}{The user the filter belongs to.}

\item{\code{isFavorite}}{This flag specifies whether the filter is marked as
favorite item, and thus shown in the favorites tab on the home module
in PetroVisor. Defaults to \code{FALSE}.}

\item{\code{labels}}{A list of strings holding the labels of the filter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Filter-toList"></a>}}
\if{latex}{\out{\hypertarget{method-Filter-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Filter$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Filter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Filter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Filter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
