% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hierarchy.R
\name{Hierarchy}
\alias{Hierarchy}
\title{Hierarchy}
\description{
Class representing a PetroVisor hierarchy object.
}
\examples{
\dontrun{
Hierarchy$new(name = "MyHierarchy",
              relationship = list(Well1 = "Parent1",
                                  Well2 = "Parent1",
                                  Parent1 = NA))
}
}
\seealso{
\itemize{
\item \link{RepositoryService} for loading and saving hierarchies
\item \link{Context} for using hierarchies in data contexts
\item \link{Entity} for entity definitions
\item \code{vignette("repository-service")} for hierarchy examples
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the hierarchy.}

\item{\code{relationship}}{The child-parent relationships stored in the hierarchy.
A single data frame with the columns \code{child} and \code{parent} if the
hierarchy is static. For time-dependent hierarchies a list of data frames.
The names of the list refer to the date of the relationship.}

\item{\code{is_time_dependent}}{Whether the hierarchy is time dependent or static.}

\item{\code{time_stamp}}{The (first) time stamp of the time dependent hierarchy.}

\item{\code{description}}{The description of the item.}

\item{\code{labels}}{A list of strings holding the labels of the scope.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Hierarchy-new}{\code{Hierarchy$new()}}
\item \href{#method-Hierarchy-toList}{\code{Hierarchy$toList()}}
\item \href{#method-Hierarchy-clone}{\code{Hierarchy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Hierarchy-new"></a>}}
\if{latex}{\out{\hypertarget{method-Hierarchy-new}{}}}
\subsection{Method \code{new()}}{
Create a new Scope instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hierarchy$new(
  name = NULL,
  relationship = NULL,
  is_time_dependent = FALSE,
  time_stamp = NULL,
  description = NULL,
  labels = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the hierarchy.}

\item{\code{relationship}}{The child-parent relationships stored in the
hierarchy. A single data frame with the columns \code{child} and
\code{parent} if the hierarchy is static. For time-dependent
hierarchies a list of data frames. The names of the list refer to the
date of the relationship.}

\item{\code{is_time_dependent}}{Whether the hierarchy is time dependent or
static.}

\item{\code{time_stamp}}{The (first) time stamp of the time dependent hierarchy.}

\item{\code{description}}{The description of the item.}

\item{\code{labels}}{A list of strings holding the labels of the scope.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Hierarchy-toList"></a>}}
\if{latex}{\out{\hypertarget{method-Hierarchy-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hierarchy$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Hierarchy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Hierarchy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hierarchy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
