% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MachineLearning.R
\name{MLCustomTrainingOptions}
\alias{MLCustomTrainingOptions}
\title{MLCustomTrainingOptions}
\description{
Options for custom trainers.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{l2_regularization}}{OLS: L2 regularization weight (0-1). Adding L2
regularization turns this algorithm into a form of ridge regression.}

\item{\code{enable_pruning}}{GAM: Enable post-training tree pruning to avoid
overfitting. Requires a validation set.}

\item{\code{entropy_coefficient}}{GAM: The entropy (regularization) coefficient
between 0 and 1.}

\item{\code{learning_rate}}{GAM: The learning rate (0-1).}

\item{\code{maximum_bin_count_per_feature}}{GAM: The maximum number of distinct
values (bins) per feature (2-500).}

\item{\code{minimum_example_count_per_leaf}}{GAM: The minimal number of data points
required to form a new tree leaf (1-100).}

\item{\code{number_of_iterations}}{GAM: Total number of passes over the training
data (100-20000).}

\item{\code{pruning_metrics}}{GAM: Determines what metric to use for pruning.
One of \code{"LeastAbsoluteDeviation"}, \code{"LeastSquares"}.}

\item{\code{rank}}{Randomized PCA: The number of components in the PCA.}

\item{\code{oversampling}}{Randomized PCA: Oversampling parameter for randomized
PCA training (1-200).}

\item{\code{ensure_zero_mean}}{Randomized PCA: If true, data is centered to have
zero mean. Defaults to true.}

\item{\code{onnx_model}}{Gets or sets the ONNX model (string, nullable).}

\item{\code{neural_network_training_options}}{Neural Network Training Options.
Object of type MLNeuralNetworkTrainingOptions.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MLCustomTrainingOptions-new}{\code{MLCustomTrainingOptions$new()}}
\item \href{#method-MLCustomTrainingOptions-toList}{\code{MLCustomTrainingOptions$toList()}}
\item \href{#method-MLCustomTrainingOptions-clone}{\code{MLCustomTrainingOptions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLCustomTrainingOptions-new"></a>}}
\if{latex}{\out{\hypertarget{method-MLCustomTrainingOptions-new}{}}}
\subsection{Method \code{new()}}{
Create a new MLCustomTrainingOptions instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLCustomTrainingOptions$new(
  l2_regularization = NULL,
  enable_pruning = NULL,
  entropy_coefficient = NULL,
  learning_rate = NULL,
  maximum_bin_count_per_feature = NULL,
  minimum_example_count_per_leaf = NULL,
  number_of_iterations = NULL,
  pruning_metrics = NULL,
  rank = NULL,
  oversampling = NULL,
  ensure_zero_mean = TRUE,
  onnx_model = NULL,
  neural_network_training_options = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{l2_regularization}}{OLS: L2 regularization weight (0-1). Adding L2
regularization turns this algorithm into a form of ridge regression.}

\item{\code{enable_pruning}}{GAM: Enable post-training tree pruning to avoid
overfitting. Requires a validation set.}

\item{\code{entropy_coefficient}}{GAM: The entropy (regularization) coefficient
between 0 and 1.}

\item{\code{learning_rate}}{GAM: The learning rate (0-1).}

\item{\code{maximum_bin_count_per_feature}}{GAM: The maximum number of distinct
values (bins) per feature (2-500).}

\item{\code{minimum_example_count_per_leaf}}{GAM: The minimal number of data
points required to form a new tree leaf (1-100).}

\item{\code{number_of_iterations}}{GAM: Total number of passes over the training
data (100-20000).}

\item{\code{pruning_metrics}}{GAM: Determines what metric to use for pruning.
One of \code{"LeastAbsoluteDeviation"}, \code{"LeastSquares"}.}

\item{\code{rank}}{Randomized PCA: The number of components in the PCA.}

\item{\code{oversampling}}{Randomized PCA: Oversampling parameter for randomized
PCA training (1-200).}

\item{\code{ensure_zero_mean}}{Randomized PCA: If true, data is centered to have
zero mean. Defaults to true.}

\item{\code{onnx_model}}{Gets or sets the ONNX model (string, nullable).}

\item{\code{neural_network_training_options}}{Neural Network Training Options.
Object of type MLNeuralNetworkTrainingOptions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLCustomTrainingOptions-toList"></a>}}
\if{latex}{\out{\hypertarget{method-MLCustomTrainingOptions-toList}{}}}
\subsection{Method \code{toList()}}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLCustomTrainingOptions$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representation of the MLCustomTrainingOptions object
compatible with the API.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLCustomTrainingOptions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MLCustomTrainingOptions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLCustomTrainingOptions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
