% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MachineLearning.R
\name{MLDataProviderConfig}
\alias{MLDataProviderConfig}
\title{MLDataProviderConfig}
\description{
Configuration for ML data providers.
}
\examples{
\dontrun{
# Create a MLDataProviderConfig
MLDataProviderConfig$new(
  provider_type = "ReferenceTable",
  source_name = "My Source Table",
  storage_name = "My Target Table",
  columns = list()
)
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{provider_type}}{The data provider type (PSharp or ReferenceTable).}

\item{\code{source_name}}{The name of the data source.}

\item{\code{storage_name}}{The name of the data storage.}

\item{\code{columns}}{The configuration of the data columns. List of MLDataColumns.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MLDataProviderConfig-new}{\code{MLDataProviderConfig$new()}}
\item \href{#method-MLDataProviderConfig-toList}{\code{MLDataProviderConfig$toList()}}
\item \href{#method-MLDataProviderConfig-clone}{\code{MLDataProviderConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLDataProviderConfig-new"></a>}}
\if{latex}{\out{\hypertarget{method-MLDataProviderConfig-new}{}}}
\subsection{Method \code{new()}}{
Create a new MLDataProviderConfig instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLDataProviderConfig$new(
  provider_type = c("PSharp", "ReferenceTable"),
  source_name = NULL,
  storage_name = NULL,
  columns = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{provider_type}}{The data provider type (PSharp or ReferenceTable).}

\item{\code{source_name}}{The name of the data source.}

\item{\code{storage_name}}{The name of the data storage.}

\item{\code{columns}}{The configuration of the data columns. List of
MLDataColumns.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLDataProviderConfig-toList"></a>}}
\if{latex}{\out{\hypertarget{method-MLDataProviderConfig-toList}{}}}
\subsection{Method \code{toList()}}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLDataProviderConfig$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representation of the MLDataProviderConfig object
compatible with the API.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLDataProviderConfig-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MLDataProviderConfig-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLDataProviderConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
