% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MachineLearning.R
\name{MLNeuralNetworkTrainingOptions}
\alias{MLNeuralNetworkTrainingOptions}
\title{MLNeuralNetworkTrainingOptions}
\description{
Represents ML neural network training options.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{library}}{The name of the neural network library.}

\item{\code{model_path}}{The path to the model.}

\item{\code{transfer_learning}}{Whether to use transfer learning.}

\item{\code{hyperparameter_tuning}}{Whether to tune hyperparameters.}

\item{\code{features_mapping}}{Features and label names mapping from internal names
to trained model names.}

\item{\code{optimizer}}{The optimizer to use.}

\item{\code{learning_rate}}{The learning rate used (value between 0 and 1).}

\item{\code{epochs}}{The number of training iterations.}

\item{\code{batch_size}}{The number of samples to use for mini-batch training.}

\item{\code{sequence_window_size}}{Series sequence window size.}

\item{\code{activation_function}}{The activation function.}

\item{\code{number_of_dense_layers}}{Number of dense layers.}

\item{\code{number_of_neurons}}{Number of neurons.}

\item{\code{number_of_conv1d_layers}}{Number of Conv1D layers.}

\item{\code{number_of_conv1d_filters}}{Number of Conv1D filters.}

\item{\code{conv1d_kernel_size}}{Conv1D kernel size.}

\item{\code{number_of_lstm_layers}}{Number of LSTM layers.}

\item{\code{number_of_lstm_neurons}}{Number of LSTM neurons.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MLNeuralNetworkTrainingOptions-new}{\code{MLNeuralNetworkTrainingOptions$new()}}
\item \href{#method-MLNeuralNetworkTrainingOptions-toList}{\code{MLNeuralNetworkTrainingOptions$toList()}}
\item \href{#method-MLNeuralNetworkTrainingOptions-clone}{\code{MLNeuralNetworkTrainingOptions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLNeuralNetworkTrainingOptions-new"></a>}}
\if{latex}{\out{\hypertarget{method-MLNeuralNetworkTrainingOptions-new}{}}}
\subsection{Method \code{new()}}{
Create a new MLNeuralNetworkTrainingOptions instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLNeuralNetworkTrainingOptions$new(
  library = NULL,
  model_path = NULL,
  transfer_learning = NULL,
  hyperparameter_tuning = NULL,
  features_mapping = NULL,
  optimizer = NULL,
  learning_rate = NULL,
  epochs = NULL,
  batch_size = NULL,
  sequence_window_size = NULL,
  activation_function = NULL,
  number_of_dense_layers = NULL,
  number_of_neurons = NULL,
  number_of_conv1d_layers = NULL,
  number_of_conv1d_filters = NULL,
  conv1d_kernel_size = NULL,
  number_of_lstm_layers = NULL,
  number_of_lstm_neurons = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{library}}{The name of the neural network library.}

\item{\code{model_path}}{The path to the model.}

\item{\code{transfer_learning}}{Whether to use transfer learning.}

\item{\code{hyperparameter_tuning}}{Whether to tune hyperparameters.}

\item{\code{features_mapping}}{Features and label names mapping from internal
names to trained model names.}

\item{\code{optimizer}}{The optimizer to use.}

\item{\code{learning_rate}}{The learning rate used (value between 0 and 1).}

\item{\code{epochs}}{The number of training iterations.}

\item{\code{batch_size}}{The number of samples to use for mini-batch training.}

\item{\code{sequence_window_size}}{Series sequence window size.}

\item{\code{activation_function}}{The activation function.}

\item{\code{number_of_dense_layers}}{Number of dense layers.}

\item{\code{number_of_neurons}}{Number of neurons.}

\item{\code{number_of_conv1d_layers}}{Number of Conv1D layers.}

\item{\code{number_of_conv1d_filters}}{Number of Conv1D filters.}

\item{\code{conv1d_kernel_size}}{Conv1D kernel size.}

\item{\code{number_of_lstm_layers}}{Number of LSTM layers.}

\item{\code{number_of_lstm_neurons}}{Number of LSTM neurons.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLNeuralNetworkTrainingOptions-toList"></a>}}
\if{latex}{\out{\hypertarget{method-MLNeuralNetworkTrainingOptions-toList}{}}}
\subsection{Method \code{toList()}}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLNeuralNetworkTrainingOptions$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representation of the MLNeuralNetworkTrainingOptions
object compatible with the API.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLNeuralNetworkTrainingOptions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MLNeuralNetworkTrainingOptions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLNeuralNetworkTrainingOptions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
