% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MachineLearning.R
\name{MLPreProcessor}
\alias{MLPreProcessor}
\title{MLPreProcessor}
\description{
Represents a ML data pre-processor.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{normalization_type}}{The type of the pre-processor.}

\item{\code{transformer_options}}{Options for the pre-processor. Object of type
MLTransformerOptions.}

\item{\code{order}}{Ordinal number.}

\item{\code{is_enabled}}{Whether the pre-processor is enabled.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MLPreProcessor-new}{\code{MLPreProcessor$new()}}
\item \href{#method-MLPreProcessor-toList}{\code{MLPreProcessor$toList()}}
\item \href{#method-MLPreProcessor-clone}{\code{MLPreProcessor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLPreProcessor-new"></a>}}
\if{latex}{\out{\hypertarget{method-MLPreProcessor-new}{}}}
\subsection{Method \code{new()}}{
Create a new MLPreProcessor instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLPreProcessor$new(
  normalization_type = c("MinMax", "MeanVariance", "LogMeanVariance", "Binning",
    "SupervisedBinning", "RobustScaling", "LpNorm", "GlobalContrast",
    "ProjectToPrincipalComponents", "ApproximatedKernelMap", "BoxCox", "BoxTidwell"),
  transformer_options = NULL,
  order = NULL,
  is_enabled = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{normalization_type}}{The type of the pre-processor.}

\item{\code{transformer_options}}{Options for the pre-processor. Object of type
MLTransformerOptionsation.}

\item{\code{order}}{Ordinal number.}

\item{\code{is_enabled}}{Whether the pre-processor is enabled.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLPreProcessor-toList"></a>}}
\if{latex}{\out{\hypertarget{method-MLPreProcessor-toList}{}}}
\subsection{Method \code{toList()}}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLPreProcessor$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representation of the MLPreProcessor object compatible
with the API.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLPreProcessor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MLPreProcessor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLPreProcessor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
