% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MachineLearning.R
\name{MLTrainingContext}
\alias{MLTrainingContext}
\title{MLTrainingContext}
\description{
Represents ML training context information.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{validation_data}}{Serialized validation data (byte format).}

\item{\code{test_data}}{Serialized test data (byte format).}

\item{\code{test_data_fraction}}{Test data fraction (nullable).}

\item{\code{validation_data_fraction}}{Validation data fraction (nullable).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MLTrainingContext-new}{\code{MLTrainingContext$new()}}
\item \href{#method-MLTrainingContext-toList}{\code{MLTrainingContext$toList()}}
\item \href{#method-MLTrainingContext-clone}{\code{MLTrainingContext$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLTrainingContext-new"></a>}}
\if{latex}{\out{\hypertarget{method-MLTrainingContext-new}{}}}
\subsection{Method \code{new()}}{
Create a new MLTrainingContext instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLTrainingContext$new(
  validation_data = NULL,
  test_data = NULL,
  test_data_fraction = NULL,
  validation_data_fraction = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{validation_data}}{Serialized validation data (byte format).}

\item{\code{test_data}}{Serialized test data (byte format).}

\item{\code{test_data_fraction}}{Test data fraction (nullable).}

\item{\code{validation_data_fraction}}{Validation data fraction (nullable).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLTrainingContext-toList"></a>}}
\if{latex}{\out{\hypertarget{method-MLTrainingContext-toList}{}}}
\subsection{Method \code{toList()}}{
Convert the object to a list for API compatibility.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLTrainingContext$toList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLTrainingContext-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MLTrainingContext-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLTrainingContext$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
