% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MachineLearning.R
\name{MLTrainingFeature}
\alias{MLTrainingFeature}
\title{MLTrainingFeature}
\description{
Represents a feature with importance and weights in ML training
results.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the feature.}

\item{\code{importance}}{The importance value of the feature.}

\item{\code{weights}}{List of MLTrainingNumericOutcome objects.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MLTrainingFeature-new}{\code{MLTrainingFeature$new()}}
\item \href{#method-MLTrainingFeature-toList}{\code{MLTrainingFeature$toList()}}
\item \href{#method-MLTrainingFeature-clone}{\code{MLTrainingFeature$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLTrainingFeature-new"></a>}}
\if{latex}{\out{\hypertarget{method-MLTrainingFeature-new}{}}}
\subsection{Method \code{new()}}{
Create a new MLTrainingFeature instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLTrainingFeature$new(name = NULL, importance = NULL, weights = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the feature.}

\item{\code{importance}}{The importance value of the feature.}

\item{\code{weights}}{List of MLTrainingNumericOutcome objects.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLTrainingFeature-toList"></a>}}
\if{latex}{\out{\hypertarget{method-MLTrainingFeature-toList}{}}}
\subsection{Method \code{toList()}}{
Convert the object to a list for API compatibility.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLTrainingFeature$toList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLTrainingFeature-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MLTrainingFeature-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLTrainingFeature$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
