% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MachineLearning.R
\name{MLTrainingNumericOutcome}
\alias{MLTrainingNumericOutcome}
\title{MLTrainingNumericOutcome}
\description{
Class representing a numeric training outcome for machine
learning models. Contains the name and value of a training metric.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the outcome (required, string, max 255 chars).}

\item{\code{value}}{The value of the metric (numeric, optional).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MLTrainingNumericOutcome-new}{\code{MLTrainingNumericOutcome$new()}}
\item \href{#method-MLTrainingNumericOutcome-toList}{\code{MLTrainingNumericOutcome$toList()}}
\item \href{#method-MLTrainingNumericOutcome-clone}{\code{MLTrainingNumericOutcome$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLTrainingNumericOutcome-new"></a>}}
\if{latex}{\out{\hypertarget{method-MLTrainingNumericOutcome-new}{}}}
\subsection{Method \code{new()}}{
Create a new MLTrainingNumericOutcome instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLTrainingNumericOutcome$new(name, value = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the outcome (required, string, max 255 chars).}

\item{\code{value}}{The value of the metric (numeric, optional).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLTrainingNumericOutcome-toList"></a>}}
\if{latex}{\out{\hypertarget{method-MLTrainingNumericOutcome-toList}{}}}
\subsection{Method \code{toList()}}{
Convert the object to a list for API compatibility.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLTrainingNumericOutcome$toList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLTrainingNumericOutcome-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MLTrainingNumericOutcome-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLTrainingNumericOutcome$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
