% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MachineLearning.R
\name{MLTrainingOptions}
\alias{MLTrainingOptions}
\title{MLTrainingOptions}
\description{
ML model training options.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{test_fraction}}{Test set fraction.}

\item{\code{test_latin_hypercube}}{Logical indicating whether the test set sampling
is latin hypercube (if true) or uniform (otherwise).}

\item{\code{validation_fraction}}{Validation set fraction.}

\item{\code{optimization_metric}}{Name of the optimization metric.}

\item{\code{time_to_train}}{Maximal time to train in seconds.}

\item{\code{num_clusters}}{Number of clusters (when applicable).}

\item{\code{num_cv_folds}}{Number of cross validation folds.}

\item{\code{trainers_to_exclude}}{The names of the trainers (algorithms) to exclude.}

\item{\code{scope}}{The training scope. Object of type Scope.}

\item{\code{entity_set}}{The training entity set. Object of type EntitySet.}

\item{\code{include_incomplete_cases}}{Logical whether to include incomplete cases
(rows with some features equal to null) for training.}

\item{\code{custom_training_type}}{Customized training type. One of Auto, Ols, Gam,
RandomizedPca, GamOva, BernoulliNaiveBayes, OnnxModel, NeuralNetwork.}

\item{\code{custom_training_options}}{Options for custom trainer. Object of type
MLCustomTrainingOptions.}

\item{\code{pre_processors}}{Data pre-processors. List of MLPreProcessor.}

\item{\code{apply_pre_processors_before_training}}{When to apply pre-processors:
before (true) or after (false) training.}

\item{\code{tuner}}{Tuner name.}

\item{\code{maximum_models_to_train}}{Maximum models to train.}

\item{\code{survival_data}}{Names of the columns containing survival data.}

\item{\code{trainer_hyperparameters}}{Trainer hyperparameters.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MLTrainingOptions-new}{\code{MLTrainingOptions$new()}}
\item \href{#method-MLTrainingOptions-toList}{\code{MLTrainingOptions$toList()}}
\item \href{#method-MLTrainingOptions-clone}{\code{MLTrainingOptions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLTrainingOptions-new"></a>}}
\if{latex}{\out{\hypertarget{method-MLTrainingOptions-new}{}}}
\subsection{Method \code{new()}}{
Create a new MLTrainingOptions instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLTrainingOptions$new(
  scope,
  entity_set,
  test_fraction = NULL,
  test_latin_hypercube = FALSE,
  validation_fraction = NULL,
  optimization_metric = NULL,
  time_to_train = 60,
  num_clusters = 1,
  num_cv_folds = 1,
  trainers_to_exclude = NULL,
  include_incomplete_cases = FALSE,
  custom_training_type = "Auto",
  custom_training_options = NULL,
  pre_processors = NULL,
  apply_pre_processors_before_training = TRUE,
  tuner = "EciCostFrugal",
  maximum_models_to_train = 1000,
  survival_data = NULL,
  trainer_hyperparameters = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scope}}{The training scope. Object of type Scope (required).}

\item{\code{entity_set}}{The training entity set. Object of type EntitySet
(required).}

\item{\code{test_fraction}}{Test set fraction.}

\item{\code{test_latin_hypercube}}{Logical indicating whether the test set
sampling is latin hypercube (if true) or uniform (otherwise).}

\item{\code{validation_fraction}}{Validation set fraction.}

\item{\code{optimization_metric}}{Name of the optimization metric.}

\item{\code{time_to_train}}{Maximal time to train in seconds.}

\item{\code{num_clusters}}{Number of clusters (when applicable).}

\item{\code{num_cv_folds}}{Number of cross validation folds.}

\item{\code{trainers_to_exclude}}{The names of the trainers (algorithms) to
exclude.}

\item{\code{include_incomplete_cases}}{Logical whether to include incomplete
cases (rows with some features equal to null) for training.}

\item{\code{custom_training_type}}{Customized training type. One of Auto, Ols,
Gam, RandomizedPca, GamOva, BernoulliNaiveBayes, OnnxModel,
NeuralNetwork.}

\item{\code{custom_training_options}}{Options for custom trainer. Object of type
MLCustomTrainingOptions.}

\item{\code{pre_processors}}{Data pre-processors. List of MLPreProcessor.}

\item{\code{apply_pre_processors_before_training}}{When to apply pre-processors:
before (true) or after (false) training.}

\item{\code{tuner}}{Tuner name.}

\item{\code{maximum_models_to_train}}{Maximum models to train.}

\item{\code{survival_data}}{Names of the columns containing survival data.}

\item{\code{trainer_hyperparameters}}{Trainer hyperparameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLTrainingOptions-toList"></a>}}
\if{latex}{\out{\hypertarget{method-MLTrainingOptions-toList}{}}}
\subsection{Method \code{toList()}}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLTrainingOptions$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representation of the MLTrainingOptions object compatible
with the API.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLTrainingOptions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MLTrainingOptions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLTrainingOptions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
