% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MachineLearning.R
\name{MLTrainingOutcome}
\alias{MLTrainingOutcome}
\title{MLTrainingOutcome}
\description{
Represents the outcome metrics from ML training.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{is_best_model}}{Whether this is the best model among others.}

\item{\code{is_best_run}}{Whether this is the best run (for cross validation folds).}

\item{\code{group_by}}{Group TrainingOutcome instances by the value of this
property (for cross validation folds).}

\item{\code{sequence}}{Sequential number (the order this model was trained by
auto-ML).}

\item{\code{values}}{List of training metric values measured on the validation
set.}

\item{\code{test_values}}{List of test metric values measured on the test set.}

\item{\code{is_best_model_test}}{Whether this is the best among others, measured
on the test set.}

\item{\code{is_best_run_test}}{Whether this is the best run (for cross validation
folds), measured on the test set.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MLTrainingOutcome-new}{\code{MLTrainingOutcome$new()}}
\item \href{#method-MLTrainingOutcome-toList}{\code{MLTrainingOutcome$toList()}}
\item \href{#method-MLTrainingOutcome-clone}{\code{MLTrainingOutcome$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLTrainingOutcome-new"></a>}}
\if{latex}{\out{\hypertarget{method-MLTrainingOutcome-new}{}}}
\subsection{Method \code{new()}}{
Create a new MLTrainingOutcome instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLTrainingOutcome$new(
  is_best_model = FALSE,
  is_best_run = FALSE,
  group_by = NULL,
  sequence = NULL,
  values = list(),
  test_values = list(),
  is_best_model_test = FALSE,
  is_best_run_test = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{is_best_model}}{Whether this is the best model among others.}

\item{\code{is_best_run}}{Whether this is the best run (for cross validation
folds).}

\item{\code{group_by}}{Group TrainingOutcome instances by the value of this
property (for cross validation folds).}

\item{\code{sequence}}{Sequential number (the order this model was trained by
auto-ML).}

\item{\code{values}}{List of training metric values measured on the validation
set.}

\item{\code{test_values}}{List of test metric values measured on the test set.}

\item{\code{is_best_model_test}}{Whether this is the best among others,
measured on the test set.}

\item{\code{is_best_run_test}}{Whether this is the best run (for cross
validation folds), measured on the test set.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLTrainingOutcome-toList"></a>}}
\if{latex}{\out{\hypertarget{method-MLTrainingOutcome-toList}{}}}
\subsection{Method \code{toList()}}{
Convert the object to a list for API compatibility.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLTrainingOutcome$toList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLTrainingOutcome-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MLTrainingOutcome-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLTrainingOutcome$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
