% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MachineLearning.R
\name{MLTrainingResult}
\alias{MLTrainingResult}
\title{MLTrainingResult}
\description{
Represents a single training result.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{entity_name}}{Entity name (nullable).}

\item{\code{trained_model}}{Trained model (byte format).}

\item{\code{training_context}}{Model training context (MLTrainingContext object).}

\item{\code{confusion_matrix}}{Confusion matrix for classification models.}

\item{\code{outcome}}{Training metrics (MLTrainingOutcome object).}

\item{\code{error}}{Training error (nullable).}

\item{\code{trainer_name}}{Trainer name.}

\item{\code{features}}{List of MLTrainingFeature objects.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MLTrainingResult-new}{\code{MLTrainingResult$new()}}
\item \href{#method-MLTrainingResult-toList}{\code{MLTrainingResult$toList()}}
\item \href{#method-MLTrainingResult-clone}{\code{MLTrainingResult$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLTrainingResult-new"></a>}}
\if{latex}{\out{\hypertarget{method-MLTrainingResult-new}{}}}
\subsection{Method \code{new()}}{
Create a new MLTrainingResult instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLTrainingResult$new(
  entity_name = NULL,
  trained_model = NULL,
  training_context = NULL,
  confusion_matrix = NULL,
  outcome = NULL,
  error = NULL,
  trainer_name = NULL,
  features = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{entity_name}}{Entity name (nullable).}

\item{\code{trained_model}}{Trained model (byte format).}

\item{\code{training_context}}{Model training context
(MLTrainingContext object).}

\item{\code{confusion_matrix}}{Confusion matrix for classification models.}

\item{\code{outcome}}{Training metrics (MLTrainingOutcome object).}

\item{\code{error}}{Training error (nullable).}

\item{\code{trainer_name}}{Trainer name.}

\item{\code{features}}{List of MLTrainingFeature objects.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLTrainingResult-toList"></a>}}
\if{latex}{\out{\hypertarget{method-MLTrainingResult-toList}{}}}
\subsection{Method \code{toList()}}{
Convert the object to a list for API compatibility.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLTrainingResult$toList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLTrainingResult-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MLTrainingResult-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLTrainingResult$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
