% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MachineLearning.R
\name{MLTrainingResults}
\alias{MLTrainingResults}
\title{MLTrainingResults}
\description{
Represents the complete ML training results collection.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{runs_metrics_group_by}}{Training runs grouping (nullable).}

\item{\code{results}}{List of MLTrainingResult objects.}

\item{\code{label_unit_name}}{Unit name for the label.}

\item{\code{status}}{The training status (added by API wrapper).}

\item{\code{request_id}}{The training request ID (added by API wrapper).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MLTrainingResults-new}{\code{MLTrainingResults$new()}}
\item \href{#method-MLTrainingResults-get_best_models}{\code{MLTrainingResults$get_best_models()}}
\item \href{#method-MLTrainingResults-toList}{\code{MLTrainingResults$toList()}}
\item \href{#method-MLTrainingResults-clone}{\code{MLTrainingResults$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLTrainingResults-new"></a>}}
\if{latex}{\out{\hypertarget{method-MLTrainingResults-new}{}}}
\subsection{Method \code{new()}}{
Create a new MLTrainingResults instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLTrainingResults$new(
  runs_metrics_group_by = NULL,
  results = list(),
  label_unit_name = NULL,
  status = NULL,
  request_id = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{runs_metrics_group_by}}{Training runs grouping (nullable).}

\item{\code{results}}{List of MLTrainingResult objects.}

\item{\code{label_unit_name}}{Unit name for the label.}

\item{\code{status}}{The training status (added by API wrapper).}

\item{\code{request_id}}{The training request ID (added by API wrapper).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLTrainingResults-get_best_models"></a>}}
\if{latex}{\out{\hypertarget{method-MLTrainingResults-get_best_models}{}}}
\subsection{Method \code{get_best_models()}}{
Get the best models from the results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLTrainingResults$get_best_models()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List of MLTrainingResult objects that are marked as best models.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLTrainingResults-toList"></a>}}
\if{latex}{\out{\hypertarget{method-MLTrainingResults-toList}{}}}
\subsection{Method \code{toList()}}{
Convert the object to a list for API compatibility.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLTrainingResults$toList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLTrainingResults-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MLTrainingResults-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLTrainingResults$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
