% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MachineLearning.R
\name{MLTransformerOptions}
\alias{MLTransformerOptions}
\title{MLTransformerOptions}
\description{
Options for ML transformers.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{rank}}{The dimension of the feature space to map the input to.}

\item{\code{fix_zero}}{Whether to map zero to zero, preserving sparsity.}

\item{\code{maximum_bin_count}}{Maximum number of bins (power of 2 recommended).}

\item{\code{minimum_examples_per_bin}}{Minimum number of examples per bin.}

\item{\code{ensure_zero_mean}}{If true, subtract mean from each value before
normalizing and use the raw input otherwise. Defaults to false for
MLNormalizationType.LpNorm, true for MLNormalizationType.GlobalContrast.}

\item{\code{norm}}{Type of norm to use to normalize each sample. The indicated norm
of the resulting vector will be normalized to one. One of \code{L2},
\code{StandardDeviation}, \code{L1}, \code{Infinity}.}

\item{\code{center_data}}{Whether to center the data around 0 by removing the
median.}

\item{\code{quantile_min}}{Quantile min used to scale the data.}

\item{\code{quantile_max}}{Quantile max used to scale the data.}

\item{\code{ensure_unit_standard_deviation}}{If true, the resulting vector's
standard deviation would be one. Otherwise, the resulting vector's L2-norm
would be one.}

\item{\code{scale}}{Scale features by this value.}

\item{\code{use_cdf}}{Whether to use CDF as the output. Defaults to false for
MLNormalizationType.MeanVariance, true for
MLNormalizationType.LogMeanVariance.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MLTransformerOptions-new}{\code{MLTransformerOptions$new()}}
\item \href{#method-MLTransformerOptions-toList}{\code{MLTransformerOptions$toList()}}
\item \href{#method-MLTransformerOptions-clone}{\code{MLTransformerOptions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLTransformerOptions-new"></a>}}
\if{latex}{\out{\hypertarget{method-MLTransformerOptions-new}{}}}
\subsection{Method \code{new()}}{
Create a new MLTransformerOptions instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLTransformerOptions$new(
  rank = NULL,
  fix_zero = NULL,
  maximum_bin_count = NULL,
  minimum_examples_per_bin = NULL,
  ensure_zero_mean = NULL,
  norm = c(NULL, "L2", "StandardDeviation", "L1", "Infinity"),
  center_data = NULL,
  quantile_min = NULL,
  quantile_max = NULL,
  ensure_unit_standard_deviation = NULL,
  scale = NULL,
  use_cdf = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rank}}{The dimension of the feature space to map the input to.}

\item{\code{fix_zero}}{Whether to map zero to zero, preserving sparsity.}

\item{\code{maximum_bin_count}}{Maximum number of bins (power of 2 recommended).}

\item{\code{minimum_examples_per_bin}}{Minimum number of examples per bin.}

\item{\code{ensure_zero_mean}}{If true, subtract mean from each value before
normalizing and use the raw input otherwise. Defaults to false for
MLNormalizationType.LpNorm, true for MLNormalizationType.GlobalContrast.}

\item{\code{norm}}{Type of norm to use to normalize each sample. The indicated
norm of the resulting vector will be normalized to one. One of \code{L2},
\code{StandardDeviation}, \code{L1}, \code{Infinity}.}

\item{\code{center_data}}{Whether to center the data around 0 by removing the
median.}

\item{\code{quantile_min}}{Quantile min used to scale the data.}

\item{\code{quantile_max}}{Quantile max used to scale the data.}

\item{\code{ensure_unit_standard_deviation}}{If true, the resulting vector's
standard deviation would be one. Otherwise, the resulting vector's
L2-norm would be one.}

\item{\code{scale}}{Scale features by this value.}

\item{\code{use_cdf}}{Whether to use CDF as the output. Defaults to false for
MLNormalizationType.MeanVariance, true for
MLNormalizationType.LogMeanVariance.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLTransformerOptions-toList"></a>}}
\if{latex}{\out{\hypertarget{method-MLTransformerOptions-toList}{}}}
\subsection{Method \code{toList()}}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLTransformerOptions$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representation of the MLTransformerOptions object
compatible with the API.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLTransformerOptions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MLTransformerOptions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLTransformerOptions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
