% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Myrconn.PetroVisor.Client-package.R
\docType{package}
\name{Myrconn.PetroVisor.Client-package}
\alias{Myrconn.PetroVisor.Client}
\alias{Myrconn.PetroVisor.Client-package}
\title{Myrconn.PetroVisor.Client: Wrapper for PetroVisor's API}
\description{
A wrapper for PetroVisor's REST API that structures calls into services (data, items, logs, files, tag_entries). These services may be used to interact with the respective PetroVisor objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://datagration.github.io/petrovisor-r-api/}
}

}
\author{
\strong{Maintainer}: Andreas Schweiger \email{andreas.schweiger@weatherford.com}

Authors:
\itemize{
  \item Ramez Abdalla \email{ramez.abdalla@weatherford.com}
}

}
\keyword{internal}
