% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PivotTable.R
\name{PivotTable}
\alias{PivotTable}
\title{PivotTable}
\description{
Class representing a PetroVisor pivot table object.
}
\examples{
\dontrun{
PivotTable$new()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the pivot table.}

\item{\code{add_entity_alias_column}}{This flag specifies whether a column
containing the entity's alias shall be added to the pivot table.}

\item{\code{entity_parent_columns}}{Defines the number of parent columns added to
the pivot table (when a hierarchy is specified).}

\item{\code{add_entity_type_column}}{This flag specifies whether a column containing
the entity's type shall be added to the pivot table.}

\item{\code{scope_formula}}{The scope definition as string (P# syntax).}

\item{\code{entity_set_formula}}{The entity set definition as string (P# syntax).}

\item{\code{table_formula}}{The table definition as string (P# syntax).}

\item{\code{hierarchy_name}}{The name of the hierarchy used to retrieve the parent
name.}

\item{\code{tag_entry_date}}{The date used to evaluate tag entries.}

\item{\code{saved_date}}{The date the pivot table was saved to the database.}

\item{\code{used_tags}}{A list of tag names used in the pivot table.}

\item{\code{skip_empty_rows}}{This flag specifies whether empty rows shall be
skipped during table generation.}

\item{\code{add_is_opportunity_column}}{This flag specifies whether the column
IsOpportunity shall be added to the pivot table.}

\item{\code{append_data}}{This flag specifies whether new data shall be appended to
the saved table.}

\item{\code{description}}{The description of the item.}

\item{\code{labels}}{A list of strings holding the labels of the pivot table.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PivotTable-new}{\code{PivotTable$new()}}
\item \href{#method-PivotTable-toList}{\code{PivotTable$toList()}}
\item \href{#method-PivotTable-clone}{\code{PivotTable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PivotTable-new"></a>}}
\if{latex}{\out{\hypertarget{method-PivotTable-new}{}}}
\subsection{Method \code{new()}}{
Create a new PivotTable instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PivotTable$new(
  name = NULL,
  add_entity_alias_column = NULL,
  entity_parent_columns = 0,
  add_entity_type_column = NULL,
  scope_formula = NULL,
  entity_set_formula = NULL,
  table_formula = NULL,
  hierarchy_name = "",
  tag_entry_date = NULL,
  saved_date = NULL,
  used_tags = list(),
  skip_empty_rows = NULL,
  add_is_opportunity_column = NULL,
  append_data = NULL,
  description = NULL,
  labels = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the pivot table.}

\item{\code{add_entity_alias_column}}{This flag specifies whether a column
containing the entity's alias shall be added to the pivot table.}

\item{\code{entity_parent_columns}}{Defines the number of parent columns added
to the pivot table (when a hierarchy is specified).}

\item{\code{add_entity_type_column}}{This flag specifies whether a column
containing the entity's type shall be added to the pivot table.}

\item{\code{scope_formula}}{The scope definition as string (P# syntax).}

\item{\code{entity_set_formula}}{The entity set definition as string
(P# syntax).}

\item{\code{table_formula}}{The table definition as string (P# syntax).}

\item{\code{hierarchy_name}}{The name of the hierarchy used to retrieve the
parent name.}

\item{\code{tag_entry_date}}{The date used to evaluate tag entries.}

\item{\code{saved_date}}{The date the pivot table was saved to the database.}

\item{\code{used_tags}}{A list of tag names used in the pivot table.}

\item{\code{skip_empty_rows}}{This flag specifies whether empty rows shall be
skipped during table generation.}

\item{\code{add_is_opportunity_column}}{This flag specifies whether the column
IsOpportunity shall be added to the pivot table.}

\item{\code{append_data}}{This flag specifies whether new data shall be appended
to the saved table.}

\item{\code{description}}{The description of the item.}

\item{\code{labels}}{A list of strings holding the labels of the pivot table.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PivotTable-toList"></a>}}
\if{latex}{\out{\hypertarget{method-PivotTable-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PivotTable$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PivotTable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PivotTable-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PivotTable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
