% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{Plot}
\alias{Plot}
\title{Plot}
\description{
Class representing a PetroVisor plot object.
}
\examples{
\dontrun{
Plot$new( )
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the plot.}

\item{\code{plotType}}{The type of the plot.}

\item{\code{inputTableNames}}{A list giving the names of the input tables used
to create the plot.}

\item{\code{xAxisColumnName}}{The name of the column for the x-axis.}

\item{\code{xAxisUnitName}}{The name of the unit of the x-axis.}

\item{\code{yAxisColumnName}}{The name of the column for the y-axis.}

\item{\code{yAxisUnitName}}{The name of the unit of the y-axis.}

\item{\code{ryAxisColumnName}}{The name of the column for the right y-axis.}

\item{\code{ryAxisUnitName}}{The name of the unit of the right y-axis.}

\item{\code{sizeColumnName}}{The name of the column for the size (if scatter-plot).}

\item{\code{sizeUnitName}}{The name of the unit of the size (if scatter-plot).}

\item{\code{formula}}{The plot's definition as string (P# syntax).}

\item{\code{isLocked}}{This flag specifies whether the plot is locked.
Defaults to \code{FALSE}.}

\item{\code{user}}{The user the plot belongs to.}

\item{\code{isFavorite}}{This flag specifies whether the plot is marked as
favorite item, and thus shown in the favorites tab on the home module in
PetroVisor. Defaults to \code{FALSE}.}

\item{\code{labels}}{A list of strings holding the labels of the plot.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Plot-new}{\code{Plot$new()}}
\item \href{#method-Plot-toList}{\code{Plot$toList()}}
\item \href{#method-Plot-clone}{\code{Plot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-new"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-new}{}}}
\subsection{Method \code{new()}}{
Create a new Plot instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plot$new(
  name = NULL,
  plotType = NULL,
  inputTableNames = NULL,
  xAxisColumnName = NULL,
  xAxisUnitName = NULL,
  yAxisColumnName = NULL,
  yAxisUnitName = NULL,
  ryAxisColumnName = NULL,
  ryAxisUnitName = NULL,
  sizeColumnName = NULL,
  sizeUnitName = NULL,
  formula = NULL,
  isLocked = FALSE,
  user = NULL,
  isFavorite = FALSE,
  labels = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the plot.}

\item{\code{plotType}}{The type of the plot.}

\item{\code{inputTableNames}}{A list giving the names of the input tables used
to create the plot.}

\item{\code{xAxisColumnName}}{The name of the column for the x-axis.}

\item{\code{xAxisUnitName}}{The name of the unit of the x-axis.}

\item{\code{yAxisColumnName}}{The name of the column for the y-axis.}

\item{\code{yAxisUnitName}}{The name of the unit of the y-axis.}

\item{\code{ryAxisColumnName}}{The name of the column for the right y-axis.}

\item{\code{ryAxisUnitName}}{The name of the unit of the right y-axis.}

\item{\code{sizeColumnName}}{The name of the column for the size (if
scatter-plot).}

\item{\code{sizeUnitName}}{The name of the unit of the size (if scatter-plot).}

\item{\code{formula}}{The plot's definition as string (P# syntax).}

\item{\code{isLocked}}{This flag specifies whether the plot is locked.
Defaults to \code{FALSE}.}

\item{\code{user}}{The user the plot belongs to.}

\item{\code{isFavorite}}{This flag specifies whether the plot is marked as
favorite item, and thus shown in the favorites tab on the home module
in PetroVisor. Defaults to \code{FALSE}.}

\item{\code{labels}}{A list of strings holding the labels of the plot.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-toList"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plot$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
