% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Polygon.R
\name{Polygon}
\alias{Polygon}
\title{Polygon}
\description{
Class representing a PetroVisor polygon object.
}
\examples{
\dontrun{
Polygon$new(name = "MyPolygon",
          points = list(Point$new(x = 10, y = 20),
                        Point$new(x = 20, y = 10),
                        Point$new(x = 30, y = 40)))
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the polygon.}

\item{\code{points}}{A list of points (objects of class Point) that define the
polygon.}

\item{\code{linkedEntityName}}{The name of the entity linked to the polygon.}

\item{\code{isLocked}}{This flag specifies whether the polygon is locked.
Defaults to \code{FALSE}.}

\item{\code{user}}{The user the polygon belongs to.}

\item{\code{isFavorite}}{This flag specifies whether the polygon is marked as
favorite item, and thus shown in the favorites tab on the home module in
PetroVisor. Defaults to \code{FALSE}.}

\item{\code{labels}}{A list of strings holding the labels of the polygon.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Polygon-new}{\code{Polygon$new()}}
\item \href{#method-Polygon-toList}{\code{Polygon$toList()}}
\item \href{#method-Polygon-clone}{\code{Polygon$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Polygon-new"></a>}}
\if{latex}{\out{\hypertarget{method-Polygon-new}{}}}
\subsection{Method \code{new()}}{
Create a new Polygon instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Polygon$new(
  name = NULL,
  points = NULL,
  linkedEntityName = NULL,
  isLocked = FALSE,
  user = NULL,
  isFavorite = FALSE,
  labels = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the polygon.}

\item{\code{points}}{A list of points (objects of class Point) that define the
polygon.}

\item{\code{linkedEntityName}}{The name of the entity linked to the polygon.}

\item{\code{isLocked}}{This flag specifies whether the polygon is locked.
Defaults to \code{FALSE}.}

\item{\code{user}}{The user the polygon belongs to.}

\item{\code{isFavorite}}{This flag specifies whether the polygon is marked as
favorite item, and thus shown in the favorites tab on the home module
in PetroVisor. Defaults to \code{FALSE}.}

\item{\code{labels}}{A list of strings holding the labels of the polygon.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Polygon-toList"></a>}}
\if{latex}{\out{\hypertarget{method-Polygon-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Polygon$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Polygon-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Polygon-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Polygon$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
