% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessTemplate.R
\name{ProcessTemplate}
\alias{ProcessTemplate}
\title{ProcessTemplate}
\description{
Class representing a PetroVisor process template object.
}
\examples{
\dontrun{
ProcessTemplate$new()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the process template.}

\item{\code{description}}{The discription of the process template.}

\item{\code{processTemplateGroup}}{The group of the process template.}

\item{\code{dueIntervals}}{The interval unti a step is due.}

\item{\code{priority}}{The priority of the process.}

\item{\code{severity}}{The severity of the process.}

\item{\code{steps}}{A list containing the definition of the steps of the
process template. The items of the list must be instances of the class
\code{Step}.}

\item{\code{customFields}}{A list of custom fields of the process template.
The items of the list must be instances of the class \code{CustomField}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProcessTemplate-new}{\code{ProcessTemplate$new()}}
\item \href{#method-ProcessTemplate-toList}{\code{ProcessTemplate$toList()}}
\item \href{#method-ProcessTemplate-clone}{\code{ProcessTemplate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProcessTemplate-new"></a>}}
\if{latex}{\out{\hypertarget{method-ProcessTemplate-new}{}}}
\subsection{Method \code{new()}}{
Create a new ProcessTemplate instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProcessTemplate$new(
  name = NULL,
  description = NULL,
  processTemplateGroup = NULL,
  dueIntervals = NULL,
  priority = NULL,
  severity = NULL,
  steps = NULL,
  customFields = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the process template.}

\item{\code{description}}{The discription of the process template.}

\item{\code{processTemplateGroup}}{The group of the process template.}

\item{\code{dueIntervals}}{The interval unti a step is due.}

\item{\code{priority}}{The priority of the process.}

\item{\code{severity}}{The severity of the process.}

\item{\code{steps}}{A list containing the definition of the steps of the
process template. The items of the list must be instances of the class
\code{Step}.}

\item{\code{customFields}}{A list of custom fields of the process template.
The items of the list must be instances of the class \code{CustomField}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProcessTemplate-toList"></a>}}
\if{latex}{\out{\hypertarget{method-ProcessTemplate-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProcessTemplate$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProcessTemplate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ProcessTemplate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProcessTemplate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
