% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Script.R
\name{PSharpScript}
\alias{PSharpScript}
\title{PSharpScript}
\description{
Class representing a PetroVisor P# script object.
}
\examples{
\dontrun{
PSharpScript$new(name = "Myscript",
          content = "content dummy",
          isLocked = FALSE,
          isFavorite = FALSE)
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the script.}

\item{\code{content}}{The script's content as string (P# syntax).}

\item{\code{isLocked}}{This flag specifies whether the script is locked.
Defaults to \code{FALSE}.}

\item{\code{user}}{The user the script belongs to.}

\item{\code{isFavorite}}{This flag specifies whether the script is marked as
favorite item, and thus shown in the favorites tab on the home module
in PetroVisor. Defaults to \code{FALSE}.}

\item{\code{labels}}{A list of strings holding the labels of the script.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PSharpScript-new}{\code{PSharpScript$new()}}
\item \href{#method-PSharpScript-toList}{\code{PSharpScript$toList()}}
\item \href{#method-PSharpScript-clone}{\code{PSharpScript$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PSharpScript-new"></a>}}
\if{latex}{\out{\hypertarget{method-PSharpScript-new}{}}}
\subsection{Method \code{new()}}{
Create a new P# script instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PSharpScript$new(
  name = NULL,
  content = NULL,
  isLocked = FALSE,
  user = NULL,
  isFavorite = FALSE,
  labels = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the P# script.}

\item{\code{content}}{The script's content as string (P# syntax).}

\item{\code{isLocked}}{This flag specifies whether the script is locked.
Defaults to \code{FALSE}.}

\item{\code{user}}{The user the script belongs to.}

\item{\code{isFavorite}}{This flag specifies whether the script is marked as
favorite item, and thus shown in the favorites tab on the home module
in PetroVisor. Defaults to \code{FALSE}.}

\item{\code{labels}}{A list of strings holding the labels of the script.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PSharpScript-toList"></a>}}
\if{latex}{\out{\hypertarget{method-PSharpScript-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PSharpScript$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PSharpScript-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PSharpScript-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PSharpScript$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
