% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataTypes.R
\name{PVTData}
\alias{PVTData}
\title{PVTData}
\description{
Class representing pvt data.
}
\examples{
\dontrun{
PVTData$new(signal_name = "produced oil per time increment pvt",
            entity_name = "Well01",
            unit_name = "m3",
            data = list(list(Pressure = 300,
                             Temperature = 150,
                             Value = 20)))
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{signal_name}}{The name of the signal.}

\item{\code{entity_name}}{The name of the entity.}

\item{\code{unit_name}}{The data's unit name.}

\item{\code{data}}{The data. A list of pressure-temperature-value
triplets (named list).}

\item{\code{scenario}}{The scenario the data is part of.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PVTData-new}{\code{PVTData$new()}}
\item \href{#method-PVTData-to_list}{\code{PVTData$to_list()}}
\item \href{#method-PVTData-clone}{\code{PVTData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PVTData-new"></a>}}
\if{latex}{\out{\hypertarget{method-PVTData-new}{}}}
\subsection{Method \code{new()}}{
Create a new PVTData instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PVTData$new(
  signal_name = NULL,
  entity_name = NULL,
  unit_name = NULL,
  data = list(),
  scenario = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{signal_name}}{The name of the signal.}

\item{\code{entity_name}}{The name of the entity.}

\item{\code{unit_name}}{The data's unit name.}

\item{\code{data}}{The data. The data. A list of pressure-temperature-value
triplets (named list).}

\item{\code{scenario}}{The scenario the data is part of.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PVTData-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-PVTData-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert the object to a list. This function is mainly used
by the DataServices to convert the objects to lists and then
call the web API.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PVTData$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PVTData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PVTData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PVTData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
