% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReferenceTable.R
\name{ReferenceTable}
\alias{ReferenceTable}
\title{ReferenceTable}
\description{
Class representing a PetroVisor ReferenceTable object.
}
\examples{
\dontrun{
referenceTable <- ReferenceTable$new(
    name = "Test R ReferenceTable",
    description = "Test Description",
    labels = list("A", "B"),
    key = ReferenceTableColumn$new(name = "TestKey",
                                   column_type = "Numeric",
                                   unit_name = "m3"),
    values = list(
        ReferenceTableColumn$new(name = "Value One",
                                 column_type = "Numeric",
                                 unit_name = "m3"),
        ReferenceTableColumn$new(name = "Value Two",
                                 column_type = "Bool",
                                 unit_name = " ")
    ))
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the reference table.}

\item{\code{description}}{The description of the item.}

\item{\code{labels}}{A list of strings holding the labels of the reference table.}

\item{\code{key}}{The key column definition. Object of type ReferenceTableColumn.}

\item{\code{values}}{List of value column definitions. I.e. a list of objects of
type ReferenceTableColumn.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ReferenceTable-new}{\code{ReferenceTable$new()}}
\item \href{#method-ReferenceTable-toList}{\code{ReferenceTable$toList()}}
\item \href{#method-ReferenceTable-clone}{\code{ReferenceTable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReferenceTable-new"></a>}}
\if{latex}{\out{\hypertarget{method-ReferenceTable-new}{}}}
\subsection{Method \code{new()}}{
Create a new ReferenceTable instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReferenceTable$new(
  name = NULL,
  description = NULL,
  labels = list(),
  key = NULL,
  values = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the scope.}

\item{\code{description}}{The description of the item.}

\item{\code{labels}}{A list of strings holding the labels of the scope.}

\item{\code{key}}{The key column definition. Object of type
ReferenceTableColumn.}

\item{\code{values}}{List of value column definitions. I.e. a list of objects of
type ReferenceTableColumn.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReferenceTable-toList"></a>}}
\if{latex}{\out{\hypertarget{method-ReferenceTable-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReferenceTable$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReferenceTable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ReferenceTable-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReferenceTable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
