% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReferenceTable.R
\name{ReferenceTableColumn}
\alias{ReferenceTableColumn}
\title{ReferenceTableColumn}
\description{
Class representing a PetroVisor ReferenceTableColumn object.
}
\examples{
\dontrun{
column <- ReferenceTableColumn$new(name = "Column One",
                                   column_type = "Numeric",
                                   unit_name = "m3")
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the column.}

\item{\code{column_type}}{The type of the column.}

\item{\code{unit_name}}{The name of the unit associated with the column.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ReferenceTableColumn-new}{\code{ReferenceTableColumn$new()}}
\item \href{#method-ReferenceTableColumn-toList}{\code{ReferenceTableColumn$toList()}}
\item \href{#method-ReferenceTableColumn-clone}{\code{ReferenceTableColumn$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReferenceTableColumn-new"></a>}}
\if{latex}{\out{\hypertarget{method-ReferenceTableColumn-new}{}}}
\subsection{Method \code{new()}}{
Create a new ReferenceTableColumn instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReferenceTableColumn$new(
  name = NULL,
  column_type = c("Numeric", "String", "DateTime", "Bool"),
  unit_name = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the column.}

\item{\code{column_type}}{The type of the column.}

\item{\code{unit_name}}{The name of the unit associated with the column.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReferenceTableColumn-toList"></a>}}
\if{latex}{\out{\hypertarget{method-ReferenceTableColumn-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReferenceTableColumn$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReferenceTableColumn-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ReferenceTableColumn-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReferenceTableColumn$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
