% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workflow.R
\name{RProviderConfiguration}
\alias{RProviderConfiguration}
\title{RProviderConfiguration}
\description{
Class representing an R provider configuration object.
}
\examples{
\dontrun{
RProviderConfiguration$new()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{providerType}}{A string specifying the provider type. Allowed values
are: \code{RNET} and \code{Rserve}.}

\item{\code{rServerAddress}}{The URL or host name of the server running R.}

\item{\code{rServerPort}}{The port Rserve is available at.}

\item{\code{rServerScriptsFolder}}{The path to the folder containing the scripts.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RProviderConfiguration-new}{\code{RProviderConfiguration$new()}}
\item \href{#method-RProviderConfiguration-toList}{\code{RProviderConfiguration$toList()}}
\item \href{#method-RProviderConfiguration-clone}{\code{RProviderConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RProviderConfiguration-new"></a>}}
\if{latex}{\out{\hypertarget{method-RProviderConfiguration-new}{}}}
\subsection{Method \code{new()}}{
Create a new R provider configuration instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RProviderConfiguration$new(
  providerType = c("RNET", "Rserve"),
  rServerAddress = NULL,
  rServerPort = NULL,
  rServerScriptsFolder = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{providerType}}{A string specifying the provider type. Allowed values
are: \code{RNET} and \code{Rserve}.}

\item{\code{rServerAddress}}{The URL or host name of the server running R.}

\item{\code{rServerPort}}{The port Rserve is available at.}

\item{\code{rServerScriptsFolder}}{The path to the folder containing the
scripts.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RProviderConfiguration-toList"></a>}}
\if{latex}{\out{\hypertarget{method-RProviderConfiguration-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RProviderConfiguration$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RProviderConfiguration-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RProviderConfiguration-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RProviderConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
