% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workflow.R
\name{RWorkflowActivity}
\alias{RWorkflowActivity}
\title{RWorkflowActivity}
\description{
Class representing an R workflow activity object.
}
\examples{
\dontrun{
RWorkflowActivity$new()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the R workflow activity.}

\item{\code{providerConfiguration}}{The configuration of the provider. Must be an
object of the class \code{RProviderConfiguration}.}

\item{\code{scriptName}}{The name of the activity's R script.}

\item{\code{functionName}}{The name of the activity's R function. The function has
to present in the referenced R script.}

\item{\code{input}}{A list of input arguments for the activity. The items in the
list must be objects of the class \code{ActivityArgument}.}

\item{\code{output}}{A list of output arguments for the activity. The items in the
list must be objects of the class \code{ActivityArgument}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RWorkflowActivity-new}{\code{RWorkflowActivity$new()}}
\item \href{#method-RWorkflowActivity-toList}{\code{RWorkflowActivity$toList()}}
\item \href{#method-RWorkflowActivity-clone}{\code{RWorkflowActivity$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RWorkflowActivity-new"></a>}}
\if{latex}{\out{\hypertarget{method-RWorkflowActivity-new}{}}}
\subsection{Method \code{new()}}{
Create a new R workflow activity instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RWorkflowActivity$new(
  name = NULL,
  providerConfiguration = NULL,
  scriptName = NULL,
  functionName = NULL,
  input = NULL,
  output = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the R workflow activity.}

\item{\code{providerConfiguration}}{The configuration of the provider. Must be
an object of the class \code{RProviderConfiguration}.}

\item{\code{scriptName}}{The name of the activity's R script.}

\item{\code{functionName}}{The name of the activity's R function. The function
has to present in the referenced R script.}

\item{\code{input}}{A list of input arguments for the activity. The items in the
list must be objects of the class \code{ActivityArgument}.}

\item{\code{output}}{A list of output arguments for the activity. The items in
the list must be objects of the class \code{ActivityArgument}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RWorkflowActivity-toList"></a>}}
\if{latex}{\out{\hypertarget{method-RWorkflowActivity-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RWorkflowActivity$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RWorkflowActivity-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RWorkflowActivity-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RWorkflowActivity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
