% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scenario.R
\name{Scenario}
\alias{Scenario}
\title{Scenario}
\description{
Class representing a PetroVisor scenario object.
}
\examples{
\dontrun{
Scenario$new(
  name = "Test R Scenario",
  description = "Test Scenario")
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the scenario}

\item{\code{configuration_settings}}{A list holding configuration items associated
with the scenario. Objects in this list are of type ConfigurationSetting.}

\item{\code{description}}{The description of the item.}

\item{\code{labels}}{A list of strings holding the labels of the item.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Scenario-new}{\code{Scenario$new()}}
\item \href{#method-Scenario-toList}{\code{Scenario$toList()}}
\item \href{#method-Scenario-clone}{\code{Scenario$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scenario-new"></a>}}
\if{latex}{\out{\hypertarget{method-Scenario-new}{}}}
\subsection{Method \code{new()}}{
Create a new Scenario instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scenario$new(
  name = NULL,
  configuration_settings = list(),
  description = NULL,
  labels = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the scenario}

\item{\code{configuration_settings}}{A list holding configuration items
associated with the scenario. Objects in this list are of type
ConfigurationSetting.}

\item{\code{description}}{The description of the item.}

\item{\code{labels}}{A list of strings holding the labels of the item.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scenario-toList"></a>}}
\if{latex}{\out{\hypertarget{method-Scenario-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scenario$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scenario-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Scenario-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scenario$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
