% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workflow.R
\name{ScheduleEnd}
\alias{ScheduleEnd}
\title{ScheduleEnd}
\description{
Class representing a schedule end object.
}
\examples{
\dontrun{
ScheduleEnd$new()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{endType}}{The type of the schedule end (integer).}

\item{\code{repetitionsCount}}{The number of times the schedule was executed.}

\item{\code{totalRepetitions}}{The number of times the schedule shall be executed.}

\item{\code{endDate}}{The date when the schedule shall end.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ScheduleEnd-new}{\code{ScheduleEnd$new()}}
\item \href{#method-ScheduleEnd-toList}{\code{ScheduleEnd$toList()}}
\item \href{#method-ScheduleEnd-clone}{\code{ScheduleEnd$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScheduleEnd-new"></a>}}
\if{latex}{\out{\hypertarget{method-ScheduleEnd-new}{}}}
\subsection{Method \code{new()}}{
Create a new schedule end instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduleEnd$new(
  endType = NULL,
  repetitionsCount = NULL,
  totalRepetitions = NULL,
  endDate = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endType}}{The type of the schedule end (integer).}

\item{\code{repetitionsCount}}{The number of times the schedule was executed.}

\item{\code{totalRepetitions}}{The number of times the schedule shall be
executed.}

\item{\code{endDate}}{The date when the schedule shall end.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScheduleEnd-toList"></a>}}
\if{latex}{\out{\hypertarget{method-ScheduleEnd-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduleEnd$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScheduleEnd-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ScheduleEnd-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduleEnd$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
