% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workflow.R
\name{ScheduleRecurrence}
\alias{ScheduleRecurrence}
\title{ScheduleRecurrence}
\description{
Class representing a schedule recurrence object.
}
\examples{
\dontrun{
ScheduleRecurrence$new()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{recurrenceType}}{The type of the recurrence (string).}

\item{\code{intervalsAmount}}{Run the schedule every n days.}

\item{\code{skipWeekend}}{Flag specifying whether to skip execution on weekends.}

\item{\code{repeatOnSpecificDayMonth}}{Flag specifying whether to reapeat the
schedule on a specific day of a month.}

\item{\code{specificDayNumber}}{The number of the day the schedule shall be repeated
on if \code{repeatOnSpecificDayMonth} is \code{TRUE}.}

\item{\code{orderNumberDay}}{??}

\item{\code{typeDayMonthDay}}{??}

\item{\code{runOnMonday}}{Flag specifying if the schedule is run on Mondays.}

\item{\code{runOnTuesday}}{Flag specifying if the schedule is run on Tuesdays.}

\item{\code{runOnWednesday}}{Flag specifying if the schedule is run on Wednesdays.}

\item{\code{runOnThursday}}{Flag specifying if the schedule is run on Thursdays.}

\item{\code{runOnFriday}}{Flag specifying if the schedule is run on Fridays.}

\item{\code{runOnSaturday}}{Flag specifying if the schedule is run on Saturdays.}

\item{\code{runOnSunday}}{Flag specifying if the schedule is run on Sundays.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ScheduleRecurrence-new}{\code{ScheduleRecurrence$new()}}
\item \href{#method-ScheduleRecurrence-toList}{\code{ScheduleRecurrence$toList()}}
\item \href{#method-ScheduleRecurrence-clone}{\code{ScheduleRecurrence$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScheduleRecurrence-new"></a>}}
\if{latex}{\out{\hypertarget{method-ScheduleRecurrence-new}{}}}
\subsection{Method \code{new()}}{
Create a new schedule recurrence instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduleRecurrence$new(
  recurrenceType = NULL,
  intervalsAmount = NULL,
  skipWeekend = NULL,
  repeatOnSpecificDayMonth = NULL,
  specificDayNumber = NULL,
  orderNumberDay = NULL,
  typeDayMonthDay = NULL,
  runOnMonday = NULL,
  runOnTuesday = NULL,
  runOnWednesday = NULL,
  runOnThursday = NULL,
  runOnFriday = NULL,
  runOnSaturday = NULL,
  runOnSunday = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{recurrenceType}}{The type of the recurrence (string).}

\item{\code{intervalsAmount}}{Run the schedule every n days.}

\item{\code{skipWeekend}}{Flag specifying whether to skip execution on weekends.}

\item{\code{repeatOnSpecificDayMonth}}{Flag specifying whether to reapeat the
schedule on a specific day of a month.}

\item{\code{specificDayNumber}}{The number of the day the schedule shall be
repeated on if \code{repeatOnSpecificDayMonth} is \code{TRUE}.}

\item{\code{orderNumberDay}}{??}

\item{\code{typeDayMonthDay}}{??}

\item{\code{runOnMonday}}{Flag specifying if the schedule is run on Mondays.}

\item{\code{runOnTuesday}}{Flag specifying if the schedule is run on Tuesdays.}

\item{\code{runOnWednesday}}{Flag specifying if the schedule is run on
Wednesdays.}

\item{\code{runOnThursday}}{Flag specifying if the schedule is run on Thursdays.}

\item{\code{runOnFriday}}{Flag specifying if the schedule is run on Fridays.}

\item{\code{runOnSaturday}}{Flag specifying if the schedule is run on Saturdays.}

\item{\code{runOnSunday}}{Flag specifying if the schedule is run on Sundays.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScheduleRecurrence-toList"></a>}}
\if{latex}{\out{\hypertarget{method-ScheduleRecurrence-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduleRecurrence$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScheduleRecurrence-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ScheduleRecurrence-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduleRecurrence$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
