% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chart.R
\name{SeriesSetting}
\alias{SeriesSetting}
\title{SeriesSetting}
\description{
Class representing a series setting object. These settings are
used when creating a chart.
}
\examples{
\dontrun{
SeriesSetting$new(scatterSize = 2,
          lineThickness = 3,
          color = "#123456")
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{scatterSize}}{An integer giving the scatter size of the series.}

\item{\code{lineThickness}}{An integer specifiying the series' line thickness.}

\item{\code{color}}{A string specifying the series' color.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeriesSetting-new}{\code{SeriesSetting$new()}}
\item \href{#method-SeriesSetting-toList}{\code{SeriesSetting$toList()}}
\item \href{#method-SeriesSetting-clone}{\code{SeriesSetting$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeriesSetting-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeriesSetting-new}{}}}
\subsection{Method \code{new()}}{
Create a new SeriesSetting instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeriesSetting$new(scatterSize = NULL, lineThickness = NULL, color = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scatterSize}}{An integer giving the scatter size of the series.}

\item{\code{lineThickness}}{An integer specifiying the series' line thickness.}

\item{\code{color}}{A string specifying the series' color.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeriesSetting-toList"></a>}}
\if{latex}{\out{\hypertarget{method-SeriesSetting-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeriesSetting$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeriesSetting-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeriesSetting-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeriesSetting$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
