% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signal.R
\name{Signal}
\alias{Signal}
\title{Signal}
\description{
Class representing a PetroVisor signal object.
}
\examples{
\dontrun{
Signal$new(name = "my signal",
           short_name = "my sig",
           measurement_name = "Length",
           storage_unit_name = "m",
           aggregation_type = "Average",
           container_aggregation_type = "Sum",
           signal_type = "TimeDependent",
           default_color = 0,
           default_line_type = "Solid",
           setting_name = NULL,
           labels = list(),
           description = NULL)
}
}
\seealso{
\itemize{
\item \link{DataServices} for loading and saving signal data
\item \link{RepositoryService} for managing signal definitions
\item \link{Unit} for working with measurement units
\item \link{UnitMeasurement} for unit measurements
\item \code{vignette("working-with-data")} for data examples
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the signal.}

\item{\code{short_name}}{The signal's short name.}

\item{\code{measurement_name}}{The signal's measurement name.}

\item{\code{storage_unit_name}}{The name of the signal's storage \link{Unit}.}

\item{\code{aggregation_type}}{The signal's aggregation type.}

\item{\code{container_aggregation_type}}{The signal's container aggregation type.}

\item{\code{signal_type}}{The signal's type.}

\item{\code{default_color}}{The signal's default color.}

\item{\code{default_line_type}}{The signal's default line type.}

\item{\code{setting_name}}{(Optional) The name of the enumeration linked to the
signal.}

\item{\code{labels}}{A list of strings holding the labels of the signal.}

\item{\code{description}}{The description of the signal.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Signal-new}{\code{Signal$new()}}
\item \href{#method-Signal-toList}{\code{Signal$toList()}}
\item \href{#method-Signal-clone}{\code{Signal$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Signal-new"></a>}}
\if{latex}{\out{\hypertarget{method-Signal-new}{}}}
\subsection{Method \code{new()}}{
Create a new Signal instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Signal$new(
  name = NULL,
  short_name = NULL,
  measurement_name = NULL,
  storage_unit_name = NULL,
  aggregation_type = c("Sum", "Average", "Mode", "Max", "Min", "Count", "Count Distinct",
    "First", "Last", "None"),
  container_aggregation_type = c("Sum", "Average", "Mode", "Max", "Min", "Count",
    "Count Distinct", "First", "Last", "None"),
  signal_type = c("Static", "TimeDependent", "DepthDependent", "String", "PVT",
    "StringTimeDependent", "StringDepthDependent"),
  default_color = 0,
  default_line_type = c("Solid", "Dash", "DashDot"),
  setting_name = NULL,
  labels = list(),
  description = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the signal.}

\item{\code{short_name}}{The signal's short name.}

\item{\code{measurement_name}}{The signal's measurement name.}

\item{\code{storage_unit_name}}{The name of the signal's storage unit.}

\item{\code{aggregation_type}}{The signal's aggregation type.}

\item{\code{container_aggregation_type}}{The signal's container aggregation
type.}

\item{\code{signal_type}}{The signal's type.}

\item{\code{default_color}}{The signal's default color.}

\item{\code{default_line_type}}{The signal's default line type.}

\item{\code{setting_name}}{(Optional) The name of the enumeration linked to the
signal.}

\item{\code{labels}}{A list of strings holding the labels of the signal.}

\item{\code{description}}{The description of the signal.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Signal-toList"></a>}}
\if{latex}{\out{\hypertarget{method-Signal-toList}{}}}
\subsection{Method \code{toList()}}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Signal$toList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Signal-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Signal-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Signal$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
