% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataTypes.R
\name{StaticData}
\alias{StaticData}
\title{StaticData}
\description{
Class representing static data.
}
\examples{
\dontrun{
StaticData$new(signalName = "surface x-coordinate",
               entityName = "Well01",
               unitName = "m",
               data = 10)
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{signalName}}{The name of the signal.}

\item{\code{entityName}}{The name of the entity.}

\item{\code{unitName}}{The data's unit name.}

\item{\code{data}}{The data. In this case case a single numberic or string value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StaticData-new}{\code{StaticData$new()}}
\item \href{#method-StaticData-toList}{\code{StaticData$toList()}}
\item \href{#method-StaticData-clone}{\code{StaticData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StaticData-new"></a>}}
\if{latex}{\out{\hypertarget{method-StaticData-new}{}}}
\subsection{Method \code{new()}}{
Create a new StaticData instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StaticData$new(
  signalName = NULL,
  entityName = NULL,
  unitName = NULL,
  data = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{signalName}}{The name of the signal.}

\item{\code{entityName}}{The name of the entity.}

\item{\code{unitName}}{The data's unit name.}

\item{\code{data}}{The data. In this case case a single numberic or string value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StaticData-toList"></a>}}
\if{latex}{\out{\hypertarget{method-StaticData-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StaticData$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the DataServices to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StaticData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StaticData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StaticData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
