% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataServices.R
\name{StaticNumeric}
\alias{StaticNumeric}
\title{StaticNumeric}
\description{
Provides access to functionality related to static numeric data.
}
\details{
A new instance of this class will be created by the DataServices
when calling the method 'GetStaticNumeric'.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StaticNumeric-new}{\code{StaticNumeric$new()}}
\item \href{#method-StaticNumeric-Delete}{\code{StaticNumeric$Delete()}}
\item \href{#method-StaticNumeric-Load}{\code{StaticNumeric$Load()}}
\item \href{#method-StaticNumeric-Save}{\code{StaticNumeric$Save()}}
\item \href{#method-StaticNumeric-clone}{\code{StaticNumeric$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StaticNumeric-new"></a>}}
\if{latex}{\out{\hypertarget{method-StaticNumeric-new}{}}}
\subsection{Method \code{new()}}{
Create a new StaticNumeric object instance. This is done by
the DataServices automatically if needed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StaticNumeric$new(url, tokenType, token)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{The URL for the API calls.}

\item{\code{tokenType}}{The type of the issued token.}

\item{\code{token}}{The issued token.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StaticNumeric-Delete"></a>}}
\if{latex}{\out{\hypertarget{method-StaticNumeric-Delete}{}}}
\subsection{Method \code{Delete()}}{
Delete the static numeric data specified by the requests.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StaticNumeric$Delete(dataSetRequests)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataSetRequests}}{A list of data set requests specifying which data
shall be deleted. The items in the list must be objects of the class
'DataSetRequest'.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StaticNumeric-Load"></a>}}
\if{latex}{\out{\hypertarget{method-StaticNumeric-Load}{}}}
\subsection{Method \code{Load()}}{
Load the static numeric data specified by the requests.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StaticNumeric$Load(dataSetRequests)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataSetRequests}}{A list of data set requests specifying which data
shall be loaded. The items in the list must be objects of the class
'DataSetRequest'.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of type list containing the data and other fields.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StaticNumeric-Save"></a>}}
\if{latex}{\out{\hypertarget{method-StaticNumeric-Save}{}}}
\subsection{Method \code{Save()}}{
Save the specified static numeric data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StaticNumeric$Save(staticData)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{staticData}}{The data to save. Has to be an object of the class
'StaticData'. If the unit is not specified, the data will be saved using
the signal's storage unit.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StaticNumeric-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StaticNumeric-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StaticNumeric$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
