% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataServices.R
\name{StaticString}
\alias{StaticString}
\title{StaticString}
\description{
Provides access to functionality related to static string data.
}
\details{
A new instance of this class will be created by the DataServices
when calling the method 'GetStaticString'.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StaticString-new}{\code{StaticString$new()}}
\item \href{#method-StaticString-Delete}{\code{StaticString$Delete()}}
\item \href{#method-StaticString-Load}{\code{StaticString$Load()}}
\item \href{#method-StaticString-Save}{\code{StaticString$Save()}}
\item \href{#method-StaticString-clone}{\code{StaticString$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StaticString-new"></a>}}
\if{latex}{\out{\hypertarget{method-StaticString-new}{}}}
\subsection{Method \code{new()}}{
Create a new StaticString object instance. This is done by
the DataServices automatically if needed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StaticString$new(url, tokenType, token)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{The URL for the API calls.}

\item{\code{tokenType}}{The type of the issued token.}

\item{\code{token}}{The issued token.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StaticString-Delete"></a>}}
\if{latex}{\out{\hypertarget{method-StaticString-Delete}{}}}
\subsection{Method \code{Delete()}}{
Delete the static string data specified by the requests.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StaticString$Delete(dataSetRequests)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataSetRequests}}{A list of data set requests specifying which data
shall be deleted. The items in the list must be objects of the class
'DataSetRequest'.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StaticString-Load"></a>}}
\if{latex}{\out{\hypertarget{method-StaticString-Load}{}}}
\subsection{Method \code{Load()}}{
Load the static string data specified by the requests.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StaticString$Load(dataSetRequests)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataSetRequests}}{A list of data set requests specifying which data
shall be loaded. The items in the list must be objects of the class
'DataSetRequest'.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of type list containing the data and other fields.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StaticString-Save"></a>}}
\if{latex}{\out{\hypertarget{method-StaticString-Save}{}}}
\subsection{Method \code{Save()}}{
Save the specified static string data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StaticString$Save(staticData)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{staticData}}{The data to save. Has to be an object of the class
'StaticData'. If the unit is not specified, the data will be saved using
the signal's storage unit.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StaticString-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StaticString-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StaticString$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
