% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TagEntry.R
\name{TagEntry}
\alias{TagEntry}
\title{TagEntry}
\description{
Class representing a PetroVisor tag entry.
}
\examples{
\dontrun{
# create a new instance of the service provider
sp <- ServiceProvider$new("Host", 8095, "WorkspaceA", "UserX", "Password")

# create a new tag entry
TagEntry$new(entity_name = "Well 01" ,
             tag_name = "Active",
             start = "2020-02-01T00:00:00.000Z")
}
}
\seealso{
\itemize{
\item \link{Tag} for tag definitions
\item \link{TagEntriesService} for loading and saving tag entries
\item \link{Entity} for entity definitions
\item \link{RepositoryService} for repository operations
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tag_name}}{The name of the \link{Tag}.}

\item{\code{entity_name}}{The name of the tagged \link{Entity}.}

\item{\code{start}}{The start date of the tag entry.}

\item{\code{end}}{The end date of the tag entry.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TagEntry-new}{\code{TagEntry$new()}}
\item \href{#method-TagEntry-to_list}{\code{TagEntry$to_list()}}
\item \href{#method-TagEntry-clone}{\code{TagEntry$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TagEntry-new"></a>}}
\if{latex}{\out{\hypertarget{method-TagEntry-new}{}}}
\subsection{Method \code{new()}}{
Create a new TagEntry instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TagEntry$new(tag_name = NULL, entity_name = NULL, start = NULL, end = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tag_name}}{The name of the \link{Tag}.}

\item{\code{entity_name}}{The name of the tagged \link{Entity}.}

\item{\code{start}}{The start date of the tag entry.}

\item{\code{end}}{(Optional) The end date of the tag entry.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TagEntry-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-TagEntry-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert the object to a list. This function is mainly used
by the TagEntriesService to convert the objects to lists and then
call the web API.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TagEntry$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TagEntry-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TagEntry-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TagEntry$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
