% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TagEntry.R
\name{TagEntryFilter}
\alias{TagEntryFilter}
\title{TagEntryFilter}
\description{
Class representing a PetroVisor tag entry filter.
}
\examples{
\dontrun{
TagEntryFilter$new(entityName = "Well01")
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{entityName}}{The name of the entity for which tag entries shall be
retrieved.}

\item{\code{entityNames}}{A list of entity names. Only tag entries for the given
entities will be returned.}

\item{\code{tagName}}{The name of the tag for which tag entries shall be retrieved.}

\item{\code{tagNames}}{A lisf of tag names. Only tag entries for the given tags will
be returned.}

\item{\code{tagGroupName}}{The name of a tag group for which tag entries shall be
retrieved.}

\item{\code{tagGroupNames}}{A list of tag group names. Only tag entries for the
given tag groups will be returned.}

\item{\code{start}}{Start of the range for which tag entries shall be retrieved.}

\item{\code{end}}{End of the range for which tag entries shall be retrieved.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TagEntryFilter-new}{\code{TagEntryFilter$new()}}
\item \href{#method-TagEntryFilter-toList}{\code{TagEntryFilter$toList()}}
\item \href{#method-TagEntryFilter-clone}{\code{TagEntryFilter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TagEntryFilter-new"></a>}}
\if{latex}{\out{\hypertarget{method-TagEntryFilter-new}{}}}
\subsection{Method \code{new()}}{
Create a new TagEntryFilter instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TagEntryFilter$new(
  entityName = NULL,
  entityNames = NULL,
  tagName = NULL,
  tagNames = NULL,
  tagGroupName = NULL,
  tagGroupNames = NULL,
  start = NULL,
  end = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{entityName}}{The name of the entity for which tag entries shall be
retrieved.}

\item{\code{entityNames}}{A list of entity names. Only tag entries for the given
entities will be returned.}

\item{\code{tagName}}{The name of the tag for which tag entries shall be
retrieved.}

\item{\code{tagNames}}{A lisf of tag names. Only tag entries for the given tags
will be returned.}

\item{\code{tagGroupName}}{The name of a tag group for which tag entries shall
be retrieved.}

\item{\code{tagGroupNames}}{A list of tag group names. Only tag entries for the
given tag groups will be returned.}

\item{\code{start}}{Start of the range for which tag entries shall be retrieved.}

\item{\code{end}}{End of the range for which tag entries shall be retrieved}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TagEntryFilter-toList"></a>}}
\if{latex}{\out{\hypertarget{method-TagEntryFilter-toList}{}}}
\subsection{Method \code{toList()}}{
Convert the object to a list. This function is mainly used
by the TagEntriesService to convert the objects to lists and then
call the web API.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TagEntryFilter$toList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TagEntryFilter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TagEntryFilter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TagEntryFilter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
