% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataServices.R
\name{TimeNumeric}
\alias{TimeNumeric}
\title{TimeNumeric}
\description{
Provides access to functionality related to time-dependent
numeric data.
}
\details{
A new instance of this class will be created by the DataServices
when calling the method 'GetTimeNumeric'.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TimeNumeric-new}{\code{TimeNumeric$new()}}
\item \href{#method-TimeNumeric-Delete}{\code{TimeNumeric$Delete()}}
\item \href{#method-TimeNumeric-Load}{\code{TimeNumeric$Load()}}
\item \href{#method-TimeNumeric-Save}{\code{TimeNumeric$Save()}}
\item \href{#method-TimeNumeric-clone}{\code{TimeNumeric$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeNumeric-new"></a>}}
\if{latex}{\out{\hypertarget{method-TimeNumeric-new}{}}}
\subsection{Method \code{new()}}{
Create a new TimeNumeric object instance. This is done by
the DataServices automatically if needed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeNumeric$new(url, tokenType, token)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{The URL for the API calls.}

\item{\code{tokenType}}{The type of the issued token.}

\item{\code{token}}{The issued token.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeNumeric-Delete"></a>}}
\if{latex}{\out{\hypertarget{method-TimeNumeric-Delete}{}}}
\subsection{Method \code{Delete()}}{
Delete the time-dependent numeric data specified by the
requests.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeNumeric$Delete(start, end, dataSetRequests)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start}}{The start of the time-range for which the data shall be
deleted.}

\item{\code{end}}{The end of the time-range for which the data shall be deleted.}

\item{\code{dataSetRequests}}{A list of data set requests specifying which data
shall be deleted. The items in the list must be objects of the class
'DataSetRequest'.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeNumeric-Load"></a>}}
\if{latex}{\out{\hypertarget{method-TimeNumeric-Load}{}}}
\subsection{Method \code{Load()}}{
Load the time-dependent numeric data specified by the
requests.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeNumeric$Load(
  start,
  end,
  increment = c("EveryMinute", "EveryFiveMinutes", "EveryFifteenMinutes", "Hourly",
    "Daily", "Monthly", "Quarterly", "Yearly"),
  hierarchyName = NULL,
  dataSetRequests
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start}}{The start of the time-range for which the data shall be
loaded.}

\item{\code{end}}{The end of the time-range for which the data shall be loaded.}

\item{\code{increment}}{The time increment in which the data shall be loaded.
Allowed values are: \code{"EveryMinute"}, \code{"EveryFiveMinutes"},
\code{"EveryFifteenMinutes"}, \code{"Hourly"}, \code{"Daily"},
\code{"Monthly"}, \code{"Quarterly"}, \code{"Yearly"}}

\item{\code{hierarchyName}}{(Optional) If the name of an existing hierarchy is
specified, the data will be aggregated according to the hierarchy
relationships if needed.}

\item{\code{dataSetRequests}}{A list of data set requests specifying which data
shall be loaded. The items in the list must be objects of the class
'DataSetRequest'.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of type list containing the data and other fields.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeNumeric-Save"></a>}}
\if{latex}{\out{\hypertarget{method-TimeNumeric-Save}{}}}
\subsection{Method \code{Save()}}{
Save the specified time-dependent numeric data. Existing
data in the range of the new data is deleted first. The range of the data
is defined by the specified data to save.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeNumeric$Save(timeData)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timeData}}{The data to save. Has to be an object of the class
'TimeData'. If the unit is not specified, the data will be saved using
the signal's storage unit.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeNumeric-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TimeNumeric-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeNumeric$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
