% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unit.R
\name{Unit}
\alias{Unit}
\title{Unit}
\description{
Class representing a PetroVisor unit object.
}
\examples{
\dontrun{
Unit$new(name = "hyper m",
         measurement_name = "Length",
         factor = 10000000000000,
         summand = 0)
}
}
\seealso{
\itemize{
\item \link{UnitMeasurement} for measurement definitions
\item \link{Signal} for signal definitions using units
\item \link{RepositoryService} for loading and saving units
\item \link{DataServices} for data operations with units
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the unit.}

\item{\code{measurement_name}}{The unit's measurement name.}

\item{\code{factor}}{The unit's factor. Used for conversion between the unit and
it's base (SI) unit.}

\item{\code{summand}}{The unit's summand. Used for conversion between the unit and
it's base (SI) unit.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Unit-new}{\code{Unit$new()}}
\item \href{#method-Unit-toList}{\code{Unit$toList()}}
\item \href{#method-Unit-clone}{\code{Unit$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Unit-new"></a>}}
\if{latex}{\out{\hypertarget{method-Unit-new}{}}}
\subsection{Method \code{new()}}{
Create a new Unit instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Unit$new(name = NULL, measurement_name = NULL, factor = NULL, summand = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the unit.}

\item{\code{measurement_name}}{The unit's measurement name.}

\item{\code{factor}}{The unit's factor. Used for conversion between the unit and
it's base (SI) unit.}

\item{\code{summand}}{The unit's summand. Used for conversion between the unit
and it's base (SI) unit.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Unit-toList"></a>}}
\if{latex}{\out{\hypertarget{method-Unit-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Unit$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Unit-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Unit-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Unit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
