% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserSetting.R
\name{UserSetting}
\alias{UserSetting}
\title{UserSetting}
\description{
Class representing a PetroVisor UserSetting object.
}
\examples{
\dontrun{
UserSetting$new(name = "MySetting", value = "MySettingValue")
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the user setting.}

\item{\code{value}}{The value of the user setting.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserSetting-new}{\code{UserSetting$new()}}
\item \href{#method-UserSetting-toList}{\code{UserSetting$toList()}}
\item \href{#method-UserSetting-clone}{\code{UserSetting$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserSetting-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserSetting-new}{}}}
\subsection{Method \code{new()}}{
Create a new UserSetting instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserSetting$new(name = NULL, value = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the user setting.}

\item{\code{value}}{The value of the user setting.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserSetting-toList"></a>}}
\if{latex}{\out{\hypertarget{method-UserSetting-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserSetting$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserSetting-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserSetting-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserSetting$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
