% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VoronoiGrid.R
\name{VoronoiCell}
\alias{VoronoiCell}
\title{VoronoiCell}
\description{
Class representing a PetroVisor VoronoiCell object.
}
\examples{
\dontrun{
VoronoiCell$new()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the Voronoi cell.}

\item{\code{center}}{The center of the Voronoi Cell (instances of class Point).}

\item{\code{points}}{A list of points of the Voronoi cell (instances of class
Point).}

\item{\code{area}}{The area of the vornoi cell.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VoronoiCell-new}{\code{VoronoiCell$new()}}
\item \href{#method-VoronoiCell-toList}{\code{VoronoiCell$toList()}}
\item \href{#method-VoronoiCell-clone}{\code{VoronoiCell$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VoronoiCell-new"></a>}}
\if{latex}{\out{\hypertarget{method-VoronoiCell-new}{}}}
\subsection{Method \code{new()}}{
Create a new VoronoiCell instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VoronoiCell$new(name = NULL, center = NULL, points = NULL, area = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the Voronoi cell.}

\item{\code{center}}{The center of the Voronoi Cell (instances of class Point).}

\item{\code{points}}{A list of points of the Voronoi cell (instances of class
Point).}

\item{\code{area}}{The area of the vornoi cell.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VoronoiCell-toList"></a>}}
\if{latex}{\out{\hypertarget{method-VoronoiCell-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VoronoiCell$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VoronoiCell-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VoronoiCell-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VoronoiCell$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
