% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VoronoiGrid.R
\name{VoronoiGrid}
\alias{VoronoiGrid}
\title{VoronoiGrid}
\description{
Class representing a PetroVisor Voronoi grid object.
}
\examples{
\dontrun{
VoronoiGrid$new()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the Voronoi grid.}

\item{\code{centralPoints}}{A list of named points (instances of class NamedPoint)
defining the central points of the Voronoi grid.}

\item{\code{boundaries}}{A list of named points (instances of class NamedPoint)
defining the boundaries of the Voronoi grid.}

\item{\code{borders}}{A list of points (instances of class Point) defining the
Voronoi grid's borders.}

\item{\code{cells}}{A list of cells (instances of class VoronoiCell) giving the
cells of the Voronoi grid.}

\item{\code{areaUnitName}}{The name of the unit the cells' area is computed in.}

\item{\code{isLocked}}{This flag specifies whether the Voronoi grid is locked.
Defaults to \code{FALSE}.}

\item{\code{user}}{The user the Voronoi grid belongs to.}

\item{\code{isFavorite}}{This flag specifies whether the Voronoi grid is marked as
favorite item, and thus shown in the favorites tab on the home module in
PetroVisor. Defaults to \code{FALSE}.}

\item{\code{labels}}{A list of strings holding the labels of the Voronoi grid.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VoronoiGrid-new}{\code{VoronoiGrid$new()}}
\item \href{#method-VoronoiGrid-toList}{\code{VoronoiGrid$toList()}}
\item \href{#method-VoronoiGrid-clone}{\code{VoronoiGrid$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VoronoiGrid-new"></a>}}
\if{latex}{\out{\hypertarget{method-VoronoiGrid-new}{}}}
\subsection{Method \code{new()}}{
Create a new VoronoiGrid instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VoronoiGrid$new(
  name = NULL,
  centralPoints = NULL,
  boundaries = NULL,
  borders = NULL,
  cells = NULL,
  areaUnitName = NULL,
  isLocked = FALSE,
  user = NULL,
  isFavorite = FALSE,
  labels = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the Voronoi grid.}

\item{\code{centralPoints}}{A list of named points (instances of class
NamedPoint) defining the central points of the Voronoi grid.}

\item{\code{boundaries}}{A list of named points (instances of class NamedPoint)
defining the boundaries of the Voronoi grid.}

\item{\code{borders}}{A list of points (instances of class Point) defining the
Voronoi grid's borders.}

\item{\code{cells}}{A list of cells (instances of class VoronoiCell) giving the
cells of the Voronoi grid.}

\item{\code{areaUnitName}}{The name of the unit the cells' area is computed in.}

\item{\code{isLocked}}{This flag specifies whether the Voronoi grid is locked.
Defaults to \code{FALSE}.}

\item{\code{user}}{The user the Voronoi grid belongs to.}

\item{\code{isFavorite}}{This flag specifies whether the Voronoi grid is marked
as favorite item, and thus shown in the favorites tab on the home
module in PetroVisor. Defaults to \code{FALSE}.}

\item{\code{labels}}{A list of strings holding the labels of the Voronoi grid.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VoronoiGrid-toList"></a>}}
\if{latex}{\out{\hypertarget{method-VoronoiGrid-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VoronoiGrid$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VoronoiGrid-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VoronoiGrid-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VoronoiGrid$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
