% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workflow.R
\name{WorkflowActivity}
\alias{WorkflowActivity}
\title{WorkflowActivity}
\description{
Class representing a workflow activity object.
}
\examples{
\dontrun{
WorkflowActivity$new()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{activityType}}{String giving the activity type.}

\item{\code{activityName}}{The activity's name.}

\item{\code{settings}}{The activity's settings as string.}

\item{\code{mappedInputArguments}}{A list of mapped input arguments. The items in
the list must be objects of the class \code{ActivityMappedArgument}.}

\item{\code{mappedOutputArguments}}{a list of mapped output arguments. The items
in the list must be objects of the class \code{ActivityMappedArgument}.}

\item{\code{inputArgumentsInfo}}{A list of argument information. The items in the
list must be objects of the class \code{ActivityArgument}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-WorkflowActivity-new}{\code{WorkflowActivity$new()}}
\item \href{#method-WorkflowActivity-toList}{\code{WorkflowActivity$toList()}}
\item \href{#method-WorkflowActivity-clone}{\code{WorkflowActivity$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkflowActivity-new"></a>}}
\if{latex}{\out{\hypertarget{method-WorkflowActivity-new}{}}}
\subsection{Method \code{new()}}{
Create a new workflow activity instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkflowActivity$new(
  activityType = NULL,
  activityName = NULL,
  settings = NULL,
  mappedInputArguments = NULL,
  mappedOutputArguments = NULL,
  inputArgumentsInfo = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{activityType}}{String giving the activity type.}

\item{\code{activityName}}{The activity's name.}

\item{\code{settings}}{The activity's settings as string.}

\item{\code{mappedInputArguments}}{A list of mapped input arguments. The items
in the list must be objects of the class \code{ActivityMappedArgument}.}

\item{\code{mappedOutputArguments}}{a list of mapped output arguments. The items
in the list must be objects of the class \code{ActivityMappedArgument}.}

\item{\code{inputArgumentsInfo}}{A list of argument information. The items in
the list must be objects of the class \code{ActivityArgument}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkflowActivity-toList"></a>}}
\if{latex}{\out{\hypertarget{method-WorkflowActivity-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkflowActivity$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkflowActivity-clone"></a>}}
\if{latex}{\out{\hypertarget{method-WorkflowActivity-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkflowActivity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
