% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfigurationSetting.R
\name{ConfigurationSetting}
\alias{ConfigurationSetting}
\title{ConfigurationSetting}
\description{
Class representing a PetroVisor ConfigurationSetting object.
}
\examples{
\dontrun{
ConfigurationSetting$new(name = "MyStringValue",
                         valueType = "String",
                         stringValue = "Hello World!")
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the configuration setting.}

\item{\code{numericValue}}{The numeric value of the configuration setting.}

\item{\code{stringValue}}{The string value of the configuration setting.}

\item{\code{listValue}}{The list value of the configuration setting (type list).}

\item{\code{enumerationValue}}{The enumeration value of the configuration setting
(named list).}

\item{\code{dictionaryValue}}{The dictionary value of the configuration setting
(named list).}

\item{\code{valueType}}{The value type of the configuration setting. Must be one of:
\code{"Numeric"}, \code{"String"}, \code{"Numeric with unit"},
\code{"List"}, \code{"Enumeration"}, \code{"Dictionary"}}

\item{\code{unitName}}{The name of the unit associated with the numeric value if the
valueType is \code{"Numeric with unit"}.}

\item{\code{isSystem}}{This flaf specifies, whether the configuration setting is a
PetroVisor system setting (immutable).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ConfigurationSetting-new}{\code{ConfigurationSetting$new()}}
\item \href{#method-ConfigurationSetting-toList}{\code{ConfigurationSetting$toList()}}
\item \href{#method-ConfigurationSetting-clone}{\code{ConfigurationSetting$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfigurationSetting-new"></a>}}
\if{latex}{\out{\hypertarget{method-ConfigurationSetting-new}{}}}
\subsection{Method \code{new()}}{
Create a new ConfigurationSetting instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigurationSetting$new(
  name = NULL,
  numericValue = NULL,
  stringValue = NULL,
  listValue = NULL,
  enumerationValue = NULL,
  dictionaryValue = NULL,
  valueType = c("Numeric", "String", "Numeric with unit", "List", "Enumeration",
    "Dictionary"),
  unitName = NULL,
  isSystem = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the configuration setting.}

\item{\code{numericValue}}{The numeric value of the configuration setting.}

\item{\code{stringValue}}{The string value of the configuration setting.}

\item{\code{listValue}}{The list value of the configuration setting (type list).}

\item{\code{enumerationValue}}{The enumeration value of the configuration
setting (named list).}

\item{\code{dictionaryValue}}{The dictionary value of the configuration setting
(named list).}

\item{\code{valueType}}{The value type of the configuration setting. Must be one
of: \code{"Numeric"}, \code{"String"}, \code{"Numeric with unit"},
\code{"List"}, \code{"Enumeration"}, \code{"Dictionary"}}

\item{\code{unitName}}{The name of the unit associated with the numeric value if
the valueType is \code{"Numeric with unit"}.}

\item{\code{isSystem}}{This flaf specifies, whether the configuration setting is
a PetroVisor system setting (immutable).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfigurationSetting-toList"></a>}}
\if{latex}{\out{\hypertarget{method-ConfigurationSetting-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigurationSetting$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfigurationSetting-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ConfigurationSetting-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigurationSetting$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
