% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Context.R
\name{Context}
\alias{Context}
\title{Context}
\description{
Class representing a PetroVisor context object.
}
\examples{
\dontrun{
Context$new(name = "MyContext",
          entitySet = EntitySet$new(name = "MyEntities",
              entities = c(Entity$new(name = "Well1",
                                      entityTypeName = "Well",
                                      alias = "WellAlias1"),
                           Entity$new(name = "Well2",
                                      entityTypeName = "Well",
                                      alias = "WellAlias2")),
          scope = Scope$new(name = "MyScope",
                            start = "2020-01-01T00:00:00.000Z",
                            end = "2020-03-01T00:00:00.000Z",
                            timeIncrement = "Daily"))
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the context.}

\item{\code{entitySet}}{The context's entity set (object of class EntitySet).}

\item{\code{scope}}{The context's scope (object of class Scope).}

\item{\code{hierarchy}}{(Optional) The hierarchy used for automatic aggregation
(Object of type Hierarchy).}

\item{\code{formula}}{The context's definition as string (P# syntax).}

\item{\code{isLocked}}{This flag specifies whether the context is locked.
Defaults to \code{FALSE}.}

\item{\code{user}}{The user the context belongs to.}

\item{\code{isFavorite}}{This flag specifies whether the context is marked as
favorite item, and thus shown in the favorites tab on the home module in
PetroVisor. Defaults to \code{FALSE}.}

\item{\code{labels}}{A list of strings holding the labels of the context.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Context-new}{\code{Context$new()}}
\item \href{#method-Context-toList}{\code{Context$toList()}}
\item \href{#method-Context-clone}{\code{Context$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Context-new"></a>}}
\if{latex}{\out{\hypertarget{method-Context-new}{}}}
\subsection{Method \code{new()}}{
Create a new Context instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Context$new(
  name = NULL,
  entitySet = NULL,
  scope = NULL,
  hierarchy = NULL,
  formula = NULL,
  isLocked = FALSE,
  user = NULL,
  isFavorite = FALSE,
  labels = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the context.}

\item{\code{entitySet}}{The context's entity set (object of class EntitySet).}

\item{\code{scope}}{The context's scope (object of class Scope).}

\item{\code{hierarchy}}{(Optional) The hierarchy used for automatic aggregation
(Object of type Hierarchy).}

\item{\code{formula}}{The context's definition as string (P# syntax).}

\item{\code{isLocked}}{This flag specifies whether the context is locked.
Defaults to \code{FALSE}.}

\item{\code{user}}{The user the context belongs to.}

\item{\code{isFavorite}}{This flag specifies whether the context is marked as
favorite item, and thus shown in the favorites tab on the home module
in PetroVisor. Defaults to \code{FALSE}.}

\item{\code{labels}}{A list of strings holding the labels of the context.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Context-toList"></a>}}
\if{latex}{\out{\hypertarget{method-Context-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Context$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Context-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Context-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Context$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
