% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataServices.R
\name{DataServices}
\alias{DataServices}
\title{DataServices}
\description{
Provides access to logging functionality provided through the
web API.
}
\details{
A new instance of this class will be created by the ServiceProvider
automatically.
}
\examples{
\dontrun{
# create a new instance of the service provider
sp <- ServiceProvider$new("Host", 8095, "WorkspaceA", "UserX", "Password")

# load static numeric data
dataSetRequest1 <- DataSetRequest$new(
  entityName = "Well01",
  signalName = "surface x-coordinate",
  unitName = "m"
)
dataSetRequest2 <- DataSetRequest$new(
  entityName = "Well01",
  signalName = "surface y-coordinate",
  unitName = "m"
)
data <- sp$dataServices$GetStaticNumeric()$Load(list(
  dataSetRequest1,
  dataSetRequest2
))

# save static string data
staticStringData <- StaticData$new(
  signalName = "model",
  entiyName = "Well01",
  unitName = " ",
  data = "string one"
)
sp$dataServices$GetStaticString()$Save(list(staticStringData))
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataServices-new}{\code{DataServices$new()}}
\item \href{#method-DataServices-GetStaticNumeric}{\code{DataServices$GetStaticNumeric()}}
\item \href{#method-DataServices-GetStaticString}{\code{DataServices$GetStaticString()}}
\item \href{#method-DataServices-GetTimeNumeric}{\code{DataServices$GetTimeNumeric()}}
\item \href{#method-DataServices-GetTimeString}{\code{DataServices$GetTimeString()}}
\item \href{#method-DataServices-clone}{\code{DataServices$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataServices-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataServices-new}{}}}
\subsection{Method \code{new()}}{
Create a new DataServices instance. This is done by the
ServiceProvider automatically.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataServices$new(url, tokenType, token)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{the URL for the API calls.}

\item{\code{tokenType}}{the type of the issued token.}

\item{\code{token}}{the issued token.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataServices-GetStaticNumeric"></a>}}
\if{latex}{\out{\hypertarget{method-DataServices-GetStaticNumeric}{}}}
\subsection{Method \code{GetStaticNumeric()}}{
Get functionality related to static numeric data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataServices$GetStaticNumeric()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new instance of the class 'StaticNumeric'.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataServices-GetStaticString"></a>}}
\if{latex}{\out{\hypertarget{method-DataServices-GetStaticString}{}}}
\subsection{Method \code{GetStaticString()}}{
Get functionality related to static string data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataServices$GetStaticString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new instance of the class 'StaticString'.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataServices-GetTimeNumeric"></a>}}
\if{latex}{\out{\hypertarget{method-DataServices-GetTimeNumeric}{}}}
\subsection{Method \code{GetTimeNumeric()}}{
Get functionality related to time-dependent numeric data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataServices$GetTimeNumeric()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new instance of the class 'TimeNumeric'.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataServices-GetTimeString"></a>}}
\if{latex}{\out{\hypertarget{method-DataServices-GetTimeString}{}}}
\subsection{Method \code{GetTimeString()}}{
Get functionality related to time-dependent string data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataServices$GetTimeString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new instance of the class 'TimeString'.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataServices-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataServices-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataServices$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
