% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetRequest.R
\name{DataSetRequest}
\alias{DataSetRequest}
\title{DataSetRequest}
\description{
Class representing a PetroVisor dataset request object.
}
\examples{
\dontrun{
DataSetRequest$new(entityName = "Well01",
                   signalName = "surface x-coordinate",
                   unitName = "m")
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{entityName}}{The name of the entity for which data is requested.}

\item{\code{signalName}}{The name of the requested signal.}

\item{\code{unitName}}{The name of the unit the data is requested in.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataSetRequest-new}{\code{DataSetRequest$new()}}
\item \href{#method-DataSetRequest-toList}{\code{DataSetRequest$toList()}}
\item \href{#method-DataSetRequest-clone}{\code{DataSetRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSetRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataSetRequest-new}{}}}
\subsection{Method \code{new()}}{
Create a new DataSetRequest instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSetRequest$new(entityName = NULL, signalName = NULL, unitName = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{entityName}}{The name of the entity for which data is requested.}

\item{\code{signalName}}{The name of the requested signal.}

\item{\code{unitName}}{The name of the unit the data is requested in.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSetRequest-toList"></a>}}
\if{latex}{\out{\hypertarget{method-DataSetRequest-toList}{}}}
\subsection{Method \code{toList()}}{
Convert the object to a list. This function is mainly used
by the DataServices to convert the objects to lists and then
call the web API.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSetRequest$toList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSetRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataSetRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSetRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
