% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataTypes.R
\name{DepthData}
\alias{DepthData}
\title{DepthData}
\description{
Class representing depth-dependent data.
}
\examples{
\dontrun{
TimeData$new(singal_name = "produced oil per depth increment",
             entity_name = "Well01",
             unit_name = "m3",
             data = list(list(Depth = 100,
                              Value = 20)))
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{signal_name}}{The name of the signal.}

\item{\code{entity_name}}{The name of the entity.}

\item{\code{unit_name}}{The data's unit name.}

\item{\code{data}}{The data. A list of depth-value-pairs (named list).}

\item{\code{scenario}}{The scenario the data is part of.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DepthData-new}{\code{DepthData$new()}}
\item \href{#method-DepthData-to_list}{\code{DepthData$to_list()}}
\item \href{#method-DepthData-clone}{\code{DepthData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DepthData-new"></a>}}
\if{latex}{\out{\hypertarget{method-DepthData-new}{}}}
\subsection{Method \code{new()}}{
Create a new DepthData instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DepthData$new(
  signal_name = NULL,
  entity_name = NULL,
  unit_name = NULL,
  data = list(),
  scenario = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{signal_name}}{The name of the signal.}

\item{\code{entity_name}}{The name of the entity.}

\item{\code{unit_name}}{The data's unit name.}

\item{\code{data}}{The data. A list of date-value-pairs (named list)}

\item{\code{scenario}}{The scenario the data is part of.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DepthData-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-DepthData-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert the object to a list. This function is mainly used
by the DataServices to convert the objects to lists and then
call the web API.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DepthData$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DepthData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DepthData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DepthData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
