% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Entity.R
\name{Entity}
\alias{Entity}
\title{Entity}
\description{
Class representing a PetroVisor entity object.
}
\examples{
\dontrun{
Entity$new(
  name = "NewWell",
  entity_type_name = "Well",
  alias = "NewWellAlias",
  is_opportunity = FALSE)
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the entity.}

\item{\code{entity_type_name}}{The name of the entity's type.}

\item{\code{alias}}{The alias of the entity.}

\item{\code{is_opportunity}}{Whether the entity is an opportunity or not.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Entity-new}{\code{Entity$new()}}
\item \href{#method-Entity-toList}{\code{Entity$toList()}}
\item \href{#method-Entity-clone}{\code{Entity$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Entity-new"></a>}}
\if{latex}{\out{\hypertarget{method-Entity-new}{}}}
\subsection{Method \code{new()}}{
Create a new Entity instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Entity$new(
  name = NULL,
  entity_type_name = NULL,
  alias = NULL,
  is_opportunity = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of entity.}

\item{\code{entity_type_name}}{The name of the entity' type.}

\item{\code{alias}}{The alias of the entity.}

\item{\code{is_opportunity}}{Whether the entity is an opportunity or not.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Entity-toList"></a>}}
\if{latex}{\out{\hypertarget{method-Entity-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Entity$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Entity-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Entity-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Entity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
