% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntitySet.R
\name{EntitySet}
\alias{EntitySet}
\title{EntitySet}
\description{
Class representing a PetroVisor entity set object.
}
\examples{
\dontrun{
EntitySet$new(name = "MyEntities",
              entities = c(Entity$new(name = "Well1",
                                      entity_type_name = "Well",
                                      alias = "WellAlias1"),
                           Entity$new(name = "Well2",
                                      entity_type_name = "Well",
                                      alias = "WellAlias2")))
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the entity set.}

\item{\code{entities}}{list of entity objects.}

\item{\code{formula}}{The entity set's definition as string (P# syntax).}

\item{\code{description}}{The description of the item.}

\item{\code{labels}}{A list of strings holding the labels of the entity set.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EntitySet-new}{\code{EntitySet$new()}}
\item \href{#method-EntitySet-toList}{\code{EntitySet$toList()}}
\item \href{#method-EntitySet-clone}{\code{EntitySet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EntitySet-new"></a>}}
\if{latex}{\out{\hypertarget{method-EntitySet-new}{}}}
\subsection{Method \code{new()}}{
Create a new EntitySet instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EntitySet$new(
  name = NULL,
  entities = NULL,
  formula = NULL,
  description = NULL,
  labels = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the entity set.}

\item{\code{entities}}{list of entity objects.}

\item{\code{formula}}{The entity set's definition as string (P# syntax).}

\item{\code{description}}{The description of the item.}

\item{\code{labels}}{A list of strings holding the labels of the entity set.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EntitySet-toList"></a>}}
\if{latex}{\out{\hypertarget{method-EntitySet-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EntitySet$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EntitySet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EntitySet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EntitySet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
