% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileService.R
\name{FileService}
\alias{FileService}
\title{FileService}
\value{
A character vector containing the names of the file in the
workspace's blob storage.

The content of the file as string.
}
\description{
Provides access to file related functionality provided through
the web API.
}
\details{
A new instance of this class will be created by the ServiceProvider
automatically.
}
\examples{
\dontrun{
# create a new instance of the service provider
sp <- ServiceProvider$new(
  url = discovery_url,
  workspace = workspace,
  user = user,
  password = password
)

# load file
file <- sp$files$load("Test_File.csv")

# save file
result <- sp$files$save("Test_File.csv")
}
}
\seealso{
\itemize{
\item \link{ServiceProvider} for accessing the file service via \code{sp$files}
\item \code{vignette("getting-started")} for basic usage
}
}
\section{Super class}{
\code{\link[Myrconn.PetroVisor.Client:ApiRequests]{Myrconn.PetroVisor.Client::ApiRequests}} -> \code{FileService}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FileService-new}{\code{FileService$new()}}
\item \href{#method-FileService-load_names}{\code{FileService$load_names()}}
\item \href{#method-FileService-load}{\code{FileService$load()}}
\item \href{#method-FileService-save}{\code{FileService$save()}}
\item \href{#method-FileService-delete}{\code{FileService$delete()}}
\item \href{#method-FileService-clone}{\code{FileService$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileService-new"></a>}}
\if{latex}{\out{\hypertarget{method-FileService-new}{}}}
\subsection{Method \code{new()}}{
Create a new FileService instance. This is done by the
\link{ServiceProvider} automatically.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileService$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileService-load_names"></a>}}
\if{latex}{\out{\hypertarget{method-FileService-load_names}{}}}
\subsection{Method \code{load_names()}}{
Retrieve the names of the files in the workspace's blob
storage.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileService$load_names(prefix = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prefix}}{If specified, only file names with the given prefix are
returned.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileService-load"></a>}}
\if{latex}{\out{\hypertarget{method-FileService-load}{}}}
\subsection{Method \code{load()}}{
Download the file with the specified name from the
workspace's blob storage.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileService$load(name, target_path = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the file to download.}

\item{\code{target_path}}{The path to download the file to.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileService-save"></a>}}
\if{latex}{\out{\hypertarget{method-FileService-save}{}}}
\subsection{Method \code{save()}}{
Upload a file to the workspace's blob storage.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileService$save(file)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{The file to upload (path incl. file name).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileService-delete"></a>}}
\if{latex}{\out{\hypertarget{method-FileService-delete}{}}}
\subsection{Method \code{delete()}}{
Delete the file with the specified name from the
workspace's blob storage.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileService$delete(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the file to delete.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileService-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FileService-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileService$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
