% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogEntry.R
\name{LogEntry}
\alias{LogEntry}
\title{LogEntry}
\description{
Class to define and create new log entries.
}
\examples{
\dontrun{
LogEntry$new(category = "MyCategory",
             message = "This is the message",
             messageDetails = "and here are some details")
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{category}}{The category of the log entry. May be any string.}

\item{\code{elapsedTime}}{The time an operation needed to execute.}

\item{\code{endTime}}{The end time of an operation.}

\item{\code{entityName}}{Thename of the entity affected by the operation.}

\item{\code{message}}{The message of the log entry.}

\item{\code{messageDetails}}{Details of the log entry message.}

\item{\code{numberOfItems}}{The number of items changed by the operation.}

\item{\code{scheduleName}}{The name of the schedule affected by the operation.}

\item{\code{scriptName}}{The name of the script affected by the operation}

\item{\code{severity}}{The severity of the tag entry.}

\item{\code{signalName}}{The name of the signal affected by the operation.}

\item{\code{startTime}}{The start time of an operation.}

\item{\code{tagName}}{The name of the tag affected by the operation.}

\item{\code{timestamp}}{The timestamp of the tag entry. If none is specified, the
current date and time will be inserted automatically.}

\item{\code{unitName}}{The name of the unit affected by the operation.}

\item{\code{userName}}{If not specified, the current user will be inserted
automatically.}

\item{\code{valueAfter}}{The value after a value change.}

\item{\code{valueBefore}}{The value before a value change.}

\item{\code{workflowName}}{The name of the executed workflow. May be any string.}

\item{\code{workspace}}{The workspace of the log entry. If none is specified,
the current workspace will be inserted automatically.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LogEntry-new}{\code{LogEntry$new()}}
\item \href{#method-LogEntry-toList}{\code{LogEntry$toList()}}
\item \href{#method-LogEntry-clone}{\code{LogEntry$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogEntry-new"></a>}}
\if{latex}{\out{\hypertarget{method-LogEntry-new}{}}}
\subsection{Method \code{new()}}{
Create a new LogEntry instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogEntry$new(
  category = NULL,
  elapsedTime = NULL,
  endTime = NULL,
  entityName = NULL,
  message = NULL,
  messageDetails = NULL,
  numberOfItems = NULL,
  scheduleName = NULL,
  scriptName = NULL,
  severity = NULL,
  signalName = NULL,
  startTime = NULL,
  tagName = NULL,
  timestamp = NULL,
  unitName = NULL,
  userName = NULL,
  valueAfter = NULL,
  valueBefore = NULL,
  workflowName = NULL,
  workspace = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{category}}{The category of the log entry. May be any string.}

\item{\code{elapsedTime}}{The time an operation needed to execute.}

\item{\code{endTime}}{The end time of an operation.}

\item{\code{entityName}}{Thename of the entity affected by the operation.}

\item{\code{message}}{The message of the log entry.}

\item{\code{messageDetails}}{Details of the log entry message.}

\item{\code{numberOfItems}}{The number of items changed by the operation.}

\item{\code{scheduleName}}{The name of the schedule affected by the operation.}

\item{\code{scriptName}}{The name of the script affected by the operation}

\item{\code{severity}}{The severity of the tag entry.}

\item{\code{signalName}}{The name of the signal affected by the operation.}

\item{\code{startTime}}{The start time of an operation.}

\item{\code{tagName}}{The name of the tag affected by the operation.}

\item{\code{timestamp}}{The timestamp of the tag entry. If none is specified,
the current date and time will be inserted automatically.}

\item{\code{unitName}}{The name of the unit affected by the operation.}

\item{\code{userName}}{If not specified, the current user will be inserted
automatically.}

\item{\code{valueAfter}}{The value after a value change.}

\item{\code{valueBefore}}{The value before a value change.}

\item{\code{workflowName}}{The name of the executed workflow. May be any string.}

\item{\code{workspace}}{The workspace of the log entry. If none is specified,
the current workspace will be inserted automatically.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogEntry-toList"></a>}}
\if{latex}{\out{\hypertarget{method-LogEntry-toList}{}}}
\subsection{Method \code{toList()}}{
Convert the object to a list. This function is mainly used
by the LoggingService to convert the objects to lists and then call
the web API.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogEntry$toList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogEntry-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LogEntry-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogEntry$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
