% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoggingService.R
\name{LoggingService}
\alias{LoggingService}
\title{LoggingService}
\description{
Provides access to logging functionality provided through the
web API.
}
\details{
A new instance of this class will be created by the ServiceProvider
automatically.
}
\examples{
\dontrun{
# create a new instance of the service provider
sp <- ServiceProvider$new("Host", 8095, "WorkspaceA", "UserX", "Password")

# get available categories
availableCategories <- sp$logs$GetAvailableCategories()

# get log entries
allLogEntries <- sp$logs$GetLogEntries()
warnings <- sp$logs$GetLogEntries(severity = "WARNING")
signIns <- sp$logs$GetLogEntries(category = "SignIn")

# add log entry
entry <- LogEntry$new(message = "Test", category = "Tag")
sp$logs$AddLogEntry(entry)

# add several log entries at once
entry1 <- LogEntry$new(message = "Test1", category = "Tag")
entry2 <- LogEntry$new(message = "Test2", category = "Tag")
sp$logs$AddLogEntries(list(entry1, entry2))

# remove all log entries
sp$logs$CleanUpLogEntries()
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LoggingService-new}{\code{LoggingService$new()}}
\item \href{#method-LoggingService-GetAvailableCategories}{\code{LoggingService$GetAvailableCategories()}}
\item \href{#method-LoggingService-GetLogEntries}{\code{LoggingService$GetLogEntries()}}
\item \href{#method-LoggingService-AddLogEntry}{\code{LoggingService$AddLogEntry()}}
\item \href{#method-LoggingService-AddLogEntries}{\code{LoggingService$AddLogEntries()}}
\item \href{#method-LoggingService-CleanUpLogEntries}{\code{LoggingService$CleanUpLogEntries()}}
\item \href{#method-LoggingService-clone}{\code{LoggingService$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LoggingService-new"></a>}}
\if{latex}{\out{\hypertarget{method-LoggingService-new}{}}}
\subsection{Method \code{new()}}{
Create a new LoggingService instance. This is done by the
ServiceProvider automatically.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LoggingService$new(url, tokenType, token)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{the URL for the API calls.}

\item{\code{tokenType}}{the type of the issued token.}

\item{\code{token}}{the issued token.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LoggingService-GetAvailableCategories"></a>}}
\if{latex}{\out{\hypertarget{method-LoggingService-GetAvailableCategories}{}}}
\subsection{Method \code{GetAvailableCategories()}}{
Get all available categories from the existing log entries.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LoggingService$GetAvailableCategories()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character vector containing all available categories.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LoggingService-GetLogEntries"></a>}}
\if{latex}{\out{\hypertarget{method-LoggingService-GetLogEntries}{}}}
\subsection{Method \code{GetLogEntries()}}{
Retrieve all log entrties matching the given filter from the
database.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LoggingService$GetLogEntries(
  lastEntries = NULL,
  start = NULL,
  end = NULL,
  category = NULL,
  categories = NULL,
  userName = NULL,
  severity = NULL,
  logEntryContentContains = NULL,
  entityName = NULL,
  signalName = NULL,
  tagName = NULL,
  unitName = NULL,
  scriptName = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lastEntries}}{Return only the latest n tag entries.}

\item{\code{start}}{Return log entries after the specified date (inclusive).}

\item{\code{end}}{Return log entries before the specified date (inclusive).}

\item{\code{category}}{Return log entries of the specified category.}

\item{\code{categories}}{Return log entries of the specified categaories.}

\item{\code{userName}}{Return log entries of the specified user.}

\item{\code{severity}}{Return log entries of the given severity.}

\item{\code{logEntryContentContains}}{Return log entries whose massage
contains the specified text.}

\item{\code{entityName}}{Return log entries of the given entity.}

\item{\code{signalName}}{Return log entries of the specified signal.}

\item{\code{tagName}}{Return log entries of the specified tag.}

\item{\code{unitName}}{Return log entries of the given unit.}

\item{\code{scriptName}}{Return log entries of the specified script.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A dataframe containing the requested log entries. The number of
returned columns depends on the available information. Columns that
contain no information are not shown in the output.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LoggingService-AddLogEntry"></a>}}
\if{latex}{\out{\hypertarget{method-LoggingService-AddLogEntry}{}}}
\subsection{Method \code{AddLogEntry()}}{
Add a logentry to the database.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LoggingService$AddLogEntry(logEntry)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{logEntry}}{A logEntry-object created by instantiating the class
'LogEntry'.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LoggingService-AddLogEntries"></a>}}
\if{latex}{\out{\hypertarget{method-LoggingService-AddLogEntries}{}}}
\subsection{Method \code{AddLogEntries()}}{
Add several log entries to the database at once.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LoggingService$AddLogEntries(logEntries)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{logEntries}}{a list of logEntry-objects.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LoggingService-CleanUpLogEntries"></a>}}
\if{latex}{\out{\hypertarget{method-LoggingService-CleanUpLogEntries}{}}}
\subsection{Method \code{CleanUpLogEntries()}}{
Remove log entries from the database.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LoggingService$CleanUpLogEntries(daysToKeep = 0, category)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{daysToKeep}}{Keep the log entries of the last n days in the log.
Defaults to 0.}

\item{\code{category}}{Remove log entries of the specified category only.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LoggingService-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LoggingService-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LoggingService$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
