% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaschineLearning.R
\name{MLModel}
\alias{MLModel}
\title{MLModel}
\description{
Class representing a PetroVisor ml model object.
}
\examples{
\dontrun{
MLModel$new()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the ml model.}

\item{\code{modelType}}{The model's type.}

\item{\code{trainedModel}}{This field holds the string representation of the
trained ML model.}

\item{\code{tableFormula}}{The model's table formula as string.}

\item{\code{contextFormula}}{The model's context formula as string.}

\item{\code{labelColumnName}}{The model's label column name.}

\item{\code{isLocked}}{This flag specifies whether the model is locked.
Defaults to \code{FALSE}.}

\item{\code{user}}{The user the model belongs to.}

\item{\code{isFavorite}}{This flag specifies whether the model is marked as favorite
item, and thus shown in the favorites tab on the home module in PetroVisor.
Defaults to \code{FALSE}.}

\item{\code{labels}}{A list of strings holding the labels of the model.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MLModel-new}{\code{MLModel$new()}}
\item \href{#method-MLModel-toList}{\code{MLModel$toList()}}
\item \href{#method-MLModel-clone}{\code{MLModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-MLModel-new}{}}}
\subsection{Method \code{new()}}{
Create a new ML model instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLModel$new(
  name = NULL,
  modelType = NULL,
  trainedModel = NULL,
  tableFormula = NULL,
  contextFormula = NULL,
  labelColumnName = NULL,
  isLocked = FALSE,
  user = NULL,
  isFavorite = FALSE,
  labels = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the ml model.}

\item{\code{modelType}}{The model's type.}

\item{\code{trainedModel}}{This field holds the string representation of the
trained ML model.}

\item{\code{tableFormula}}{The model's table formula as string.}

\item{\code{contextFormula}}{The model's context formula as string.}

\item{\code{labelColumnName}}{The model's label column name.}

\item{\code{isLocked}}{This flag specifies whether the model is locked.
Defaults to \code{FALSE}.}

\item{\code{user}}{The user the model belongs to.}

\item{\code{isFavorite}}{This flag specifies whether the model is marked as favorite
item, and thus shown in the favorites tab on the home module in PetroVisor.
Defaults to \code{FALSE}.}

\item{\code{labels}}{A list of strings holding the labels of the model.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLModel-toList"></a>}}
\if{latex}{\out{\hypertarget{method-MLModel-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLModel$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MLModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
