% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PivotTable.R
\name{PivotTable}
\alias{PivotTable}
\title{PivotTable}
\description{
Class representing a PetroVisor pivot table object.
}
\examples{
\dontrun{
PivotTable$new()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the pivot table.}

\item{\code{addEntityAliasColumn}}{This flag specifies whether a column containing
the entity's alias shall be added to the pivot table.}

\item{\code{addEntityParentColumn}}{This flag specifies whether a column containing
the entity's parent shall be added to the pivot table.}

\item{\code{addEntityTypeColumn}}{This flag specifies whether a column containing
the entity's type shall be added to the pivot table.}

\item{\code{scopeFormula}}{The scope definition as string (P# syntax).}

\item{\code{entitySetFormula}}{The entity set definition as string (P# syntax).}

\item{\code{tableFormula}}{The table definition as string (P# syntax).}

\item{\code{hierarchyName}}{The name of the hierarchy used to retrieve the parent
name.}

\item{\code{useDataPreloading}}{This flag specifies whether data preloading shall be
used when calculating the pivot table.}

\item{\code{allowScriptExecution}}{This flag specifies whether calculations in the
table formula are executed or not.}

\item{\code{tagEntryDate}}{The date used to evaluate tag entries.}

\item{\code{savedDate}}{The date the pivot table was saved to the database.}

\item{\code{usedTags}}{A list of tag names used in the pivot table.}

\item{\code{dateTimeFormat}}{A string specifying the time format in the pivot table.}

\item{\code{numberFormat}}{A string specifying the number format in the pivot table.}

\item{\code{nanValueString}}{A string specifying the NaN value in the pivot table.}

\item{\code{tagActiveString}}{A string specifying the text that is stored in the
table if a tag is active.}

\item{\code{isLocked}}{This flag specifies whether the pivot table is locked.
Defaults to \code{FALSE}.}

\item{\code{user}}{The user the pivot table belongs to.}

\item{\code{isFavorite}}{This flag specifies whether the pivot table is marked as
favorite item, and thus shown in the favorites tab on the home module in
PetroVisor. Defaults to \code{FALSE}.}

\item{\code{labels}}{A list of strings holding the labels of the pivot table.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PivotTable-new}{\code{PivotTable$new()}}
\item \href{#method-PivotTable-toList}{\code{PivotTable$toList()}}
\item \href{#method-PivotTable-clone}{\code{PivotTable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PivotTable-new"></a>}}
\if{latex}{\out{\hypertarget{method-PivotTable-new}{}}}
\subsection{Method \code{new()}}{
Create a new PivotTable instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PivotTable$new(
  name = NULL,
  addEntityAliasColumn = NULL,
  addEntityParentColumn = NULL,
  addEntityTypeColumn = NULL,
  scopeFormula = NULL,
  entitySetFormula = NULL,
  tableFormula = NULL,
  hierarchyName = NULL,
  useDataPreloading = NULL,
  allowScriptExecution = NULL,
  tagEntryDate = NULL,
  savedDate = NULL,
  usedTags = NULL,
  dateTimeFormat = NULL,
  numberFormat = NULL,
  nanValueString = NULL,
  tagActiveString = NULL,
  isLocked = FALSE,
  user = NULL,
  isFavorite = FALSE,
  labels = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the pivot table.}

\item{\code{addEntityAliasColumn}}{This flag specifies whether a column
containing the entity's alias shall be added to the pivot table.}

\item{\code{addEntityParentColumn}}{This flag specifies whether a column
containing the entity's parent shall be added to the pivot table.}

\item{\code{addEntityTypeColumn}}{This flag specifies whether a column
containing the entity's type shall be added to the pivot table.}

\item{\code{scopeFormula}}{The scope definition as string (P# syntax).}

\item{\code{entitySetFormula}}{The entity set definition as string (P# syntax).}

\item{\code{tableFormula}}{The table definition as string (P# syntax).}

\item{\code{hierarchyName}}{The name of the hierarchy used to retrieve the
parent name.}

\item{\code{useDataPreloading}}{This flag specifies whether data preloading
shall be used when calculating the pivot table.}

\item{\code{allowScriptExecution}}{This flag specifies whether calculations in
the table formula are executed or not.}

\item{\code{tagEntryDate}}{The date used to evaluate tag entries.}

\item{\code{savedDate}}{The date the pivot table was saved to the database.}

\item{\code{usedTags}}{A list of tag names used in the pivot table.}

\item{\code{dateTimeFormat}}{A string specifying the time format in the pivot
table.}

\item{\code{numberFormat}}{A string specifying the number format in the pivot
table.}

\item{\code{nanValueString}}{A string specifying the NaN value in the pivot
table.}

\item{\code{tagActiveString}}{A string specifying the text that is stored in the
table if a tag is active.}

\item{\code{isLocked}}{This flag specifies whether the pivot table is locked.
Defaults to \code{FALSE}.}

\item{\code{user}}{The user the pivot table belongs to.}

\item{\code{isFavorite}}{This flag specifies whether the pivot table is marked
as favorite item, and thus shown in the favorites tab on the home
module in PetroVisor. Defaults to \code{FALSE}.}

\item{\code{labels}}{A list of strings holding the labels of the pivot table.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PivotTable-toList"></a>}}
\if{latex}{\out{\hypertarget{method-PivotTable-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PivotTable$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PivotTable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PivotTable-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PivotTable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
