% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepositoryService.R
\name{RepositoryService}
\alias{RepositoryService}
\title{RepositoryService}
\description{
Provides access to PetroVisor item related functionalities.
}
\details{
A new instance of this class will be created by the ServiceProvider
automatically.
}
\examples{
\dontrun{
# create a new instance of the service provider
sp <- ServiceProvider$new("Host", 8095, "WorkspaceA", "UserX", "Password")

# get the names of all available entities
entityNames <- sp$repositoryService$GetNamesOfItems("Entity")

# delete an item (delete the hierarchy with the name "test")
sp$repositoryService$DeleteItem("Hierarchy", "test")

# get an item by name
well01 <- sp$repositoryService$GetItemByName("Well", "Well01")

# add or edit an item
entity <- Entity$new(
  name = "TestWell01",
  entityTypeName = "Well",
  alias = "TestAlias01"
)
sp$repositoryService$AddOrEditItem("Entity", entity)
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RepositoryService-new}{\code{RepositoryService$new()}}
\item \href{#method-RepositoryService-GetNamesOfItems}{\code{RepositoryService$GetNamesOfItems()}}
\item \href{#method-RepositoryService-DeleteItem}{\code{RepositoryService$DeleteItem()}}
\item \href{#method-RepositoryService-GetItemByName}{\code{RepositoryService$GetItemByName()}}
\item \href{#method-RepositoryService-AddOrEditItem}{\code{RepositoryService$AddOrEditItem()}}
\item \href{#method-RepositoryService-clone}{\code{RepositoryService$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RepositoryService-new"></a>}}
\if{latex}{\out{\hypertarget{method-RepositoryService-new}{}}}
\subsection{Method \code{new()}}{
Create a new RepositoryService instance. This is done by the
ServiceProvider automatically.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RepositoryService$new(url, tokenType, token)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{the URL for the API calls.}

\item{\code{tokenType}}{the type of the issued token.}

\item{\code{token}}{the issued token.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RepositoryService-GetNamesOfItems"></a>}}
\if{latex}{\out{\hypertarget{method-RepositoryService-GetNamesOfItems}{}}}
\subsection{Method \code{GetNamesOfItems()}}{
Get the names of all items of the given type available in
PetroVisor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RepositoryService$GetNamesOfItems(type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{The type of the item.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character array containing the names of the items.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RepositoryService-DeleteItem"></a>}}
\if{latex}{\out{\hypertarget{method-RepositoryService-DeleteItem}{}}}
\subsection{Method \code{DeleteItem()}}{
Delete an item by name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RepositoryService$DeleteItem(type, name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{The type of the item.}

\item{\code{name}}{Name of the item to delete.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RepositoryService-GetItemByName"></a>}}
\if{latex}{\out{\hypertarget{method-RepositoryService-GetItemByName}{}}}
\subsection{Method \code{GetItemByName()}}{
Get an item by name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RepositoryService$GetItemByName(type, name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{The type of the item.}

\item{\code{name}}{Name of the item to retrieve.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of the specified type class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RepositoryService-AddOrEditItem"></a>}}
\if{latex}{\out{\hypertarget{method-RepositoryService-AddOrEditItem}{}}}
\subsection{Method \code{AddOrEditItem()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RepositoryService$AddOrEditItem(type, item)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{The type of the item.}

\item{\code{item}}{Item to add or edit. Has to be an object of the respective
class.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Add or edit an item.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RepositoryService-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RepositoryService-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RepositoryService$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
