% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scope.R
\name{Scope}
\alias{Scope}
\title{Scope}
\description{
Class representing a PetroVisor scope object.
}
\examples{
\dontrun{
Scope$new(name = "MyScope",
          start = "2020-01-01T00:00:00.000Z",
          end = "2020-03-01T00:00:00.000Z",
          time_increment = "Daily")
}
}
\seealso{
\itemize{
\item \link{RepositoryService} for managing scopes
\item \link{DataServices} for using scopes in data queries
\item \link{EntitySet} for entity filtering
\item \link{MLModel} for ML training context
\item \code{vignette("working-with-data")} for data examples
\item \code{vignette("machine-learning")} for ML examples
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the scope.}

\item{\code{start}}{The scope's start date.}

\item{\code{end}}{The scope's end date.}

\item{\code{time_increment}}{The scope's time increment. Allowed values are:
\code{"EverySecond"}, \code{"EveryMinute"}, \code{"EveryFiveMinutes"},
\code{"EveryFifteenMinutes"}, \code{"Hourly"}, \code{"Daily"},
\code{"Monthly"}, \code{"Quarterly"}, \code{"Yearly"}}

\item{\code{depth_increment}}{The scope's depth increment. Allowed values are:
\code{"Meter"}, \code{"HalfMeter"}, \code{"EighthMeter"},
\code{"TenthMeter"}, \code{"Foot"}, \code{"HalfFoot"}}

\item{\code{start_depth}}{The scope's start depth.}

\item{\code{end_depth}}{The scope's end depth.}

\item{\code{formula}}{The scope's definition as string (P# syntax).}

\item{\code{description}}{The description of the item.}

\item{\code{labels}}{A list of strings holding the labels of the scope.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Scope-new}{\code{Scope$new()}}
\item \href{#method-Scope-toList}{\code{Scope$toList()}}
\item \href{#method-Scope-clone}{\code{Scope$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scope-new"></a>}}
\if{latex}{\out{\hypertarget{method-Scope-new}{}}}
\subsection{Method \code{new()}}{
Create a new Scope instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scope$new(
  name = NULL,
  start = NULL,
  end = NULL,
  time_increment = c("", "EverySecond", "EveryMinute", "EveryFiveMinutes",
    "EveryFifteenMinutes", "Hourly", "Daily", "Monthly", "Quarterly", "Yearly"),
  depth_increment = c("", "Meter", "HalfMeter", "EighthMeter", "TenthMeter", "Foot",
    "HalfFoot"),
  start_depth = NULL,
  end_depth = NULL,
  formula = NULL,
  description = NULL,
  labels = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the scope.}

\item{\code{start}}{The scope's start date.}

\item{\code{end}}{The scope's end date.}

\item{\code{time_increment}}{The scope's time increment. Allowed values are:
\code{"EverySecond"}, \code{"EveryMinute"}, \code{"EveryFiveMinutes"},
\code{"EveryFifteenMinutes"}, \code{"Hourly"}, \code{"Daily"},
\code{"Monthly"}, \code{"Quarterly"}, \code{"Yearly"}}

\item{\code{depth_increment}}{The scope's depth increment. Allowed values are:
\code{"Meter"}, \code{"HalfMeter"}, \code{"EighthMeter"},
\code{"TenthMeter"}, \code{"Foot"}, \code{"HalfFoot"}}

\item{\code{start_depth}}{The scope's start depth.}

\item{\code{end_depth}}{The scope's end depth.}

\item{\code{formula}}{The scope's definition as string (P# syntax).}

\item{\code{description}}{The description of the item.}

\item{\code{labels}}{A list of strings holding the labels of the scope.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scope-toList"></a>}}
\if{latex}{\out{\hypertarget{method-Scope-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scope$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scope-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Scope-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scope$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
