% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ServiceProvider.R
\name{ServiceProvider}
\alias{ServiceProvider}
\title{ServiceProvider}
\value{
Converted values. Either single number or collection of values.
}
\description{
Provides access to all services provided through the web API.
}
\details{
This class allows interaction with the PetroVisor API by providing
access to various services such as logging, repository, data, and tag
entries.
}
\examples{
\dontrun{
# Create a new instance of the service provider using token
sp <- ServiceProvider$new(
  url = "https://identity.us1.petrovisor.com/PetroVisor/API/WorkspaceA/",
  client_token = "your_token_here"
)

# Create a new instance of the service provider using username and password
sp <- ServiceProvider$new(
  url = "https://identity.us1.petrovisor.com",
  workspace = "WorkspaceA",
  user = "your_username",
  password = "your_password"
)
}

## ------------------------------------------------
## Method `ServiceProvider$send_mail`
## ------------------------------------------------

\dontrun{
  # Send a simple email
  sp$send_mail(
    address = "example@domain.com",
    subject = "Test Subject",
    body = "This is a test email from PetroVisor R client."
  )
}
}
\section{Super class}{
\code{\link[Myrconn.PetroVisor.Client:ApiRequests]{Myrconn.PetroVisor.Client::ApiRequests}} -> \code{ServiceProvider}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{url}}{Full URL to the PetroVisor API (either with or without token).}

\item{\code{data_url}}{The base URL for data services.}

\item{\code{workspace}}{The currently used workspace.}

\item{\code{user}}{The current user.}

\item{\code{client_token}}{Access token for the API (optional, used if URL is
provided).}

\item{\code{workspace_data_url}}{The URL for the workspace-specific data services.}

\item{\code{logs}}{Instance of class \code{LoggingService} wrapping all
logging-related functionalities.}

\item{\code{items}}{Instance of class \code{RepositoryService} wrapping all
functionality related to PetroVisor items (entities, signals, units, etc.).}

\item{\code{data}}{Instance of class \code{DataServices} wrapping all functionality
related to data.}

\item{\code{tag_entries}}{Instance of class \code{TagEntriesService} wrapping all
functionality related to tag entries.}

\item{\code{files}}{Instance of class \code{FileService} wrapping all functionality
related to files.}

\item{\code{ml}}{Instance of class \code{MlTrainingService} wrapping
all functionality related to ML model training and prediction.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ServiceProvider-new}{\code{ServiceProvider$new()}}
\item \href{#method-ServiceProvider-parse_signal}{\code{ServiceProvider$parse_signal()}}
\item \href{#method-ServiceProvider-set_workspace_data_url}{\code{ServiceProvider$set_workspace_data_url()}}
\item \href{#method-ServiceProvider-getDataUrl}{\code{ServiceProvider$getDataUrl()}}
\item \href{#method-ServiceProvider-convert_unit}{\code{ServiceProvider$convert_unit()}}
\item \href{#method-ServiceProvider-send_mail}{\code{ServiceProvider$send_mail()}}
\item \href{#method-ServiceProvider-clone}{\code{ServiceProvider$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceProvider-new"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceProvider-new}{}}}
\subsection{Method \code{new()}}{
Create a new ServiceProvider instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceProvider$new(
  url = NULL,
  workspace = NULL,
  user = NULL,
  password = NULL,
  client_token = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{Full URL to the PetroVisor API (if token is provided).}

\item{\code{workspace}}{Name of the workspace to connect to.}

\item{\code{user}}{Username for authentication.}

\item{\code{password}}{Password for the given user.}

\item{\code{client_token}}{Access token for authentication (optional).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceProvider-parse_signal"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceProvider-parse_signal}{}}}
\subsection{Method \code{parse_signal()}}{
Parse the mapped signal received from PetroVisor to a list
containing the name and unit of the mapped signal.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceProvider$parse_signal(mapped_signal)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mapped_signal}}{The mapped signal (string) as received from
PetroVisor.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceProvider-set_workspace_data_url"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceProvider-set_workspace_data_url}{}}}
\subsection{Method \code{set_workspace_data_url()}}{
Set the workspace-specific data URL.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceProvider$set_workspace_data_url(data_url, workspace)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_url}}{The base URL for data services.}

\item{\code{workspace}}{The name of the workspace.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceProvider-getDataUrl"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceProvider-getDataUrl}{}}}
\subsection{Method \code{getDataUrl()}}{
Retrieve the base data URL from the authentication service.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceProvider$getDataUrl()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceProvider-convert_unit"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceProvider-convert_unit}{}}}
\subsection{Method \code{convert_unit()}}{
Convert values from a source unit to a target unit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceProvider$convert_unit(x, source_unit, target_unit)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The value to convert. Either a single value, or a collection of
values.}

\item{\code{source_unit}}{The name of the source unit.}

\item{\code{target_unit}}{The name of the target unit.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceProvider-send_mail"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceProvider-send_mail}{}}}
\subsection{Method \code{send_mail()}}{
Send an email using the PetroVisor email configuration.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceProvider$send_mail(address, subject, body)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{address}}{The recipient email address (single string).}

\item{\code{subject}}{The subject of the email.}

\item{\code{body}}{The body content of the email.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
  # Send a simple email
  sp$send_mail(
    address = "example@domain.com",
    subject = "Test Subject",
    body = "This is a test email from PetroVisor R client."
  )
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceProvider-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceProvider-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceProvider$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
