% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ServiceProvider.R
\name{ServiceProvider}
\alias{ServiceProvider}
\title{ServiceProvider}
\description{
Provides access to all services provided through the web API.
}
\details{
This class allows interaction with the PetroVisor API by providing
access to various services such as logging, repository, data, and tag entries.
}
\examples{
\dontrun{
# Create a new instance of the service provider using token
sp <- ServiceProvider$new(
  url = "https://identity.us1.petrovisor.com/PetroVisor/API/WorkspaceA/",
  client_token = "your_token_here"
)

# Create a new instance of the service provider using username and password
sp <- ServiceProvider$new(
  url = "https://identity.us1.petrovisor.com",
  workspace = "WorkspaceA",
  user = "your_username",
  password = "your_password"
)
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{url}}{Full URL to the PetroVisor API (either with or without token).}

\item{\code{data_url}}{The base URL for data services.}

\item{\code{workspace}}{The currently used workspace.}

\item{\code{user}}{The current user.}

\item{\code{client_token}}{Access token for the API (optional, used if URL is provided).}

\item{\code{workspace_data_url}}{The URL for the workspace-specific data services.}

\item{\code{loggingService}}{Class instance wrapping all logging-related functionalities.}

\item{\code{repositoryService}}{Class instance wrapping all functionality related to
PetroVisor items (entities, signals, units, etc.).}

\item{\code{dataServices}}{Class instance wrapping all functionality related to data.}

\item{\code{tagEntriesService}}{Class instance wrapping all functionality related to
tag entries.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ServiceProvider-new}{\code{ServiceProvider$new()}}
\item \href{#method-ServiceProvider-set_workspace_data_url}{\code{ServiceProvider$set_workspace_data_url()}}
\item \href{#method-ServiceProvider-getDataUrl}{\code{ServiceProvider$getDataUrl()}}
\item \href{#method-ServiceProvider-clone}{\code{ServiceProvider$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceProvider-new"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceProvider-new}{}}}
\subsection{Method \code{new()}}{
Create a new ServiceProvider instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceProvider$new(
  url = NULL,
  workspace = NULL,
  user = NULL,
  password = NULL,
  client_token = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{Full URL to the PetroVisor API (if token is provided).}

\item{\code{workspace}}{Name of the workspace to connect to.}

\item{\code{user}}{Username for authentication.}

\item{\code{password}}{Password for the given user.}

\item{\code{client_token}}{Access token for authentication (optional).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceProvider-set_workspace_data_url"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceProvider-set_workspace_data_url}{}}}
\subsection{Method \code{set_workspace_data_url()}}{
Set the workspace-specific data URL.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceProvider$set_workspace_data_url(data_url, workspace)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_url}}{The base URL for data services.}

\item{\code{workspace}}{The name of the workspace.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceProvider-getDataUrl"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceProvider-getDataUrl}{}}}
\subsection{Method \code{getDataUrl()}}{
Retrieve the base data URL from the authentication service.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceProvider$getDataUrl()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceProvider-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceProvider-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceProvider$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
