% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signal.R
\name{Signal}
\alias{Signal}
\title{Signal}
\description{
Class representing a PetroVisor signal object.
}
\examples{
\dontrun{
Signal$new(name = "my signal",
           shortName = "my sig",
           measurementName = "Length",
           storageUnitName = "m",
           aggregationType = "Average",
           containerAggregationType = "Sum",
           signalType = "Time-dependent",
           defaultColor = 0,
           defaultLineType = "Solid",
           settingName = NULL)
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the signal.}

\item{\code{shortName}}{The signal's short name.}

\item{\code{measurementName}}{The signal's measurement name.}

\item{\code{storageUnitName}}{The name of the signal's storage unit.}

\item{\code{aggregationType}}{The signal's aggregation type.}

\item{\code{containerAggregationType}}{The signal's container aggregation type.}

\item{\code{signalType}}{The signal's type.}

\item{\code{defaultColor}}{The signal's default color.}

\item{\code{defaultLineType}}{The signal's default line type.}

\item{\code{settingName}}{(Optional) The name of the enumeration linked to the
signal.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Signal-new}{\code{Signal$new()}}
\item \href{#method-Signal-toList}{\code{Signal$toList()}}
\item \href{#method-Signal-clone}{\code{Signal$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Signal-new"></a>}}
\if{latex}{\out{\hypertarget{method-Signal-new}{}}}
\subsection{Method \code{new()}}{
Create a new Signal instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Signal$new(
  name = NULL,
  shortName = NULL,
  measurementName = NULL,
  storageUnitName = NULL,
  aggregationType = c("Sum", "Average", "Max", "Min", "Count"),
  containerAggregationType = c("Sum", "Average", "Max", "Min", "Count"),
  signalType = c("Static", "Time-dependent", "Depth-dependent", "String", "PVT",
    "String time-dependent"),
  defaultColor = 0,
  defaultLineType = c("Solid", "Dash", "DashDot"),
  settingName = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the signal.}

\item{\code{shortName}}{The signal's short name.}

\item{\code{measurementName}}{The signal's measurement name.}

\item{\code{storageUnitName}}{The name of the signal's storage unit.}

\item{\code{aggregationType}}{The signal's aggregation type.}

\item{\code{containerAggregationType}}{The signal's container aggregation type.}

\item{\code{signalType}}{The signal's type.}

\item{\code{defaultColor}}{The signal's default color.}

\item{\code{defaultLineType}}{The signal's default line type.}

\item{\code{settingName}}{(Optional) The name of the enumeration linked to the
signal.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Signal-toList"></a>}}
\if{latex}{\out{\hypertarget{method-Signal-toList}{}}}
\subsection{Method \code{toList()}}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Signal$toList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Signal-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Signal-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Signal$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
