% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataTypes.R
\name{StaticData}
\alias{StaticData}
\title{StaticData}
\description{
Class representing static data.
}
\examples{
\dontrun{
StaticData$new(signal_name = "surface x-coordinate",
               entity_name = "Well01",
               unit_name = "m",
               data = 10)
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{signal_name}}{The name of the signal.}

\item{\code{entity_name}}{The name of the entity.}

\item{\code{unit_name}}{The data's unit name.}

\item{\code{data}}{The data. In this case case a single numberic or string value.}

\item{\code{scenario}}{The scenario the data is part of.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StaticData-new}{\code{StaticData$new()}}
\item \href{#method-StaticData-to_list}{\code{StaticData$to_list()}}
\item \href{#method-StaticData-clone}{\code{StaticData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StaticData-new"></a>}}
\if{latex}{\out{\hypertarget{method-StaticData-new}{}}}
\subsection{Method \code{new()}}{
Create a new StaticData instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StaticData$new(
  signal_name = NULL,
  entity_name = NULL,
  unit_name = NULL,
  data = NULL,
  scenario = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{signal_name}}{The name of the signal.}

\item{\code{entity_name}}{The name of the entity.}

\item{\code{unit_name}}{The data's unit name.}

\item{\code{data}}{The data. In this case case a single number or string value.}

\item{\code{scenario}}{The scenario the data is part of.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StaticData-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-StaticData-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StaticData$to_list()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the DataServices to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StaticData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StaticData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StaticData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
