% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tag.R
\name{Tag}
\alias{Tag}
\title{Tag}
\description{
Class representing a PetroVisor tag object.
}
\examples{
\dontrun{
Tag$new(name = "NewTag", tag_group = "Group 1")
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the tag.}

\item{\code{tag_group}}{The tag's tag group.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Tag-new}{\code{Tag$new()}}
\item \href{#method-Tag-toList}{\code{Tag$toList()}}
\item \href{#method-Tag-clone}{\code{Tag$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tag-new"></a>}}
\if{latex}{\out{\hypertarget{method-Tag-new}{}}}
\subsection{Method \code{new()}}{
Create a new Tag instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tag$new(name = NULL, tag_group = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of tag.}

\item{\code{tag_group}}{The tag's tag group.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tag-toList"></a>}}
\if{latex}{\out{\hypertarget{method-Tag-toList}{}}}
\subsection{Method \code{toList()}}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tag$toList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tag-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Tag-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tag$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
