% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TagEntry.R
\name{TagEntry}
\alias{TagEntry}
\title{TagEntry}
\description{
Class representing a PetroVisor tag entry.
}
\examples{
\dontrun{
# create a new instance of the service provider
sp <- ServiceProvider$new("Host", 8095, "WorkspaceA", "UserX", "Password")

# retrieve an entity from the database
well <- sp$repositoryService$GetItemByName(type = "Entity", name = "Well01")
# retrieve a tag from the datbase
active <- sp$repositoryService$GetItemByName(type = "Tag", name = "Active")

# create a new tag entry
TagEntry$new(entity = well, tag = active, start = "2020-02-01T00:00:00.000Z")
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tag}}{The tag of the tag entry. Has to be an instance of the class
'Tag'.}

\item{\code{entity}}{The tagged entity. Has to be an instance of the class 'Entity'.}

\item{\code{start}}{The start date of the tag entry.}

\item{\code{end}}{The end date of the tag entry.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TagEntryFilter-new}{\code{TagEntry$new()}}
\item \href{#method-TagEntryFilter-toList}{\code{TagEntry$toList()}}
\item \href{#method-TagEntryFilter-clone}{\code{TagEntry$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TagEntryFilter-new"></a>}}
\if{latex}{\out{\hypertarget{method-TagEntryFilter-new}{}}}
\subsection{Method \code{new()}}{
Create a new TagEntry instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TagEntry$new(tag = NULL, entity = NULL, start = NULL, end = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tag}}{The tag of the tag entry. Has to be an instance of the class
'Tag'.}

\item{\code{entity}}{The tagged entity. Has to be an instance of the class
'Entity'.}

\item{\code{start}}{The start date of the tag entry.}

\item{\code{end}}{(Optional) The end date of the tag entry.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TagEntryFilter-toList"></a>}}
\if{latex}{\out{\hypertarget{method-TagEntryFilter-toList}{}}}
\subsection{Method \code{toList()}}{
Convert the object to a list. This function is mainly used
by the TagEntriesService to convert the objects to lists and then
call the web API.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TagEntry$toList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TagEntryFilter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TagEntryFilter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TagEntry$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
