% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataTypes.R
\name{TimeData}
\alias{TimeData}
\title{TimeData}
\description{
Class representing time-dependent data.
}
\examples{
\dontrun{
TimeData$new(singalName = "produced oil per time increment",
             entityName = "Well01",
             unitName = "m3",
             data = list(DataPoint$new(date = "2020-01-01T00:00:00.000Z",
                                       value = 20)))
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{signalName}}{The name of the signal.}

\item{\code{entityName}}{The name of the entity.}

\item{\code{unitName}}{The data's unit name.}

\item{\code{data}}{The data. A list of date-value-pairs (objects of type DataPoint)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TimeData-new}{\code{TimeData$new()}}
\item \href{#method-TimeData-toList}{\code{TimeData$toList()}}
\item \href{#method-TimeData-clone}{\code{TimeData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeData-new"></a>}}
\if{latex}{\out{\hypertarget{method-TimeData-new}{}}}
\subsection{Method \code{new()}}{
Create a new StaticData instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeData$new(
  signalName = NULL,
  entityName = NULL,
  unitName = NULL,
  data = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{signalName}}{The name of the signal.}

\item{\code{entityName}}{The name of the entity.}

\item{\code{unitName}}{The data's unit name.}

\item{\code{data}}{The data. A list of date-value-pairs (objects of type
DataPoint)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeData-toList"></a>}}
\if{latex}{\out{\hypertarget{method-TimeData-toList}{}}}
\subsection{Method \code{toList()}}{
Convert the object to a list. This function is mainly used
by the DataServices to convert the objects to lists and then
call the web API.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeData$toList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TimeData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
