% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnitMeasurement.R
\name{UnitMeasurement}
\alias{UnitMeasurement}
\title{UnitMeasurement}
\description{
Class representing a PetroVisor unit measurement object.
}
\examples{
\dontrun{
UnitMeasurement$new(name = "Length", canonical_unit_name = "m")
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the unit measurement.}

\item{\code{canonical_unit_name}}{The name of the unit measurement's canonical unit.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UnitMeasurement-new}{\code{UnitMeasurement$new()}}
\item \href{#method-UnitMeasurement-toList}{\code{UnitMeasurement$toList()}}
\item \href{#method-UnitMeasurement-clone}{\code{UnitMeasurement$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnitMeasurement-new"></a>}}
\if{latex}{\out{\hypertarget{method-UnitMeasurement-new}{}}}
\subsection{Method \code{new()}}{
Create a new UnitMeasurement instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitMeasurement$new(name = NULL, canonical_unit_name = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the unit measurement.}

\item{\code{canonical_unit_name}}{The name of the unit measurement's canonical
unit.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnitMeasurement-toList"></a>}}
\if{latex}{\out{\hypertarget{method-UnitMeasurement-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitMeasurement$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnitMeasurement-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UnitMeasurement-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitMeasurement$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
