% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workflow.R
\name{Workflow}
\alias{Workflow}
\title{Workflow}
\description{
Class representing a PetroVisor workflow object.
}
\examples{
\dontrun{
Workflow$new()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the workflow.}

\item{\code{activities}}{A list of activities. The items in the list must be
objects of the class \code{WorkflowActivity}.}

\item{\code{isLocked}}{This flag specifies whether the workflow is locked.
Defaults to \code{FALSE}.}

\item{\code{user}}{The user the workflow belongs to.}

\item{\code{isFavorite}}{This flag specifies whether the workflow is marked as
favorite item, and thus shown in the favorites tab on the home module in
PetroVisor. Defaults to \code{FALSE}.}

\item{\code{labels}}{A list of strings holding the labels of the workflow.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Workflow-new}{\code{Workflow$new()}}
\item \href{#method-Workflow-toList}{\code{Workflow$toList()}}
\item \href{#method-Workflow-clone}{\code{Workflow$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Workflow-new"></a>}}
\if{latex}{\out{\hypertarget{method-Workflow-new}{}}}
\subsection{Method \code{new()}}{
Create a new Workflow instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Workflow$new(
  name = NULL,
  activities = NULL,
  isLocked = FALSE,
  user = NULL,
  isFavorite = FALSE,
  labels = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the workflow.}

\item{\code{activities}}{A list of activities. The items in the list must be
objects of the class \code{WorkflowActivity}.}

\item{\code{isLocked}}{This flag specifies whether the workflow is locked.
Defaults to \code{FALSE}.}

\item{\code{user}}{The user the workflow belongs to.}

\item{\code{isFavorite}}{This flag specifies whether the workflow is marked as
favorite item, and thus shown in the favorites tab on the home module in
PetroVisor. Defaults to \code{FALSE}.}

\item{\code{labels}}{A list of strings holding the labels of the workflow.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Workflow-toList"></a>}}
\if{latex}{\out{\hypertarget{method-Workflow-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Workflow$toList()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Convert the object to a list. This function is mainly used
by the RepositoryService to convert the objects to lists and then
call the web API.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Workflow-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Workflow-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Workflow$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
